% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdupenames.R
\name{all.equal_functions}
\alias{all.equal_functions}
\title{helper function for checking possibly different versions of a function with same name in 2 packages}
\usage{
\method{all}{equal_functions}(fun = "latlon_infer", package1 = "EJAM", package2 = "EJAMejscreenapi")
}
\arguments{
\item{fun}{quoted name of function, like "latlon_infer"}

\item{package1}{quoted name of package, like "EJAM"}

\item{package2}{quoted name of package, like "EJAMejscreenapi"}
}
\value{
TRUE or FALSE
}
\description{
helper function for checking possibly different versions of a function with same name in 2 packages
}
\seealso{
\code{\link[=dupenames]{dupenames()}}
}
