% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames.R
\name{all_equal_functions}
\alias{all_equal_functions}
\title{all_equal_functions  -  helper to check different versions of function with same name in 2 packages}
\usage{
all_equal_functions(
  fun = "latlon_infer",
  package1 = "EJAM",
  package2 = "EJAMejscreenapi"
)
}
\arguments{
\item{fun}{quoted name of function, like "latlon_infer"}

\item{package1}{quoted name of package, like "EJAM"}

\item{package2}{quoted name of package, like "EJAMejscreenapi"}
}
\value{
TRUE or FALSE
}
\description{
all_equal_functions  -  helper to check different versions of function with same name in 2 packages
}
\seealso{
\code{\link[=dupenames]{dupenames()}} \code{\link[=all.equal.function]{all.equal.function()}}
}
