% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run_EJAM.R
\name{app_run_EJAM}
\alias{app_run_EJAM}
\alias{run_app}
\title{Launch EJAM as shiny web app}
\usage{
app_run_EJAM(
  onStart = NULL,
  options = list(),
  enableBookmarking = "server",
  uiPattern = "/",
  ...
)
}
\description{
launch Shiny web app from RStudio
}
\details{
An alias for run_app()
}
\seealso{
\code{\link[=run_app]{run_app()}}
}
