% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_ask_number.R
\name{ask_number}
\alias{ask_number}
\alias{askradius}
\title{utility - interactive prompt in RStudio to ask user to specify number like radius
same as askradius()}
\usage{
ask_number(
  default = 3,
  title = "Radius",
  message = "Within how many miles of each point?"
)
}
\arguments{
\item{default}{default value}

\item{title}{question}

\item{message}{title of popup dialog box}
}
\value{
a number
}
\description{
utility - interactive prompt in RStudio to ask user to specify number like radius
same as askradius()
}
\seealso{
askYesNo()
}
