% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{bg_from_county}
\alias{bg_from_county}
\title{bg_from_county - Analyze US Counties as if they were sites, to get EJ indicators summary for each county}
\usage{
bg_from_county(fips)
}
\arguments{
\item{fips}{County FIPS vector (ideally as character not numeric values)}
}
\value{
data.table with one row per blockgroup in these counties, or
all pairs of county fips - bgid, and a unique siteid assigned to each county
}
\description{
bg_from_county - Analyze US Counties as if they were sites, to get EJ indicators summary for each county
}
\details{
This function provides one row per blockgroup.
\code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}} provides one row per block.
}
\examples{
 bg_from_county(c('01001','72153'))
 # Largest US Counties by ACS Population Totals:
 blockgroupstats[ , .(ST = ST[1], countypop = sum(pop)),
  by=.(FIPS = substr(bgfips,1,5))][order(-countypop),][1:20, .(
  CountyPopulation = prettyNum(countypop, big.mark = ","), FIPS, ST)]
 
}
\seealso{
\code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}}
}
