% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockid2fips.R
\name{blockid2fips}
\alias{blockid2fips}
\title{blockid2fips (DATA) Census FIPS codes of blocks}
\description{
blockid2fips (DATA) Census FIPS codes of blocks
}
\details{
For documentation on EJScreen, see (https://www.epa.gov/ejscreen)

blockid2fips is a table of all census blocks, with the FIPS codes.

It also has a column
called blockid that can join it to other block datasets.

\if{html}{\out{<div class="sourceCode">}}\preformatted{dataload_from_pins('blockid2fips')

names(blockid2fips)
dim(blockid2fips)
}\if{html}{\out{</div>}}
}
