% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_ejam.R
\name{calc_byformula}
\alias{calc_byformula}
\title{DRAFT utility to use formulas provided as text, to calculate indicators}
\usage{
calc_byformula(mydf, formulas = NULL, keep = formula_varname(formulas))
}
\arguments{
\item{mydf}{data.frame//table of indicators or variables to use}

\item{formulas}{text strings of formulas}

\item{keep}{etc.}
}
\description{
DRAFT utility to use formulas provided as text, to calculate indicators
}
