% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_ejam.R
\name{calc_ejam}
\alias{calc_ejam}
\title{DRAFT utility to use formulas provided as text, to calculate indicators}
\usage{
calc_ejam(bg, keep.old = c("bgid", "pop"), keep.new = "all", formulas)
}
\arguments{
\item{bg}{data.frame//table of indicators or variables to use}

\item{keep.old}{names of columns (variables) to retain from among those provided in bg}

\item{keep.new}{names of calculated variables to retain in output}

\item{formulas}{text strings of formulas}
}
\value{
data.frame of calculated variables one row per bg row
}
\description{
DRAFT utility to use formulas provided as text, to calculate indicators
}
\details{
may be used by doaggregate_custom()
}
