% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_doaggregate.R
\name{calctype}
\alias{calctype}
\title{utility - what type of formula is used to aggregate this variable?}
\usage{
calctype(varnames)
}
\arguments{
\item{varnames}{vector like names_d}
}
\value{
vector same length as varnames, like c("sum of counts", "wtdmean")
}
\description{
utility - what type of formula is used to aggregate this variable?
}
\examples{
calctype(
)

}
\keyword{internal}
