% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_censusplaces.R
\docType{data}
\name{censusplaces}
\alias{censusplaces}
\title{censusplaces (DATA) Census FIPS and other basic info on 41,414 cities/towns/places}
\description{
This is just a list of US cities and similar places defined by Census
}
\details{
from (https://www2.census.gov/geo/docs/reference/codes/PLACElist.txt)
Column names:  "EPA_REGION" "STATE" "ST_FIPS" "COUNTY" "CO_FIPS" "PLACE" "PL_FIPS"
}
