% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanoutput.R
\name{cleanoutput}
\alias{cleanoutput}
\title{clean up output of fast batch tool? not used currently}
\usage{
cleanoutput(finaldata, filename)
}
\arguments{
\item{finaldata}{output of batch tool}

\item{filename}{what to save output as}
}
\description{
# this function is not currently used (yet?) -
  it would just make the outputs of a batch tool into a data.table, which it already should be I think,
  renames fields (colnames) to friendlier names,
  rounds values to specific numbers of significant digits,
  writes the output as a csv file locally
}
