% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcounter_summary.R
\name{colcounter_summary}
\alias{colcounter_summary}
\title{Summarize how many rows have N columns at or above (or below) various cutoffs?
Like colcounter or cols.above.count
but will handle multiple cutoffs to compare to each indicator, etc.
Table of counts, percents, cumulative counts, cumulative percents
of places with N, or at least N, of the indicators
at or above the benchmark(s)}
\usage{
colcounter_summary(
  x,
  cutofflist,
  or.tied = TRUE,
  na.rm = TRUE,
  below = FALSE,
  one.cut.per.col = FALSE
)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value,
like percentiles for example.}

\item{cutofflist}{vector of numeric cutoff values to compare to}

\item{or.tied}{if TRUE, include ties (value in x equals cutoff)}

\item{na.rm}{if TRUE, used by \code{\link[=colcounter]{colcounter()}} to count only the non-NA columns in given row}

\item{below}{if TRUE, count x below cutoff not above cutoff}

\item{one.cut.per.col}{if FALSE, compare each cutoff to all of x.
If TRUE, specify one cutoff to use for each column.}
}
\value{
A table of frequency counts
}
\description{
Summarize how many rows have N columns at or above (or below) various cutoffs?
Like colcounter or cols.above.count
but will handle multiple cutoffs to compare to each indicator, etc.
Table of counts, percents, cumulative counts, cumulative percents
of places with N, or at least N, of the indicators
at or above the benchmark(s)
}
\examples{
  \dontrun{
 pdata <- data.frame(a=rep(80,4),b=rep(93,4), col3=c(49,98,100,100))
  ### pdata <- EJAM::blockgroupstats[ , names_e_pctile]
  ## or ## pdata <- ejscreen::bg22[ , ejscreen::names.e.pctile]
 pcuts <-  5 * (0:20)  # <- as.vector(keystats_e['highcut', ])
colcounter_summary(        pdata, pcuts)
colcounter_summary_pct(    pdata, pcuts)
colcounter_summary_cum(    pdata, pcuts)
colcounter_summary_cum_pct(pdata, pcuts)
colcounter_summary_cum_pct(pdata, 5 * (10:20))
a3 <- colcounter_summary_all(    pdata, pcuts)

x80 <- colcounter(pdata, cutoff = 80, or.tied = T)
x95 <- colcounter(pdata, cutoff = 95, or.tied = T)
table(x95)
tablefixed(x95, NCOL(pdata))
cbind(at80=tablefixed(x80, NCOL(pdata)), at95=tablefixed(x95, NCOL(pdata)))
  }

}
\seealso{
\code{\link[=colcounter_summary_all]{colcounter_summary_all()}} \code{\link[=colcounter_summary]{colcounter_summary()}} \code{\link[=colcounter_summary_cum]{colcounter_summary_cum()}} \code{\link[=colcounter_summary_pct]{colcounter_summary_pct()}} \code{\link[=colcounter_summary_cum_pct]{colcounter_summary_cum_pct()}}

\code{\link[=tablefixed]{tablefixed()}}
}
