% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcounter_summary.R
\name{colcounter_summary_cum}
\alias{colcounter_summary_cum}
\title{Summarize how many rows have AT LEAST N columns at or above (or below) various cutoffs
See colcounter_summary() for more info and examples.}
\usage{
colcounter_summary_cum(
  x,
  cutofflist,
  or.tied = TRUE,
  na.rm = TRUE,
  below = FALSE,
  one.cut.per.col = FALSE
)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value,
like percentiles for example.}

\item{cutofflist}{vector of numeric cutoff values to compare to}

\item{or.tied}{if TRUE, include ties (value in x equals cutoff)}

\item{na.rm}{if TRUE, used by colcounter to count only the non-NA columns in given row}

\item{below}{if TRUE, count x below cutoff not above cutoff}

\item{one.cut.per.col}{if FALSE, compare each cutoff to all of x.
If TRUE, specify one cutoff to use for each column.}
}
\value{
A table of cumulative frequency counts
}
\description{
Summarize how many rows have AT LEAST N columns at or above (or below) various cutoffs
See colcounter_summary() for more info and examples.
}
\seealso{
colcounter_summary_all() colcounter_summary() colcounter_summary_cum() colcounter_summary_pct() colcounter_summary_cum_pct()
}
