% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcounter_summary.R
\name{colcounter_summary_pct}
\alias{colcounter_summary_pct}
\title{Summarize what percent of rows have N columns at or above (or below) various cutoffs}
\usage{
colcounter_summary_pct(x, cutofflist, ...)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to cutoff value,
like percentiles for example.}

\item{cutofflist}{vector of numeric cutoff values to compare to}

\item{...}{passed to colcounter_summary()
like or.tied=TRUE, na.rm=TRUE, below=FALSE, one.cut.per.col=FALSE}
}
\description{
Summarize what percent of rows have N columns at or above (or below) various cutoffs
}
\details{
See examples for colcounter_summary_cum_pct()
}
\seealso{
colcounter_summary_all() colcounter_summary() colcounter_summary_cum() colcounter_summary_pct() colcounter_summary_cum_pct()
}
