% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActualDistancefromSurfacedistance.R
\name{computeActualDistancefromSurfacedistance}
\alias{computeActualDistancefromSurfacedistance}
\title{convert surface distance to actual distance}
\usage{
computeActualDistancefromSurfacedistance(x)
}
\arguments{
\item{x}{surface distance in miles}
}
\description{
\preformatted{
    Just a simple formula:
   earthRadius_miles <- 3959
   angle_rad <- x/earthRadius_miles
   # Calculate  radius * cord length
   return( earthRadius_miles * 2*sin(angle_rad/2) )
   }
}
