% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sites_with_n_high_scores.R
\name{count_sites_with_n_high_scores}
\alias{count_sites_with_n_high_scores}
\title{count_sites_with_n_high_scores}
\usage{
count_sites_with_n_high_scores(
  scores,
  thresholds = c(1.01, 1.5, 2, 3, 5, 10),
  xwide = c("statewide", "nationwide")[1]
)
}
\arguments{
\item{scores}{scores}

\item{thresholds}{thresholds}

\item{xwide}{xwide}
}
\description{
Answers questions like What \verb{\%} of sites have at least 1 demog. indicator \verb{>2x} state avg?
}
\details{
Helps provide stat summaries such as:

(x\%) of these (sites) have

at least (N) of these (YTYPE )indicators

at least (R) times the (State/National average)
}
\examples{
x <- ejamit(testpoints_100, radius = 1)
out <- x$results_bysite
out <- setDF(copy(out))
ratio_benchmarks <- c(1.01, 1.50, 2, 3, 5, 10)
`ratiodata <- out[, names_d_ratio_to_state_avg]`

}
