% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_custom.R
\docType{data}
\name{custom}
\alias{custom}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 2.
}
\usage{
custom
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
