% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_from_aws.R
\name{data_load_from_aws}
\alias{data_load_from_aws}
\title{utility to load datasets from AWS}
\usage{
data_load_from_aws(
  fnames = c("lookup_states.rda", "bgid2fips.rda", "blockid2fips.rda", "blockwts.rda",
    "blockpoints.rda", "quaddata.rda"),
  envir = parent.frame(),
  mybucket = "dmap-data-commons-oa",
  mybucketfolder = "EJAM"
)
}
\arguments{
\item{fnames}{vector of the names of the rda files}

\item{envir}{default is parent.frame()}

\item{mybucket}{where in AWS}

\item{mybucketfolder}{where in AWS}
}
\value{
nothing - just loads data into environment
}
\description{
utility to load datasets from AWS
}
