% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_from_aws.R
\name{data_load_from_package}
\alias{data_load_from_package}
\title{utility to load a couple of datasets using data() if lazy loading not set up?}
\usage{
data_load_from_package()
}
\value{
Nothing
}
\description{
utility to load a couple of datasets using data() if lazy loading not set up?
}
\details{
These two datasets, blockgroupstats and frs, are some of the ones stored in the package as data()
}
