% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_aws.R
\name{dataload_from_aws}
\alias{dataload_from_aws}
\title{utility to load datasets from AWS}
\usage{
dataload_from_aws(
  fnames = c("bgid2fips.rda", "blockid2fips.rda", "blockwts.rda", "blockpoints.rda",
    "quaddata.rda"),
  envir = globalenv(),
  mybucket = "dmap-data-commons-oa",
  mybucketfolder = "EJAM",
  justchecking = FALSE
)
}
\arguments{
\item{fnames}{vector of the names of the rda files}

\item{envir}{default is parent.frame()}

\item{mybucket}{where in AWS}

\item{mybucketfolder}{where in AWS}

\item{justchecking}{set to TRUE to get object size (and confirm file is accessible/exists)}
}
\value{
nothing - just loads data into environment (unless justchecking=T)
}
\description{
utility to load datasets from AWS
}
\details{
See source code for details. Requires credentials.

Use dataload_from_aws(justchecking=TRUE),

or datapack("EJAM") to get info,

or tables(),

or object.size(quaddata)

List 9/2023 was:
\itemize{
\item blockid2fips (20 MB on disk, approx 600 MB RAM)
\item quaddata (168 MB on disk, 218 MB RAM)
\item blockpoints (86 MB on disk, 156 MB RAM)
\item blockwts (31 MB on disk, 125 MB RAM)
\item bgid2fips (18 MB RAM)

\if{html}{\out{<div class="sourceCode">}}\preformatted{   quaddata is >150 MB on disk and >200 MB RAM, while all others are smaller on disk.
   
   blockid2fips is roughly 600 MB in RAM because it stores 8 million block FIPS as text.
}\if{html}{\out{</div>}}
}
}
\seealso{
\code{\link[=datapack]{datapack()}} \code{\link[=dataload_from_aws]{dataload_from_aws()}} \code{\link[=dataload_from_package]{dataload_from_package()}} \code{\link[=indexblocks]{indexblocks()}} \code{\link[=.onAttach]{.onAttach()}}
}
