% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_pins.R
\name{dataload_from_pins}
\alias{dataload_from_pins}
\title{dataload_from_pins - download / load datasets from pin board}
\usage{
dataload_from_pins(
  varnames = c("blockwts", "quaddata", "blockpoints", "blockid2fips", "bgid2fips",
    "bgej"),
  boardfolder = "Mark",
  auth = "auto",
  server = "https://rstudio-connect.dmap-stage.aws.epa.gov",
  envir = globalenv(),
  justchecking = FALSE
)
}
\arguments{
\item{varnames}{character vector of names of R objects to get from board}

\item{boardfolder}{if needed to specify a different folder than default}

\item{auth}{See help documentation for \code{\link[pins:board_connect]{pins::board_connect()}}}

\item{server}{if needed to specify a server other than default (which might be
stored in envt variable CONNECT_SERVER or be registered via the rsconnect package).
Note if auth = "envvar" then it looks for CONNECT_SERVER to get name of server which
needs to be the full url starting with https:// - see help for board_connect}

\item{envir}{if needed to specify environment other than default}

\item{justchecking}{can set to TRUE to just see a list of what pins are stored in that board}
}
\value{
a vector of names of objects downloaded if justchecking = FALSE, which excludes those
already in environment so not re-downloaded and excludes those not found in pin board.
If justchecking = TRUE, returns vector of names of ALL objects found in pin board,
regardless of whether they are already in the environment, and
regardless of whether they were specified among varnames.
}
\description{
dataload_from_pins - download / load datasets from pin board
}
\details{
This does work:

board <- pins::board_connect(auth = "rsconnect")

assuming that credentials are set up for the user doing this.

This does work:

board <- pins::board_connect(auth = 'manual', server = Sys.getenv("CONNECT_SERVER"), key = Sys.getenv("CONNECT_API_KEY"))

if   Sys.setenv(CONNECT_SERVER = "https://rstudio-connect.dmap-stage.aws.epa.gov")

and if  CONNECT_API_KEY  was set to the API key created already.
}
