% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_urlpins.R
\name{dataload_from_urlpins}
\alias{dataload_from_urlpins}
\title{Utility to download / load datasets from pin board in pkgdown github pages of public repo}
\usage{
dataload_from_urlpins(
  varnames = c("blockwts", "blockpoints", "blockid2fips", "quaddata", "bgej",
    "bgid2fips", "frs", "frs_by_programid", "frs_by_naics", "frs_by_sic",
    "frs_by_mact")[1:4],
  server = "https://ejanalysis.github.io/ejscreendata/pins/",
  envir = globalenv(),
  justchecking = FALSE
)
}
\arguments{
\item{varnames}{vector of names of data objects stored in the pins board}

\item{server}{do not change}

\item{envir}{do not change default of globalenv()}

\item{justchecking}{set to TRUE if you just need to see what is stored in the pins board}
}
\value{
subset of varnames that is a vector of the ones successfully assigned to specified environment
}
\description{
Utility to download / load datasets from pin board in pkgdown github pages of public repo
}
