% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapack.R
\name{datapack}
\alias{datapack}
\title{Print list of data sets in package(s)
Wrapper for data(), for convenience}
\usage{
datapack(pkg = NULL, len = 30)
}
\arguments{
\item{pkg}{a character vector giving the package(s) to look in for data sets}

\item{len}{number of characters to limit Title to, making it easier to see in the console.}
}
\value{
data.frame with Item and Title as columns
}
\description{
Print list of data sets in package(s)
Wrapper for data(), for convenience
}
