% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetlist.R
\name{datasetlist}
\alias{datasetlist}
\title{see names of datasets a package offers}
\usage{
datasetlist(pkg)
}
\arguments{
\item{pkg}{character string name of an installed package}
}
\value{
vector of character string names of what is seen by data(package=pkg)
}
\description{
see names of datasets a package offers
}
\examples{
 datasetlist("datasets")
 datasetlist("MASS")
}
