% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrite_to_local.R
\name{datawrite_to_local}
\alias{datawrite_to_local}
\title{datawrite_to_local
Write large object(s) like EJAM datasets to local disk for convenience during app/pkg development, formatted as .arrow or .rda}
\usage{
datawrite_to_local(
  varnames = c("bgid2fips", "blockid2fips", "blockpoints", "blockwts", "quaddata"),
  ext = c(".arrow", ".rda")[1],
  folder_local_source = "~/../Downloads",
  fun = c("arrow::write_ipc_file", "save")[1],
  justchecking = F,
  overwrite = FALSE
)
}
\arguments{
\item{varnames}{vector of object names}

\item{ext}{file .extension appropriate to the format, ".rda" or ".arrow"}

\item{folder_local_source}{path to local folder without slash at end}

\item{fun}{function to use, but as a character string, like "arrow::write_ipc_file"
but fun is ignored if ext=".rda" since it then just uses save()}

\item{justchecking}{set this to FALSE to actually save instead of
just seeing in console info or the commands to be used, to test/check this}

\item{overwrite}{Set to TRUE to overwrite file if it exists already, with new copy.}
}
\value{
the paths of the objects as requested to be saved whether or not actually done
}
\description{
datawrite_to_local
Write large object(s) like EJAM datasets to local disk for convenience during app/pkg development, formatted as .arrow or .rda
}
\examples{
  # datawrite_to_local(ext = ".arrow", folder_local_source = ".", justchecking = F, overwrite = T) 
}
\seealso{
\code{\link[=datawrite_to_aws]{datawrite_to_aws()}}  \code{\link[=datawrite_to_local]{datawrite_to_local()}} \code{\link[=dataload_from_local]{dataload_from_local()}} \code{\link[=dataload_from_aws]{dataload_from_aws()}}
}
