% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{dependencies_of_ejam}
\alias{dependencies_of_ejam}
\title{utility for developing package - recursive dependencies of a package
Reminder of ways to check this is printed to console.}
\usage{
dependencies_of_ejam(
  localpkg = "EJAM",
  depth = 6,
  ignores_grep = "0912873410239478"
)
}
\arguments{
\item{localpkg}{"EJAM"}

\item{depth}{see deepdep package function deepdep()}

\item{ignores_grep}{see deepdep package function deepdep()}
}
\description{
utility for developing package - recursive dependencies of a package
Reminder of ways to check this is printed to console.
}
\keyword{internal}
