% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directsums.R
\name{directsums}
\alias{directsums}
\title{Weighted sum of counts of indicators, used to roll up population counts?}
\usage{
directsums(data, fieldnames = "POP100", scaling = 1)
}
\arguments{
\item{data}{data.table with environmental and/or demographic indicators. Just total population is default.}

\item{fieldnames}{vector of column names in data}

\item{scaling}{multipliers like 1 or 100 to put into correct units}
}
\description{
This function seems to just find the scaled and weighted sum(s) of field(s),
 using scaling value and scoringweight both multiplied by value of field in data.
 But it is unclear why the code uses  scoringweight  and what that is, where it is defined.
 Also, this seems to use only POP100 as the only fieldname in data.
}
