% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_by_groups_by_site-DRAFT.R
\name{distance_by_group_by_site}
\alias{distance_by_group_by_site}
\title{distance_by_group_by_site - DRAFT FUNCTION}
\usage{
distance_by_group_by_site(bybg)
}
\arguments{
\item{bybg}{such as \code{\link[=ejamit]{ejamit()}}$results_bybg_people}
}
\value{
table of ratios, one col per site, one row per indicator
}
\description{
distance_by_group_by_site - DRAFT FUNCTION
}
\seealso{
\code{\link[=plot_distance_cdf_by_group]{plot_distance_cdf_by_group()}} \code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}} \code{\link[=distance_by_groups]{distance_by_groups()}} \code{\link[=distance_mean_by_group]{distance_mean_by_group()}}
}
