% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{distance_by_group_plot}
\alias{distance_by_group_plot}
\title{distance_by_group_plot or plot_distance_cdf_by_group}
\usage{
distance_by_group_plot(...)
}
\value{
see returnwhat parameter
}
\description{
distance_by_group_plot or plot_distance_cdf_by_group
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plot(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=distance_by_group]{distance_by_group()}} \code{\link[=ejamit]{ejamit()}} for examples
}
