% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_comparison_ejamit.R
\name{distance_comparison_ejamit}
\alias{distance_comparison_ejamit}
\title{compare ejamit results for more than one radius
run ejamit() once per radius, get a table with a row per radius}
\usage{
distance_comparison_ejamit(
  sitepoints,
  radii = c(1, 2, 3),
  quiet = TRUE,
  silentinteractive = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{sitepoints}{like for \code{\link[=ejamit]{ejamit()}}}

\item{radii}{vector of radius values like 1:3 for \code{\link[=ejamit]{ejamit()}}}

\item{quiet}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{silentinteractive}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{...}{passed to \code{\link[=ejamit]{ejamit()}}}
}
\value{
data.table like ejamit()$results_overall but with one row
per radius.
}
\description{
compare ejamit results for more than one radius
run ejamit() once per radius, get a table with a row per radius
}
\examples{
  radii <- c(1,2,3,6,10)
  radii <- c(1, 10)
  pts <- testpoints_100
  pts <- testpoints_10
  x <- distance_comparison_ejamit(pts, radii = radii)
  EJAM:::distance_comparison_ejamit2plot(x)
  
}
\seealso{
\code{\link[=plot_distance_by_pctd]{plot_distance_by_pctd()}}, \code{\link[=distance_by_group]{distance_by_group()}} and several other related functions
}
