% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{distance_mean_by_group}
\alias{distance_mean_by_group}
\alias{distance_by_group}
\title{Avg distance of each demog group (of multiple groups) - Table or Plot}
\usage{
distance_mean_by_group(
  results_bybg_people,
  demogvarname = NULL,
  demoglabel = fixcolnames(demogvarname, "r", "shortlabel"),
  returnwhat = "table",
  graph = FALSE
)
}
\arguments{
\item{results_bybg_people}{data.table from \code{doaggregate()$results_bybg_people}}

\item{demogvarname}{vector of column names like "pctlowinc" etc.}

\item{demoglabel}{vector of labels like "Low Income Residents" etc.}

\item{returnwhat}{If returnwhat is "table", invisibly returns a
data.frame with group, ratio, avg_distance_for_group, avg_distance_for_nongroup.
If returnwhat is "plotfilename" then it returns the full path including filename of a .png in a tempdir
If returnwhat is "plot" then it returns the plot object as needed for table_xls_format() ?}
}
\value{
see parameter returnwhat
}
\description{
distance_mean_by_group() and distance_mean_by_group() are synonymous, and are like
\code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}}, but show a table not plot, by default.
}
\details{
see examples in \code{\link[=plot_distance_cdf_by_group]{plot_distance_cdf_by_group()}}
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plot(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=distance_by_group]{distance_by_group()}} \code{\link[=distance_by_group_plot]{distance_by_group_plot()}}  \code{\link[=plot_distance_cdf_by_group]{plot_distance_cdf_by_group()}}
}
