% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doaggregate_old.R
\name{doaggregate_old}
\alias{doaggregate_old}
\title{Summarize indicators in each buffer (given the blocks in each buffer and indicators for each block)}
\usage{
doaggregate_old(facilities, facilityblocks)
}
\arguments{
\item{facilities}{data.table of facility locations}

\item{facilityblocks}{data.table of blocks inside buffer around any given facility}
}
\description{
This takes a set of facilities and the set of blocks that are near each,
  (as identified previously, in other code that has identified which blocks are nearby)
  and combines those with indicator scores for block groups.

  It aggregates the scores as population weighted means, (check that - all of them?)
  creating a summary of each indicator, in each buffer (i.e., near each facility).
}
\details{
\preformatted{

 **** As currently written, this is hardcoded for specific indicators as the columns, indicators like pctlowinc.
 **** It should be made generic.

  The code USES SOME VARIABLES THAT WERE ALREADY PUT INTO THE GLOBAL ENVIRONMENT
   SO THAT THEY ARE AVAILABLE FROM WITHIN ALL FUNCTIONS WITHOUT HAVING TO PASS THE DATA TO EACH FUNCTION THAT NEEDS IT.
  It sets,  env <- globalenv()
  ### The global environment .GlobalEnv, more often known as the user's workspace, is the first item on the search path.
  It can also be accessed by globalenv().

   Check if it includes Puerto Rico (no, that was dropped)
 }
}
