% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Set up EJAM (do slow initialization steps when package is attached)}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{na}

\item{pkgname}{na}
}
\description{
Download datasets, load to memory, index block locations
}
\details{
This uses \code{\link[=dataload_from_pins]{dataload_from_pins()}} and \code{\link[=indexblocks]{indexblocks()}}

This function \code{.onAttach()} gets run when the package EJAM is attached,
which happens when library(EJAM) or require(EJAM) is used.

This code would not get run if a server ran app.R as a regular shiny app
and just used dataload or source to read the /R/*.R source files
rather than loading and attaching the EJAM package. see app.R   ***
Note this duplicates some code in global.R,
see source code here to adjust settings.
}
