% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_onLoad_create_quad_tree_index.R
\name{.onLoad}
\alias{.onLoad}
\title{Creates index to all US blocks (internal point lat lon) at package load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{}

\item{pkgname}{}
}
\description{
Creates index to all US blocks (internal point lat lon) at package load
}
