% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_latest_arrow_data.R
\name{download_latest_arrow_data}
\alias{download_latest_arrow_data}
\title{Download latest versions of arrow datasets if user doesn't have them}
\usage{
download_latest_arrow_data(varnames = .arrow_ds_names, envir = globalenv())
}
\arguments{
\item{varnames}{use defaults, or vector of names like "bgej" or use "all" to get all available}

\item{envir}{if needed to specify environment other than default, e.g., globalenv() or parent.frame()}
}
\description{
Used when EJAM is loaded
}
