% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupeRfiles.R
\name{dupeRfiles}
\alias{dupeRfiles}
\title{dupeRfiles
utility for developer to find similarly named .R files in 2 separate folders of source code,
whether or not they are packages.}
\usage{
dupeRfiles(folder1 = "./R", folder2 = "../EJAM/R")
}
\arguments{
\item{folder1}{path}

\item{folder2}{path}
}
\description{
see \code{\link[=dupenames]{dupenames()}} for when they are all packages.
}
