% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_distances.R
\name{ejam2barplot_distances}
\alias{ejam2barplot_distances}
\title{plot indicators as a function of distance from point(s)
plot results of ejamit_compare_distances()}
\usage{
ejam2barplot_distances(
  results_bydistance,
  myvars = names_d_subgroups_ratio_to_state_avg,
  ylab = "Ratio of Avg. within X miles to Avg. Statewide or Nationwide",
  ylim = c(0, 5),
  n = 1
)
}
\arguments{
\item{results_bydistance}{output of \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}, table similar to
ejamit()$results_overall except it has one row per distance.}

\item{myvars}{optional, see \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}}

\item{ylab}{optional, see \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}}

\item{ylim}{optional, see \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}}

\item{n}{optional, see \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}}
}
\value{
text vector length n, naming which indicators most strongly
increase as you get closer to the site(s)
}
\description{
plot indicators as a function of distance from point(s)
plot results of ejamit_compare_distances()
}
