% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_from_ejam.R
\name{ejam2excel}
\alias{ejam2excel}
\title{ejam2excel - alias for ejam2excel()}
\usage{
ejam2excel(...)
}
\arguments{
\item{...}{optional additional parameters passed to \code{\link[=table_xls_format]{table_xls_format()}}, such as
heatmap_colnames, heatmap_cuts, heatmap_colors, etc.}
}
\value{
returns a workbook object for use by openxlsx::saveWorkbook(wb_out, pathname)
or returns just the full path/file name of where it was saved if save_now = TRUE
}
\description{
ejam2excel - alias for ejam2excel()
}
\examples{
\dontrun{
  table_xls_from_ejam(testoutput_ejamit_10pts_1miles)
  }
}
