% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2map.R
\name{ejam2map}
\alias{ejam2map}
\title{Show EJAM results as a map of points}
\usage{
ejam2map(ejamitout, ...)
}
\value{
like what \code{\link[=mapfastej]{mapfastej()}} returns
}
\description{
Takes the output of ejamit() and uses \code{\link[=mapfastej]{mapfastej()}} to
create a map of the points.
}
\details{
Gets radius by checking ejamitout$results_overall$radius.miles
}
\examples{
# out = ejamit(testpoints_100, radius = 1)
out = testoutput_ejamit_100pts_1miles
x = ejam2map(out)
map2browser(x)
# to see it in the local browser instead of RStudio viewer pane
 
}
