% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit.R
\name{ejamit}
\alias{ejamit}
\title{Get complete EJ analysis numbers (demographic and environmental indicators) near a list of locations}
\usage{
ejamit(
  sitepoints,
  radius = 3,
  maxradius = 31.07,
  avoidorphans = TRUE,
  quadtree = NULL,
  ...
)
}
\arguments{
\item{sitepoints}{data.table with columns siteid, lat, lon giving point locations of sites or facilities around which are circular buffers}

\item{radius}{in miles, defining circular buffer around a site point}

\item{maxradius}{miles distance (max distance to check if not even 1 block point is within radius)}

\item{avoidorphans}{logical Whether to avoid case where no block points are within radius,
so if TRUE, it keeps looking past radius to find nearest one within maxradius.}

\item{quadtree}{(a pointer to the large quadtree object)
created from the SearchTree package example:
SearchTrees::createTree( quaddata, treeType = "quad", dataType = "point")
Takes about 2-5 seconds to create this each time it is needed.
It is automatically created when the package is loaded via the \code{\link[=.onLoad]{.onLoad()}} function}

\item{...}{passed to \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}
}
\value{
A list of tables of results.
}
\description{
This is the main function in EJAM for users who want to use EJAM from RStudio.
It does essentially what the webapp does to analyze/summarize near a set of points.
See help("EJAM")
}
\examples{
 # All in one step, using functions not shiny app:
 \dontrun{
 out <- ejamit(testpoints_100_dt, 2, quadtree=localtree)

 # Do not specify sitepoints and it will prompt you for a file,
 # if in RStudio in interactive mode!
 out <- ejamit(radius = 3)
 
  # Specify facilities or sites as points for test data, 
  # use 1000 test facility points from the R package 
  testsites <- testpoints_1000_dt
  # use facility points in an excel or csv file
  testsites <- latlon_from_anything(
   "./inst/testdata/testpoints_207_sites_with_signif_violations_NAICS_326_ECHO.csv")
  # use facility points from a random sample of EPA-regulated facilities
  testsites <- testpoints_n(1e3)  
  
  # Specify max distance from sites to look at (residents within X miles of site point)
  radius <- 3.1 # miles
  
  # Get summaries of all indicators near a set of points 
  out <- ejamit(testsites, radius)  
  # out <- ejamit("myfile.xlsx", 3.1)  
  
  # out2 <- EJAMejscreenapi::ejscreenit(testpoints_1000[1:3,])
  
  # View results overall
  round(t(out$results_overall), 3.1)
  
  # View plots
  # plot_distance_avg_by_group(out)  
  # plot_distance_cdf_by_group(out)
  
  # View maps
  mapfast(out$results_bysite, radius = 3.1)
  
  # view results at a single site
  t(out$results_bysite[1, ])
  t(out$results_bysite[out$results_bysite$siteid == 2, ])
  
  # if doing just 1st step of ejamit() 
  #  get distance between each site and every nearby Census block
  s2b <- testdata_sites2blocks
  s2b <- getblocksnearby(testsites, radius = radius)
  s2b <- getblocksnearbyviaQuadTree(testsites, radius = radius)
  getblocks_diagnostics(s2b)
  
  # if doing just 2d step of ejamit()
  #  get summaries of all indicators based on table of distances
  out <- doaggregate(s2b, testsites) # this works now and is simpler
  
}

}
\seealso{
\code{\link[=getblocksnearby]{getblocksnearby()}} \code{\link[=doaggregate]{doaggregate()}}
}
