% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_distances.R
\name{ejamit_compare_distances}
\alias{ejamit_compare_distances}
\title{Compare EJAM results overall for more than one radius
Run ejamit() once per radius, get a summary table with a row per radius}
\usage{
ejamit_compare_distances(
  sitepoints,
  radii = c(1, 2, 3),
  quiet = TRUE,
  silentinteractive = TRUE,
  plot = TRUE,
  myvars = names_d_subgroups_ratio_to_state_avg,
  ylab = "Ratio of Avg. within X miles to Avg. Statewide or Nationwide",
  ylim = c(0, 5),
  n = 1,
  ...
)
}
\arguments{
\item{sitepoints}{like for \code{\link[=ejamit]{ejamit()}}}

\item{radii}{vector of radius values like 1:3 for \code{\link[=ejamit]{ejamit()}}}

\item{quiet}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{silentinteractive}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{myvars}{for plot, see default value}

\item{ylab}{for plot, see default value}

\item{ylim}{for plot, see default value}

\item{n}{how many of the indicators to report on (printed to console),
when reporting which indicators most strongly increase as radius decreases.}

\item{...}{passed to \code{\link[=ejamit]{ejamit()}}}
}
\value{
data.table you can call results_bydistance,
like ejamit()$results_overall but with one row per radius
}
\description{
Compare EJAM results overall for more than one radius
Run ejamit() once per radius, get a summary table with a row per radius
}
\examples{
  radii <- c(1,2,3,6,10)
  radii <- c(1, 10)
  pts <- testpoints_100
  pts <- testpoints_10
  bydist <- ejamit_compare_distances(pts, radii = radii)
  ejamit_compare_distances2plot(bydist, myvars = c(
    "ratio.to.avg.pctlowinc", "ratio.to.avg.pcthisp", "ratio.to.avg.pctnhba"))
  
  names(bydist) <- fixcolnames(names(bydist), "r", "shortlabel")
  
}
\seealso{
\code{\link[=plot_distance_by_pctd]{plot_distance_by_pctd()}}, \code{\link[=distance_by_group]{distance_by_group()}},
and related internal functions
}
