% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_distances.R
\name{ejamit_compare_distances_fulloutput}
\alias{ejamit_compare_distances_fulloutput}
\title{Compare ejamit() full results for more than one radius
Run ejamit() once per radius, get FULL ejamit() output list per radius}
\usage{
ejamit_compare_distances_fulloutput(
  sitepoints,
  radii = c(1, 2, 3),
  quiet = TRUE,
  silentinteractive = TRUE,
  plot = TRUE,
  myvars = names_d_subgroups_ratio_to_state_avg,
  ylab = "Ratio of Avg. within X miles to Avg. Statewide or Nationwide",
  ylim = c(0, 5),
  n = 1,
  ...
)
}
\arguments{
\item{sitepoints}{like for \code{\link[=ejamit]{ejamit()}}}

\item{radii}{vector of radius values like 1:3 for \code{\link[=ejamit]{ejamit()}}}

\item{quiet}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{silentinteractive}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{myvars}{for plot, see default value}

\item{...}{passed to \code{\link[=ejamit]{ejamit()}}}
}
\value{
list you can call out_bydistance,
where each element is the full output of ejamit() for 1 radius
}
\description{
Compare ejamit() full results for more than one radius
Run ejamit() once per radius, get FULL ejamit() output list per radius
}
\examples{
  radii <- c(1,2,3,6,10)
  pts <- testpoints_10
  \dontrun{
  x <- ejamit_compare_distances_fulloutput(pts, radii = radii)
  }
}
\seealso{
\code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}
}
