% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_groups_of_places.R
\name{ejamit_compare_types_of_places}
\alias{ejamit_compare_types_of_places}
\title{Compare subsets (types) of places that are all from one list}
\usage{
ejamit_compare_types_of_places(sitepoints, typeofsite = NULL, ...)
}
\arguments{
\item{sitepoints}{see \code{\link[=ejamit]{ejamit()}}}

\item{typeofsite}{vector of length same as NROW(sitepoints), where each unique value defines a group of sites}

\item{...}{see \code{\link[=ejamit]{ejamit()}}}
}
\value{
similar to ejamit output but results_overall has one row per unique typeofsite
}
\description{
Compare subsets (types) of places that are all from one list
}
\examples{
  out = ejamit_compare_types_of_places(testpoints_10[1:4, ], 
  typeofsite = c("A", "B", "B", "C"))
  
}
