% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_see1map}
\alias{ejscreen_vs_ejam_see1map}
\title{EJAM/EJSCREEN comparisons - see map and tables of blocks near a site to explain discrepancy in pop and blocks}
\usage{
ejscreen_vs_ejam_see1map(n, x)
}
\arguments{
\item{n}{row number in x$EJAM of site to check}

\item{x}{results from  x <- ejscreen_vs_ejam(testpoints_10, radius =3, include_ejindexes = TRUE)}
}
\value{
map, tables are shown
}
\description{
EJAM/EJSCREEN comparisons - see map and tables of blocks near a site to explain discrepancy in pop and blocks
}
