% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_explain_summary}
\alias{ejscreen_vs_explain_summary}
\title{EJAM/EJSCREEN comparisons - Summarize explanations for discrepancies in pop}
\usage{
ejscreen_vs_explain_summary(whyall)
}
\arguments{
\item{whyall}{Output from \code{\link[=ejscreen_vs_explain]{ejscreen_vs_explain()}}}
}
\value{
summary table of info, on how many sites had each type of explanation
}
\description{
EJAM/EJSCREEN comparisons - Summarize explanations for discrepancies in pop
}
\examples{
vs10 <- ejscreen_vs_ejam(testpoints_10, radius = 3, save_ejscreen_output = F, save_when_report = F, calculate_ratios = F)
why10 <- ejscreen_vs_explain(vs10, ejam_uniq_id = 1:10, pause_on_each_site = F)
ejscreen_vs_explain_summary(why10)

}
\keyword{internal}
