% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam1}
\alias{ejscreenapi_vs_ejam1}
\title{compare EJScreen API vs EJAM stats near one site}
\usage{
ejscreenapi_vs_ejam1(latlon, radius = 3, nadrop = FALSE, ...)
}
\arguments{
\item{latlon}{1-row data.table or data.frame with colnames lat and lon}

\item{radius}{in miles, used in ejamit() and ejscreenapi_plus()}

\item{nadrop}{whether to drop indicators for which EJScreen API returns NA}

\item{...}{passed to ejamit() as any additional parameters,
like include_ejindexes = FALSE}
}
\value{
a data.frame with columns EJSCREEN and EJAM,
rownames are indicators like pop, blockcount_near_site, etc.
}
\description{
compare EJScreen API vs EJAM stats near one site
}
\examples{
 \dontrun{
   pts <- testpoints_100[1:5, ]
  #z <- ejscreenapi_vs_ejam1(testpoints_100[27, ], radius = 3, nadrop = T, include_ejindexes = TRUE)
  z <- ejscreenapi_vs_ejam1(pts[5, ], radius = 3, include_ejindexes = TRUE)
   
   # Reported key indicators - which ones do or don't match
   # when comparing EJSCREEN and EJAM results?
   keyreportnames <- c('pop', names_these, names_pctile, names_state_pctile)
   z[z$rname \%in\% keyreportnames &  z$same_shown, ]
   z[z$rname \%in\% keyreportnames & !z$same_shown & !is.na(z$EJSCREEN), ]
  
   # Reported (rounded) numbers match:
   z[z$same_shown , -1]
   # Reports disagree: 
   #  (and not just because of percentages being of 100 vs of 1.00)
   z[!z$same_shown & !is.na(z$EJSCREEN) & z$ratio != 0.01, -1] 
   # Reports disagree if percentages reported as 0-100 vs fractions 0-1.00
   z[z$ratio == 0.01 & !is.na(z$ratio), -1]
   
  }
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1_alreadyrun]{ejscreenapi_vs_ejam1_alreadyrun()}}
}
