% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam1_alreadyrun}
\alias{ejscreenapi_vs_ejam1_alreadyrun}
\title{compare EJScreen API vs EJAM stats near one site (if results already run)}
\usage{
ejscreenapi_vs_ejam1_alreadyrun(apisite, ejamsite, nadrop = FALSE)
}
\arguments{
\item{apisite}{1-row table output of ejscreenit()$table,
or ejscreenapi_plus()}

\item{ejamsite}{1-row table output of ejamit()$results_bysite}

\item{nadrop}{whether to drop indicators for which EJScreen API returns NA}
}
\value{
a data.frame with columns EJSCREENAPI and EJAM,
rownames are indicators like pop, blockcount_near_site, etc.
}
\description{
compare EJScreen API vs EJAM stats near one site (if results already run)
}
\examples{
 \dontrun{
   api1 <- ejscreenit(testpoints_100, radius = 3)
   api1 <- api1$table[5, ]
   ejam1 <- ejamit(testpoints_100, radius = 3)
   ejam1 <- ejam1$results_bysite[5, ]
   ejscreenapi_vs_ejam1_alreadyrun(api1, ejam1)
 }
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1]{ejscreenapi_vs_ejam1()}}
}
