% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam_see1}
\alias{ejscreenapi_vs_ejam_see1}
\title{see EJAM and EJSCREEN results for 1 site after using ejscreenapi_vs_ejam()}
\usage{
ejscreenapi_vs_ejam_see1(z, myvars = names_d, mysite = 1)
}
\arguments{
\item{z}{output of ejscreenapi_vs_ejam()}

\item{myvars}{optional to check just a subset of the colnames found in z$EJAM and z$EJSCREEN,
such as

myvars = colnames(z$EJAM) or

myvars = c(names_d, names_d_subgroups) or

myvars = grep("pctile", colnames(z$EJAM), value = T)}

\item{mysite}{rownumber corresponding to site of interest, among 1:nrow(z$EJAM)}
}
\value{
a table showing one row per indicator, and columns like EJSCREEN, EJAM, ratio, etc.
}
\description{
see EJAM and EJSCREEN results for 1 site after using ejscreenapi_vs_ejam()
}
\examples{
  dontrun{
  z <- ejscreenapi_vs_ejam(testpoints_10, radius = 3)
  mysite <- 9
  ejscreenapi_vs_ejam_see1(z, mysite = mysite, myvars = colnames(z$EJAM))[!is.na(z$EJSCREEN[mysite, ]) , 1:2]
  }
  
}
