% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam_summary}
\alias{ejscreenapi_vs_ejam_summary}
\title{see summary stats after using ejscreenapi_vs_ejam()}
\usage{
ejscreenapi_vs_ejam_summary(
  z,
  myvars = colnames(z$EJAM),
  tol = 0.01,
  na.rm = TRUE
)
}
\arguments{
\item{z}{output of ejscreenapi_vs_ejam()}

\item{myvars}{optional to check just a subset of the colnames found in z$EJAM and z$EJSCREEN,
such as myvars = c(names_d, names_d_subgroups) or myvars = grep("pctile", colnames(z$EJAM), value = T)}

\item{tol}{optional, set this so that results can be said to agree with this tolerance
if they differ by less than tol percent where tol is expressed as a fraction 0 to 1.}

\item{na.rm}{needs testing, optional, but}
}
\value{
A data.frame of summary stats showing counts and percents of analyzed sites (or those with valid data
that are found in both EJAM and EJSCREEN outputs), indicating how many of the sites
agree between EJSCREEN and EJAM estimates, exactly as reported or within some tolerance.
Columns include

"indicator" (variable name)

"sites.with.data.neither"

"sites.with.data.both"

"Rounded.Results.Agree"

"Results.within.tol"

"Rounded.Results.Agree.as.pct.of.sites.withdata"

"Results.within.tol.as.pct.of.sites.withdata"
}
\description{
see summary stats after using ejscreenapi_vs_ejam()
}
\examples{
  dontrun{
  pts <- testpoints_n(100, weighting = 'frs')
  vs100 <- ejscreenapi_vs_ejam(pts, radius = 3, include_ejindexes = TRUE)
  sum100 <- ejscreenapi_vs_ejam_summary(vs100, tol = 0.01)
  s100 <- sum100[ , c(1, 6:12)]
  
  s100[s100$indicator \%in\% names_e, ]
  s100[s100$indicator \%in\% names_d, ]
  s100[s100$indicator \%in\% names_these, ]
  s100[s100$indicator \%in\% c(names_ej_pctile, names_ej_state_pctile, names_ej_supp, names_ej_supp_state), ]
  
  sum100_within5pct <- ejscreenapi_vs_ejam_summary(vs100, tol = 0.05)
  sum100_within5pct[sum100_within5pct$indicator \%in\% names_these, ][ , c(1, 6:12)]
  
  }
}
