% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam_summary_quantiles}
\alias{ejscreenapi_vs_ejam_summary_quantiles}
\title{see quantiles over tested sites, of a stat like ratio of EJAM/EJSCREEN estimates}
\usage{
ejscreenapi_vs_ejam_summary_quantiles(
  z,
  mystat = c("ratio", "diff", "absdiff", "pctdiff", "abspctdiff")[1],
  myvars = c("pop", "", names_these),
  probs = (0:20)/20,
  na.rm = TRUE
)
}
\arguments{
\item{z}{output of ejscreenapi_vs_ejam()}

\item{mystat}{just one (not more than 1) of these: "ratio", "diff", "absdiff", "pctdiff", "abspctdiff"}

\item{myvars}{optional vector of indicators, to check just a subset of the colnames found in z$EJAM and z$EJSCREEN,
such as myvars = c(names_d, names_d_subgroups)
or myvars = grep("pctile", colnames(z$EJAM), value = T)}

\item{probs}{optional vector of probabilities 0 to 1 to pass to quantile()}

\item{na.rm}{optional, leave it as default TRUE}
}
\value{
table of percentiles across analyzed locations, of a stat like the ratio of
EJAM estimates over EJSCREEN estimates, for specified list of indicators like in names_d
}
\description{
see quantiles over tested sites, of a stat like ratio of EJAM/EJSCREEN estimates
}
