% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenit_for_ejam.R
\name{ejscreenit_for_ejam}
\alias{ejscreenit_for_ejam}
\title{ejscreenit_for_ejam
Wrapper for ejscreenit() from EJAMejscreenapi package, to use in EJAM app}
\usage{
ejscreenit_for_ejam(sitepoints, radius = 3, fillmissingcolumns = TRUE, ...)
}
\arguments{
\item{sitepoints}{table with lat and lon columns}

\item{radius}{in miles}

\item{...}{passed to ejscreenit()}
}
\value{
a data.table that looks like output of ejamit()$results_bysite
}
\description{
ejscreenit_for_ejam
Wrapper for ejscreenit() from EJAMejscreenapi package, to use in EJAM app
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1]{ejscreenapi_vs_ejam1()}}
}
