% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epa_programs.R
\docType{data}
\name{epa_programs}
\alias{epa_programs}
\title{epa_programs (DATA) named vector with program counts}
\description{
This is a named vector of EPA programs that lists the programs and has names with programs and program counts. Used by EJAM shiny app for dropdown menu.

@seealso \link{frs_by_programid}
}
