% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{fill_tbl_rowset_ej}
\alias{fill_tbl_rowset_ej}
\title{Write an EJ or EJ supp index VECTOR to an HTML table row}
\usage{
fill_tbl_rowset_ej(output_df, var_valueset = names_ej, var_nameset = names_e)
}
\arguments{
\item{output_df}{single row of results table from \code{\link[=ejamit]{ejamit()}} or
\code{\link[=doaggregate]{doaggregate()}} such as x$results_overall or one row like
x$results_bysite\link{1, }, where x is the result of x <- ejamit()}

\item{var_valueset}{character vector of raw EJ index variable names,
as in \link{names_ej}}

\item{var_nameset}{optional character vector of nicer names of
environmental indicators corresponding to var_valueset
(as a way to refer to EJ indexes)}
}
\value{
text string that is a fragment of HTML for use in fill_tbl_full_ej()
}
\description{
Write an EJ or EJ supp index VECTOR to an HTML table row
}
\seealso{
fill_tbl_row_ej()
}
\keyword{internal}
