% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips2countyname}
\alias{fips2countyname}
\title{Get county names from county FIPS codes}
\usage{
fips2countyname(fips, includestate = c("ST", "Statename", "")[1])
}
\arguments{
\item{fips}{vector of US Census FIPS codes for Counties (5 digits each). can be string or numeric, with or without leading zeroes.}

\item{includestate}{can be ST, Statename, "", or TRUE to specify what if anything comes after county name and comma}
}
\value{
vector of county names, optionally with comma and 2-character abbreviation or full state name.
}
\description{
Get county names from county FIPS codes
}
\examples{
  # names of all counties in ME and NY
  fips2countyname(fips_counties_from_state_abbrev(c("ME", "NY")), includestate = "ST")
   fips_counties_from_state_abbrev(c("AK", "LA"))
   
}
