% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips2name}
\alias{fips2name}
\title{fips2name - Get county or state names from county or state FIPS codes
Get county or state names from county or state FIPS codes}
\usage{
fips2name(fips, ...)
}
\arguments{
\item{fips}{vector of US Census FIPS codes for Counties (5 digits each) or States (2 digits).
Can be string or numeric, with or without leading zeroes.}

\item{...}{passed to fips2countyname() to control whether it appends something like , NY or , New York
after county name}
}
\value{
vector of state and/or county names,
where county names optionally have comma and 2-character abbreviation or full state name.
}
\description{
fips2name - Get county or state names from county or state FIPS codes
Get county or state names from county or state FIPS codes
}
\examples{
  fips2name(fips_counties_from_state_abbrev(c("AK", "LA"))  )
  fips2name(c(22, 02013))  # can have mix where some are a whole state and others are a county.
}
