% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips2statename}
\alias{fips2statename}
\title{fips2statename  -  get state names from Census FIPS codes}
\usage{
fips2statename(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector os state names
}
\description{
fips2statename  -  get state names from Census FIPS codes
}
\examples{
  cbind(fips_lead_zero(1:80), fips2state_abbrev(1:80), fips2statename(1:80))
}
