% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_bg_from_anyfips}
\alias{fips_bg_from_anyfips}
\title{FIPS - Get unique blockgroup fips in or containing specified fips of any type}
\usage{
fips_bg_from_anyfips(fips)
}
\arguments{
\item{fips}{vector of US FIPS codes, as character or numeric,
with or without their leading zeroes, each with as many characters}
}
\value{
vector of blockgroup FIPS (or NA values) that may be much longer than the
vector of fips passed to this function.
}
\description{
Convert any FIPS codes to the FIPS of all the blockgroups that are
among or within or containing those FIPS
}
\details{
This is a way to get a list of blockgroups, specified by state/county/tract or even block.

This function is not optimized for speed -- it is slow for large queries.

Takes a vector of one or more FIPS that could be State (2-digit), County (5-digit),
Tract (11-digit), or blockgroup (12 digit), or even block (15-digit fips).

Returns unique vector of FIPS of all US blockgroups (including DC and Puerto Rico)
that contain any specified blocks, are equal to any specified blockgroup fips,
or are contained within any provided tract/county/state FIPS.
}
\examples{

  # all blockgroups in one state
  fips_counties_from_state_abbrev("DE")
  fips_bg_from_anyfips( fips_counties_from_state_abbrev("DE") )

  blockgroupstats[,.N,by=substr(bgfips,1,2)]
  length(fips_bg_from_anyfips("72"))

  # all blockgroups in this one county
  fips_bg_from_anyfips(30001)
  fips_bg_from_anyfips("30001")
  fips_bg_from_anyfips(fips_counties_from_statename("Rhode Island")[1])

  # all blockgroups that contain any of these 6 blocks (i.e., just one bg)
  fips_bg_from_anyfips( blockid2fips$blockfips[1:6])

  # 2 counties
  fips_bg_from_anyfips(c(36009,36011))

}
\seealso{
\code{\link[=fips_lead_zero]{fips_lead_zero()}}
}
