% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_bgs_from_countyfips.R
\name{fips_bgs_from_countyfips}
\alias{fips_bgs_from_countyfips}
\title{fips_bgs_from_countyfips}
\usage{
fips_bgs_from_countyfips(fips)
}
\arguments{
\item{fips}{County FIPS vector (ideally as character not numeric values)}
}
\value{
data.table with all pairs of county fips - bgid, and a unique siteid assigned to each county
}
\description{
fips_bgs_from_countyfips
}
\examples{
 bg_from_county(c('01001','72153'))
 # Largest US Counties by ACS Population Totals:
 blockgroupstats[ , .(ST = ST[1], countypop = sum(pop)),
  by=.(FIPS = substr(bgfips,1,5))][order(-countypop),][1:20, .(
  CountyPopulation = prettyNum(countypop, big.mark = ","), FIPS, ST)]
 
}
