% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_counties_from.R
\name{fips_counties_from_statefips}
\alias{fips_counties_from_statefips}
\title{fips_counties_from_statefips}
\usage{
fips_counties_from_statefips(statefips)
}
\arguments{
\item{statefips}{vector of 2-digit state FIPS codes like c("10", "44", "44") or c(10,44)}
}
\value{
vector of 5-digit character string county FIPS of all unique counties in those states
}
\description{
fips_counties_from_statefips
}
\examples{
  fips_counties_from_statefips(c(10,44,44))
  fips_counties_from_statefips("10")
}
