% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_counties_from.R
\name{fips_counties_from_statename}
\alias{fips_counties_from_statename}
\title{fips_counties_from_statename}
\usage{
fips_counties_from_statename(statename)
}
\arguments{
\item{statename}{vector of state names like c("New York","Georgia")}
}
\value{
vector of 5-digit character string county FIPS of all unique counties in those states
}
\description{
fips_counties_from_statename
}
\examples{
fips_counties_from_statename("Delaware")
}
