% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_from_table.R
\name{fips_from_table}
\alias{fips_from_table}
\title{fips_from_table - read and clean FIPS column from a table, after inferring which col it is
Just read the codes in one column of a table obtained from something like read.csv, or excel, etc.}
\usage{
fips_from_table(fips_table, addleadzeroes = TRUE, inshiny = FALSE)
}
\arguments{
\item{fips_table}{data.frame or data.table of FIPS codes for counties, states, or tracts,
for example, in a column whose name can be interpreted as FIPS
(is one of the aliases like fips, countyfips, etc.)
Aliases are: c("FIPS", "fips", "fips_code", "fipscode", "Fips", "statefips",
"countyfips", "ST_FIPS", "st_fips", "ST_FIPS", "st_fips", "FIPS.ST",
"FIPS.COUNTY", "FIPS.TRACT")}

\item{addleadzeroes}{whether to add leading zeroes where needed as for a State whose FIPS starts with "01"}

\item{inshiny}{used by server during shiny app}
}
\value{
a vector of fips codes
}
\description{
fips_from_table - read and clean FIPS column from a table, after inferring which col it is
Just read the codes in one column of a table obtained from something like read.csv, or excel, etc.
}
\seealso{
\code{\link[=fips_bg_from_anyfips]{fips_bg_from_anyfips()}} \code{\link[=fips_lead_zero]{fips_lead_zero()}} \code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}} \code{\link[=fips_from_table]{fips_from_table()}}
}
