% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_place_from_placename}
\alias{fips_place_from_placename}
\title{search using names of cities to try to find matches and get FIPS}
\usage{
fips_place_from_placename(place_st, geocoding = FALSE, exact = TRUE)
}
\arguments{
\item{place_st}{vector of place names in format like "yonkers, ny" or "Chelsea city, MA"}

\item{geocoding}{set to TRUE to use a geocoding service to try to find hits}

\item{exact}{set to FALSE to allow partial matching but not necessarily 1-to-1 matches}
}
\value{
prints a table of possible hits but returns just the vector of fips
}
\description{
search using names of cities to try to find matches and get FIPS
}
