% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips_state_from_statename}
\alias{fips_state_from_statename}
\title{fips_state_from_statename}
\usage{
fips_state_from_statename(statename)
}
\arguments{
\item{statename}{vector of state names like c("New York","Georgia")}
}
\value{
vector of 2-digit state FIPS codes like c("10", "44", "44"),
same length as input, so including any duplicates
}
\description{
fips_state_from_statename
}
\examples{
  fips_state_from_statename("Delaware")
}
