% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fipstype}
\alias{fipstype}
\title{FIPS - Identify what type of Census geography is each FIPS code (block, county, etc.)}
\usage{
fipstype(fips)
}
\arguments{
\item{fips}{vector of one or more Census FIPS with or without leading zeroes, as strings or numeric}
}
\value{
vector of types: "block", "blockgroup", "tract", "county", or "state"
}
\description{
FIPS - Identify what type of Census geography is each FIPS code (block, county, etc.)
}
\details{
Note a number of length 11 is an ambiguous case - might be a tract fips
but might be a blockgroup fips with a missing leading zero.
}
\examples{
 testfips <- c("1", "12", "123", "1234", "12345", "", NA, "words")
 fipstype(testfips)
 
 testfips16 = c("1", "12", "123", "1234", "12345", "123456", "1234567", "12345678", "123456789",
                "1234567890", "12345678901", "123456789012", "1234567890123", "12345678901234",
                               "123456789012345", "1234567890123456")
 cbind(fipstype(testfips16), fips_lead_zero(testfips16), testfips16)

 fips_counties_from_statename(c("Connecticut", "Delaware") )
 # [1] "09001" "09003" "09005" "09007" "09009" "09011" "09013" "09015" "10001" "10003" "10005"
 fipstype(9001)
 fipstype("10001")
 # note blockid2fips is a large file, but can be obtained via [dataload_from_pins()]
 \dontrun{
 fipsexamples <- c(
   fips_state_from_statename("Alaska"),
   fips_counties_from_state_abbrev("DE")[1],
    "01001020100", # substr(blockgroupstats$bgfips[1],1,11),
    blockgroupstats$bgfips[1],
    "010010201001000" # blockid2fips$blockfips[1]
 )
 cbind(fipsexamples, type = fipstype(fipsexamples))
}

}
