% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_address.R
\name{fixcolnames_infer}
\alias{fixcolnames_infer}
\title{convert colnames to standardized names based on aliases}
\usage{
fixcolnames_infer(currentnames, alias_list = NULL)
}
\arguments{
\item{currentnames}{vector of colnames that may include aliases}

\item{alias_list}{optional named list where
names are standard colnames like "street"
and each named element in list is a vector of aliases
for that standard name}
}
\value{
vector like currentnames but some renamed to a
standard name if alias found, ignoring case.
}
\description{
Used by address_from_table
}
