% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt_table.R
\name{format_gt_table}
\alias{format_gt_table}
\title{format_gt_table}
\usage{
format_gt_table(
  df,
  type,
  my_cell_color = "#dce6f0",
  my_border_color = "#0070c0"
)
}
\arguments{
\item{df, }{a data frame with 6 columns (var_names, value, state_avg, state_pctile, usa_avg, usa_pctile), and one row per indicator}

\item{type, }{string - one of 'demog', 'envt'}

\item{my_cell_color, }{color for filling in background of table cells, can be given as string ('blue') or hex code ('#0070c0')}

\item{my_border_color, }{color for table borders and boundaries, can be given as string ('blue') or hex code ('#0070c0')}
}
\value{
a 'gt'-style table with formatting to closely match EJScreen standard report formatting
}
\description{
format_gt_table
}
