% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results_overall.R
\name{format_results_overall}
\alias{format_results_overall}
\title{format_results_overall
Format the results_overall part of the output of ejamit() or doaggregate()}
\usage{
format_results_overall(results_overall, longnames)
}
\arguments{
\item{results_overall}{data.table of 1 row, from output of ejamit() or doaggregate()}

\item{longnames}{vector of names of variables in results_overall,
from output of ejamit() or doaggregate()}
}
\value{
data.table that is one row per indicator
}
\description{
format_results_overall
Format the results_overall part of the output of ejamit() or doaggregate()
}
