% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_by_sic.R
\docType{data}
\name{frs_by_sic}
\alias{frs_by_sic}
\title{data.table of SIC code(s) for each EPA-regulated site in Facility Registry Service}
\description{
This is the format with one row per site-SIC pair,
so multiple rows for one site if it is in multiple SIC.
}
