% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_frs_clean.R
\name{frs_clean}
\alias{frs_clean}
\title{frs_clean - utility to clean Facility Registry Service dataset when updating it
This is just a helper function used to create the dataset saved as data for use in EJAM}
\usage{
frs_clean(
  frs,
  usefulcolumns = c("LATITUDE83", "LONGITUDE83", "REGISTRY_ID", "PRIMARY_NAME",
    "NAICS_CODES", "SIC_CODES", "PGM_SYS_ACRNMS")
)
}
\arguments{
\item{frs}{data.table that is the output of frs_read()}

\item{usefulcolumns}{optional, drops all columns except those in this vector of character colnames}
}
\description{
frs_clean - utility to clean Facility Registry Service dataset when updating it
This is just a helper function used to create the dataset saved as data for use in EJAM
}
\details{
Used by frs_get()
This renames some columns (lat, lon, NAICS are new names)
and it drops rows lacking lat/lon location info in the LATITUDE83 or LONGITUDE83 columns
}
