% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_siteid.R
\name{frs_from_naics}
\alias{frs_from_naics}
\title{Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_naics(naics_code_or_name, ...)
}
\arguments{
\item{naics_code_or_name}{}

\item{...}{passed to \code{\link[=naics_from_any]{naics_from_any()}}}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites
}
\examples{
  frs_from_naics("uranium")
  mapfast(frs_from_naics(naics_from_any("nuclear")$code))
  naics_from_any("silver")
  naics_from_name("silver")
  naics_from_any(212222 )
  frs_from_naics(21222)
  siteid_from_naics(21222)
  latlon_from_naics(21222)
  
}
\seealso{
\code{\link[=siteid_from_naics]{siteid_from_naics()}} \code{\link[=naics_from_any]{naics_from_any()}}
}
