% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_xyz.R
\name{frs_from_programid}
\alias{frs_from_programid}
\title{Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_programid(programname, programid)
}
\arguments{
\item{programname}{acronym for EPA program, as found in \code{epa_programs}}

\item{programid}{program system ID, such as "XJW000012435"}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites
}
\examples{
 test <- data.frame(programname = c('STATE','FIS','FIS'),
                    programid = c('#5005','0-0000-01097','0-0000-01103'))
 x = frs_from_programid(test$programname, test$programid)
 x
 mapfast(x)

}
