% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_from_any.R
\name{frs_from_sic}
\alias{frs_from_sic}
\title{frs_from_sic - Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_sic(sic_code_or_name, ...)
}
\arguments{
\item{...}{passed to \code{\link[=naics_from_any]{naics_from_any()}}}

\item{naics_code_or_name}{}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "SIC" "PGM_SYS_ACRNMS"
}
\description{
frs_from_sic - Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites
}
\examples{
  frs_from_sic("glass")
  mapfast(frs_from_sic(sic_from_any("silver")$code))
  sic_from_any("silver")
  sic_from_name("silver")
  sic_from_any('0780')
  frs_from_sic('0780')
  siteid_from_sic('0780')
  latlon_from_sic('0780')
  
}
\seealso{
\code{\link[=siteid_from_sic]{siteid_from_sic()}} \code{\link[=sic_from_any]{sic_from_any()}}
}
