% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_is_valid.R
\name{frs_is_valid}
\alias{frs_is_valid}
\title{Validate FRS uploads}
\usage{
frs_is_valid(frs_upload)
}
\arguments{
\item{frs_upload}{upload frs converted to data frame}
}
\value{
boolean value (valid or not valid)
}
\description{
Check for proper FRS facility id in uploaded data
}
