% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_make_programid_lookup.R
\name{frs_make_programid_lookup}
\alias{frs_make_programid_lookup}
\title{Reformat frs datatable to look up facilities by PGM_SYS_ACRNMS}
\usage{
frs_make_programid_lookup(x)
}
\arguments{
\item{x}{data.table frs from \code{\link[=frs_get]{frs_get()}}}
}
\value{
data.table with columns lat, lon, REGISTRY_ID, program, pgm_sys_id
}
\description{
Reformat frs datatable to look up facilities by PGM_SYS_ACRNMS
}
\seealso{
\code{\link[=frs_update_datasets]{frs_update_datasets()}}
}
