% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_html_header}
\alias{generate_html_header}
\title{generate_html_header - build HTML header for community report}
\usage{
generate_html_header(analysis_title, totalpop, locationstr, in_shiny = FALSE)
}
\arguments{
\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}
}
\description{
generate_html_header - build HTML header for community report
}
