% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getacs_epaquery.R
\name{get_any_rest_chunked_by_id}
\alias{get_any_rest_chunked_by_id}
\title{experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API}
\usage{
get_any_rest_chunked_by_id(objectIds, chunksize = 200, ...)
}
\arguments{
\item{objectIds}{see API}

\item{chunksize}{see API}

\item{...}{passed to getacs_epaquery()}
}
\value{
a table
}
\description{
experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API
}
