% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blockpoints_in_shape.R
\name{get_blockpoints_in_shape}
\alias{get_blockpoints_in_shape}
\title{find blocks that are in a polygon, using internal point of block - WORK IN PROGRESS ****}
\usage{
get_blockpoints_in_shape(polys, addedbuffermiles = 0)
}
\arguments{
\item{polys}{polygons (FORMAT???)}

\item{addedbuffermiles}{width of optional buffering to add to edges, in miles}
}
\value{
Polygons with joined block points info. **********
}
\description{
This is like getblocksnearby() but for a polygonal buffer area instead of
a circular buffer.
}
