% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_buffered.R
\name{get_shape_buffered}
\alias{get_shape_buffered}
\title{add buffer around shape}
\usage{
get_shape_buffered(shape, radius, ...)
}
\arguments{
\item{shape}{spatial object like areas at high risk or areas with facilities to be analyzed}

\item{radius}{width of buffer to add to shape.
(in case dist is a units object, it should be
convertible to arc_degree if x has geographic coordinates,
and to st_crs(x)$units otherwise)}

\item{...}{passed to st_buffer()}
}
\description{
add buffer around shape
}
\details{
Just a wrapper for \code{\link[sf:geos_unary]{sf::st_buffer()}}
}
