% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_buffered_from_shapefile_points.R
\name{get_shape_buffered_from_shapefile_points}
\alias{get_shape_buffered_from_shapefile_points}
\title{add buffer around shape (points, here)}
\usage{
get_shape_buffered_from_shapefile_points(shapefile_points, radius.miles, ...)
}
\arguments{
\item{shapefile_points}{spatial object like areas at high risk or areas with facilities to be analyzed}

\item{radius.miles}{width of buffer to add to shapefile_points
(in case dist is a units object, it should be
convertible to arc_degree if x has geographic coordinates,
and to st_crs(x)$units otherwise)}

\item{...}{passed to st_buffer()}
}
\description{
add buffer around shape (points, here)
}
\details{
Just a wrapper for \code{\link[sf:geos_unary]{sf::st_buffer()}}
}
\seealso{
\code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}} \code{\link[=get_shapefile_from_sitepoints]{get_shapefile_from_sitepoints()}} \link{get_shape_buffered_from_shapefile_points}
}
