% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_via_url.R
\name{get_via_url}
\alias{get_via_url}
\title{helper function work in progress: GET json via url of ejscreen ejquery map services}
\usage{
get_via_url(url)
}
\arguments{
\item{url}{the url for an EJScreen ejquery request}
}
\value{
json
}
\description{
helper function work in progress: GET json via url of ejscreen ejquery map services
}
