% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_summarize_blocks_per_site}
\alias{getblocks_summarize_blocks_per_site}
\title{getblocks_summarize_blocks_per_site - Get summary stats on counts of blocks near various sites}
\usage{
getblocks_summarize_blocks_per_site(x, varname = "siteid")
}
\arguments{
\item{x}{The output of getblocksnearby()}

\item{varname}{colname of variable in data.table x that is the one to summarize by}
}
\value{
invisibly, a list of stats
}
\description{
Tells you # of blocks near avg site,
how many sites have only 1 block nearby, or have <30 nearby, etc.
}
\seealso{
\code{\link[=getblocks_diagnostics]{getblocks_diagnostics()}}
}
