% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby.R
\name{getblocksnearby}
\alias{getblocksnearby}
\title{Key buffering function - wrapper redirecting to the right version of getblocksnearby()}
\usage{
getblocksnearby(
  sitepoints,
  cutoff = 1,
  maxcutoff = 31.07,
  avoidorphans = TRUE,
  quadtree,
  ...
)
}
\arguments{
\item{sitepoints}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{cutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{maxcutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{avoidorphans}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{quadtree}{a large quadtree object created from the SearchTree package example:
SearchTrees::createTree(EJAMblockdata::quaddata, treeType = "quad", dataType = "point")}

\item{...}{see \code{\link[=getblocksnearbyviaQuadTree_Clustered]{getblocksnearbyviaQuadTree_Clustered()}} or other such functions}
}
\description{
Key buffering function - wrapper redirecting to the right version of getblocksnearby()
}
\details{
As written it assumes that certain things are
already in global environment.
}
\seealso{
\code{\link[=getblocksnearby2]{getblocksnearby2()}} that was work in progress
}
