% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby2.R
\name{getblocksnearby2}
\alias{getblocksnearby2}
\title{Key buffering function - wrapper redirecting to the right version of getblocksnearby()
Like getblocksnearby() but tries to handle localtree and quadtree parameter differently - not sure how to check if they are in the right environment.}
\usage{
getblocksnearby2(
  sitepoints,
  cutoff = 1,
  maxcutoff = 31.07,
  avoidorphans = TRUE,
  quadtree = is.null,
  ...
)
}
\arguments{
\item{sitepoints}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{cutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{maxcutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{avoidorphans}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{quadtree}{a large quadtree object created from the SearchTree package example:
SearchTrees::createTree(EJAMblockdata::quaddata, treeType = "quad", dataType = "point")}

\item{...}{see \code{\link[=getblocksnearbyviaQuadTree_Clustered]{getblocksnearbyviaQuadTree_Clustered()}} or other such functions}
}
\description{
Key buffering function - wrapper redirecting to the right version of getblocksnearby()
Like getblocksnearby() but tries to handle localtree and quadtree parameter differently - not sure how to check if they are in the right environment.
}
\seealso{
\code{\link[=getblocksnearby]{getblocksnearby()}}
}
