% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby_and_doaggregate.R
\name{getblocksnearby_and_doaggregate}
\alias{getblocksnearby_and_doaggregate}
\title{Wrapper for getblocksnearby() plus doaggregate()}
\usage{
getblocksnearby_and_doaggregate(
  sitepoints,
  cutoff = 1,
  maxcutoff = 31.07,
  avoidorphans = TRUE,
  quadtree,
  ...
)
}
\arguments{
\item{sitepoints}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}

\item{cutoff}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}

\item{maxcutoff}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}

\item{avoidorphans}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}

\item{...}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}
}
\description{
Wrapper for getblocksnearby() plus doaggregate()
}
