% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby_from_fips.R
\name{getblocksnearby_from_fips}
\alias{getblocksnearby_from_fips}
\title{getblocksnearby_from_fips}
\usage{
getblocksnearby_from_fips(fips)
}
\arguments{
\item{fips}{vector of FIPS codes like for States or all counties in a State}
}
\value{
same as for \link{getblocksnearby} but one row per FIPS and the distance column is irrelevant
}
\description{
getblocksnearby_from_fips
}
\examples{
  x <- getblocksnearby_from_fips(fips_counties_from_state_abbrev("DE"))
  counties_ej <- doaggregate(x, sites2states_or_latlon = unique(x[ , .(siteid, lat, lon)]))
  mapfast(counties_ej$results_bysite)
  y =  ejamit(fips=fips_counties_from_statename("Delaware"))
  
}
