% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby_from_fips_OLD.R
\name{getblocksnearby_from_fips_OLD}
\alias{getblocksnearby_from_fips_OLD}
\title{getblocksnearby_from_fips
Actually finds all blocks within each of the FIPS codes provided}
\usage{
getblocksnearby_from_fips_OLD(fips, need_blockwt = TRUE)
}
\arguments{
\item{fips}{vector of FIPS codes identifying blockgroups, tracts, counties, or states.
This is useful if -- instead of gettings stats on and comparing circular buffers or polygons --
one will be getting stats on one or more tracts,
or analyzing and comparing blockgroups in a county,
or comparing whole counties to each other, within a State.}

\item{need_blockwt}{set to FALSE to speed it up if you do not need blockwt}
}
\value{
same as for \link{getblocksnearby} but one row per FIPS, and the distance column is irrelevant
}
\description{
getblocksnearby_from_fips
Actually finds all blocks within each of the FIPS codes provided
}
\examples{
  x <- getblocksnearby_from_fips(fips_counties_from_state_abbrev("DE"))
  counties_ej <- doaggregate(x, sites2states_or_latlon = unique(x[ , .(siteid, lat, lon)]))
  mapfast(counties_ej$results_bysite)
  y =  ejamit(fips=fips_counties_from_statename("Delaware"))
  
  # x=getblocksnearby_from_fips("482011000011") # one blockgroup only
  # y=doaggregate(x)
  
}
