% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearbyviaQuadTree_Clustered.R
\name{getblocksnearbyviaQuadTree_Clustered}
\alias{getblocksnearbyviaQuadTree_Clustered}
\title{find nearby blocks using Quad Tree data structure for speed, CLUSTERED FOR PARALLEL PROCESSING}
\usage{
getblocksnearbyviaQuadTree_Clustered(
  facilities,
  cutoff,
  maxcutoff,
  avoidorphans,
  CountCPU = 1
)
}
\arguments{
\item{facilities}{data.table with columns LAT, LONG}

\item{cutoff}{miles distance (check what this actually does)}

\item{maxcutoff}{miles distance (check what this actually does)}

\item{avoidorphans}{logical}

\item{CountCPU}{for parallel processing via makeCluster() and doSNOW::registerDoSNOW()}
}
\description{
Uses packages \link[pkg]{parallel} and snow. parallel::makePSOCKcluster is an enhanced version of snow::makeSOCKcluster in package snow.
    It runs Rscript on the specified host(s) to set up a worker process which listens on a socket for expressions to evaluate, and returns the results (as serialized objects).
}
\details{
Uses indexgridsize and quaddata  variables that come from global environment (but should pass to this function rather than assume in global env?)
}
\seealso{
\link{getblocksnearbyviaQuadTree}  \link{computeActualDistancefromSurfacedistance}
}
