% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_file_upload.R
\name{handle_file_upload}
\alias{handle_file_upload}
\title{Used in app_server.R to read in files into a data.table}
\usage{
handle_file_upload(file_input, file_type)
}
\arguments{
\item{file_input}{actual input file that's being inputed through the app}

\item{file_type}{the extension derived from the file path}
}
\value{
data.table derived from the input
}
\description{
This function replaces some of the repetitive code in app_server for reading in input data files. It checks what type of file it is then properly reads it into a data.table
}
