% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasfield.R
\name{hasfield}
\alias{hasfield}
\title{helper function}
\usage{
hasfield(data, fieldname)
}
\arguments{
\item{data}{data.table}

\item{fieldname}{colname to check}
}
\value{
logical
}
\description{
helper function
}
