% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_high_pctiles_tied_with_min.R
\docType{data}
\name{high_pctiles_tied_with_min}
\alias{high_pctiles_tied_with_min}
\title{high_pctiles_tied_with_min (DATA) internal data used to handle cases where multiple places are tied for the lowest indicator score}
\description{
high_pctiles_tied_with_min (DATA) internal data used to handle cases where multiple places are tied for the lowest indicator score
}
