% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_indexblocks.R
\name{indexblocks}
\alias{indexblocks}
\title{indexblocks
Create localtree (a quadtree index of all US block centroids) in global environment}
\usage{
indexblocks()
}
\value{
Side effect is it creates the index in memory
}
\description{
indexblocks
Create localtree (a quadtree index of all US block centroids) in global environment
}
\details{
.onLoad() can be edited to create this when the package loads,
but then it takes time each time a developer rebuilds/installs the package or others that load EJAM.
}
