% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_installejamsource.R
\name{installejamsource}
\alias{installejamsource}
\title{installejamsource
utility for when working on the source code - redoes documentation and reinstalls packages}
\usage{
installejamsource(pkg = EJAM::ejampackages)
}
\arguments{
\item{pkg}{vector of names of folders with source packages,
but not the full path - just the package name.
Assumes that R_USER points to the folder that has all those source package folders in it.}
}
\value{
nothing. side effect is it does document() and INSTALL
}
\description{
installejamsource
utility for when working on the source code - redoes documentation and reinstalls packages
}
