% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_islandareas.R
\docType{data}
\name{islandareas}
\alias{islandareas}
\title{islandareas table, bounds info on lat lon of US Island Areas}
\description{
data.frame of info on approximate lat lon bounding boxes around
American Samoa, Guam, the
Commonwealth of the Northern Mariana Islands (Northern Mariana Islands),
and the United States Virgin Islands.

See also \link{stateinfo} and []

See \url{http://www.census.gov/geo/reference/gtc/gtc_island.html}

See datacreate_islandareas.R or data-raw/islandareas.xlsx

Note the US minor outlying islands are not in that list and are widely dispersed.
They include Midway Islands, etc.
}
