% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lat_alias.R
\docType{data}
\name{lat_alias}
\alias{lat_alias}
\alias{lon_alias}
\title{lat_alias, lon_alias (DATA) Synonyms for lat and lon}
\description{
lists of synonyms for "latitude" and "longitude" used when guessing which column is what in user-provided tables of coordinates
}
