% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_aliases.R
\docType{data}
\name{lat_aliases, lon_aliases}
\alias{lat_aliases, lon_aliases}
\title{Synonyms for lat and lon}
\description{
lists of synonyms for "latitude" and "longitude" used when guessing which column is what in user-provided tables of coordinates
}
