% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_df_clean.R
\name{latlon_df_clean}
\alias{latlon_df_clean}
\title{Find and clean up latitude and longitude columns in a data.frame}
\usage{
latlon_df_clean(df)
}
\arguments{
\item{df}{data.frame With columns lat and lon or names that can be interpreted as such - see latlon_infer()}
}
\value{
Returns the same data.frame but with relevant colnames changed to lat and lon,
and invalid lat or lon values cleaned up if possible or else replaced with NA
}
\description{
Utility to identify lat and lon columns, renaming and cleaning them up.
}
\details{
Tries to figure out which columns seem to have lat lon values, renames those in the data.frame.
Cleans up lat and lon values (removes extra characters, makes numeric)
}
\examples{
#  x <- latlon_df_clean(x)
}
\seealso{
latlon_df_clean() latlon_infer() latlon_is.valid() latlon_as.numeric()
}
