% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_naics.R
\name{latlon_from_naics}
\alias{latlon_from_naics}
\title{latlon_from_naics - Find EPA-regulated facilities in FRS by NAICS code (industrial category)
Get lat lon, Registry ID, given NAICS industry code(s)
Find all EPA Facility Registry Service (FRS) sites with this exact NAICS code (not subcategories)}
\usage{
latlon_from_naics(naics, id_only = FALSE)
}
\arguments{
\item{naics}{a vector of naics codes, or
a data.table with column named code, as with output of \code{\link[=naics_from_any]{naics_from_any()}}}
}
\value{
A data.table (not just data.frame) with columns called
lat, lon, REGISTRY_ID, NAICS (but see the id_only parameter)
}
\description{
latlon_from_naics - Find EPA-regulated facilities in FRS by NAICS code (industrial category)
Get lat lon, Registry ID, given NAICS industry code(s)
Find all EPA Facility Registry Service (FRS) sites with this exact NAICS code (not subcategories)
}
\details{
NOTE: many FRS sites lack NAICS code!

Also, this function does not find the sites
identified by FRS data as being in a child NAICS (subcategory of your exact query)!

Relies on  frs_by_naics (a data.table)

See info about NAICS industry codes at \url{https://www.naics.com/search}
}
\examples{
  regid_from_naics(321114)
  latlon_from_naics(321114)
  latlon_from_naics(EJAM::naics_from_any("cheese")[,code] )
  head(latlon_from_naics(c(3366, 33661, 336611), id_only=TRUE))
  # mapfast(frs_from_naics(336611)) # simple map
}
