% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_programid.R
\name{latlon_from_programid}
\alias{latlon_from_programid}
\title{Get lat lon, Registry ID, and NAICS, for given FRS Program System ID}
\usage{
latlon_from_programid(programname, programid)
}
\arguments{
\item{programname}{acronym for EPA program, as found in \code{epa_programs}}

\item{programid}{program system ID, such as "XJW000012435"}
}
\value{
data.table with lat  lon  REGISTRY_ID  program   pgm_sys_id
}
\description{
Get lat lon, Registry ID, and NAICS, for given FRS Program System ID
}
\details{
The ID is the identification number, such as the permit number,
assigned by an information management system that represents a
facility site, waste site, operable unit, or other feature tracked by that
Environmental Information System.

Also note the FRS API:
\url{https://www.epa.gov/frs/facility-registry-service-frs-api}
\url{https://www.epa.gov/frs/frs-rest-services}
}
\examples{
 test <- data.frame(programname = c('STATE','FIS','FIS'),
                    programid = c('#5005','0-0000-01097','0-0000-01103'))
 latlon_from_programid(test$programname, test$programid)

 latlon_from_regid(latlon_from_programid(test$programname, test$programid)[,REGISTRY_ID])
 latlon_from_programid(test$programname, test$programid)[,.(lat,lon)]

}
