% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilaterate-DRAFT.R
\name{latlon_join_on_blockid}
\alias{latlon_join_on_blockid}
\title{latlon_join_on_blockid - get lat,lon of each block internal point via blockid
get expanded version of sites2blocks data.table, with new lat,lon columns}
\usage{
latlon_join_on_blockid(s2b)
}
\arguments{
\item{s2b}{like \link{testoutput_getblocksnearby_10pts_1miles}, output of \code{\link[=getblocksnearby]{getblocksnearby()}}}
}
\value{
returns the input data.table but with lat,lon columns added as block coordinates
}
\description{
latlon_join_on_blockid - get lat,lon of each block internal point via blockid
get expanded version of sites2blocks data.table, with new lat,lon columns
}
\examples{
 s2b = copy(testoutput_getblocksnearby_10pts_1miles)
 latlon_join_on_blockid(s2b) # done by trilaterate also
 
}
