% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_readclean.R
\name{latlon_readclean}
\alias{latlon_readclean}
\title{read csv or xlsx and then clean it
This function just wraps \code{\link[EJAMbatch.summarizer:read_csv_or_xl]{EJAMbatch.summarizer::read_csv_or_xl()}}
and \code{\link[=latlon_df_clean]{latlon_df_clean()}}}
\usage{
latlon_readclean(x)
}
\arguments{
\item{x}{character string, full path to a .csv or .xlsx file}
}
\value{
see \code{\link[=latlon_df_clean]{latlon_df_clean()}}
}
\description{
read csv or xlsx and then clean it
This function just wraps \code{\link[EJAMbatch.summarizer:read_csv_or_xl]{EJAMbatch.summarizer::read_csv_or_xl()}}
and \code{\link[=latlon_df_clean]{latlon_df_clean()}}
}
