% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.pctile.Region.R
\name{lookup.pctile.Region}
\alias{lookup.pctile.Region}
\title{old code to create a look up table to find percentile that a raw score is at  within EPA Region}
\usage{
lookup.pctile.Region(data, thisregion, ...)
}
\arguments{
\item{data}{passed to \link{lookup.pctile.US}}

\item{thisregion}{passed to \link{lookup.pctile.US}}

\item{...}{passed to \link{lookup.pctile.US}}
}
\value{
data.table of 100 percentiles and mean (in rows) for each indicator (in columns)
}
\description{
old code to create a look up table to find percentile that a raw score is at  within EPA Region
}
\details{
CURRENTLY HARD CODED TO USE THESE SPECIFIC FIELDS - NEEDS TO BE MADE GENERIC.
  This should be about the same as ejanalysis::lookup.pctile()
  using lookup tables from something like ejscreen::ejscreen.lookuptables()
  See user ejanalysis, ejscreen and ejanalysis packages on GitHub
}
