% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mact_categories.R
\docType{data}
\name{mact_categories}
\alias{mact_categories}
\title{MACT NESHAP subparts (the code and the description)}
\description{
This is a data.table with one row per MACT subpart such as BBBBBB or ZZZZZZ
and the title of each category. From the ECHO download of ICIS Air
@seealso \link{frs_by_mact}
}
