% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2map.R
\name{map2browser}
\alias{map2browser}
\title{quick way to open a map html widget in the local browser (by saving it as a tempfile)}
\usage{
map2browser(x)
}
\arguments{
\item{x}{output of \code{\link[=ejam2map]{ejam2map()}} or \code{\link[=mapfastej]{mapfastej()}} or \code{\link[=mapfast]{mapfast()}}}
}
\value{
launches local browser to show x, but also returns
name of tempfile that is the html widget
}
\description{
quick way to open a map html widget in the local browser (by saving it as a tempfile)
}
\examples{
x = ejam2map(testoutput_ejamit_100pts_1miles)
map2browser(x)
# to see it in the local browser instead of RStudio viewer pane
   
}
