% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_facilities.R
\name{map_facilities}
\alias{map_facilities}
\title{map_facilities}
\usage{
map_facilities(mypoints, rad = 3, highlight = FALSE, clustered)
}
\arguments{
\item{mypoints, }{data frame of uploaded points}

\item{rad, }{a size for drawing each circle (buffer search radius)}

\item{highlight, }{a logicial for whether to highlight overlapping points (defaults to FALSE)}

\item{clustered, }{a vector of T/F values for each point, indicating if they overlap with another}
}
\value{
a leaflet map with circles, circleMarkers, and basic popup
}
\description{
make a leaflet map of uploaded points
}
