% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_facilities_proxy.R
\name{map_facilities_proxy}
\alias{map_facilities_proxy}
\title{map_facilities_proxy}
\usage{
map_facilities_proxy(
  mymap,
  rad = 3,
  highlight = FALSE,
  clustered = FALSE,
  popup_vec = NULL
)
}
\arguments{
\item{mymap, }{leafletProxy map object to be added to}

\item{rad, }{a size for drawing each circle (buffer search radius)}

\item{highlight, }{a logicial for whether to highlight overlapping points (defaults to FALSE)}

\item{clustered, }{a vector of T/F values for each point, indicating if they overlap with another}

\item{popup_vec, }{a vector of popup values to display when points are clicked. Length should match number of rows in the dataset.}
}
\value{
a leaflet map with circles, circleMarkers, and basic popup
}
\description{
update a leaflet map within the EJAM shiny app with uploaded points.
}
