% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{map_shapes_leaflet}
\alias{map_shapes_leaflet}
\title{Map - Create a new leaflet map from shapefile data, in shiny app}
\usage{
map_shapes_leaflet(shapes, color = "green", popup = shapes$NAME)
}
\arguments{
\item{shapes}{like from shapes_counties_from_countyfips(fips_counties_from_state_abbrev("DE"))}

\item{color}{passed to leaflet::addPolygons()}

\item{popup}{passed to leaflet::addPolygons()}
}
\value{
html widget from leaflet::leaflet()
}
\description{
Map - Create a new leaflet map from shapefile data, in shiny app
}
