% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{map_shapes_leaflet_proxy}
\alias{map_shapes_leaflet_proxy}
\title{update existing leaflet map by adding shapefile data}
\usage{
map_shapes_leaflet_proxy(mymap, shapes, color = "green", popup = shapes$NAME)
}
\arguments{
\item{mymap}{map like from leafletProxy()}

\item{shapes}{like from shapes_counties_from_countyfips(fips_counties_from_state_abbrev("DE"))}

\item{color}{passed to leaflet::addPolygons()}

\item{popup}{passed to leaflet::addPolygons()}
}
\value{
html widget like from leaflet::leafletProxy()
}
\description{
update existing leaflet map by adding shapefile data
}
