% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{map_shapes_mapview}
\alias{map_shapes_mapview}
\title{map_shapes_mapview}
\usage{
map_shapes_mapview(shapes, col.regions = "green", map.types = "OpenStreetMap")
}
\arguments{
\item{shapes}{like from shapes_counties_from_countyfips()}

\item{col.regions}{passed to mapview() from mapview package}

\item{map.types}{passed to mapview() from mapview package}
}
\description{
map_shapes_mapview
}
