% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{map_shapes_plot}
\alias{map_shapes_plot}
\title{map_shapes_plot}
\usage{
map_shapes_plot(shapes, main = "Selected Census Units", ...)
}
\arguments{
\item{shapes}{like from shapes_counties_from_countyfips()}

\item{main}{title for map}

\item{...}{passed to plot()}
}
\value{
Just draws map using plot()
}
\description{
map_shapes_plot
}
