% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfast.R
\name{mapfast}
\alias{mapfast}
\title{Map - points - Create leaflet html widget map of points using table with lat lon}
\usage{
mapfast(
  mydf,
  radius = 3,
  column_names = "all",
  labels = column_names,
  browse = FALSE,
  color = "#03F"
)
}
\arguments{
\item{mydf}{data.frame or data.table with lat and lon columns or
columns that \code{\link[=latlon_infer]{latlon_infer()}} can infer to be that}

\item{radius}{in miles, converted to meters and passed to leaflet::addCircles()}

\item{column_names}{If "ej" then nice popup made based on just key EJScreen
indicators. If "all" then every column in the entire mydf table is shown
in the popup. If a vector of colnames, only those are shown in popups.}

\item{labels}{The labels used before the column_names, for map popups,
like  label: column_name  (ignored if column_names is ej or all)}

\item{browse}{optional logical, set to TRUE if you want the function to
launch a default browser window to show the map
and print the temp filepath and filename in the console.
Normally the map would be shown in the default RStudio viewer pane.}

\item{color}{color of circles, default is "#03F"}
}
\value{
plots a leaflet map with popups with all the columns from mydf,
and returns html widget
}
\description{
Map - points - Create leaflet html widget map of points using table with lat lon
}
\examples{
pts = testpoints_100
mapfast(pts)

# out = ejamit(pts, radius = 1)
out = testoutput_ejamit_100pts_1miles

# See in RStudio viewer pane
ejam2map(out)
mapfastej(out$results_bysite[c(12,31),])
\dontrun{

# See in local browser instead
ejam2map(out, browse = T)

# Open folder where interactive map
#  .html file is saved, so you can share it:
x = ejam2map(out)
fname = map2browser(x)
# browseURL(dirname(fname)) # to open the temp folder
# file.copy(fname, "./map.html") # to copy map file to working directory

out <- testoutput_ejscreenapi_plus_50
mapfastej(out)
}
}
\seealso{
\code{\link[=ejam2map]{ejam2map()}} \code{\link[=popup_from_any]{popup_from_any()}} \code{\link[=mapfastej]{mapfastej()}}
}
