% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfast.R
\name{mapfastej}
\alias{mapfastej}
\title{Get a leaflet html widget map of data.frame with lat lon

Like \code{\link[=mapfast]{mapfast()}} but with column_names = "ej"}
\usage{
mapfastej(...)
}
\description{
Get a leaflet html widget map of data.frame with lat lon

Like \code{\link[=mapfast]{mapfast()}} but with column_names = "ej"
}
