% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefileops.R
\name{merge_state_shapefiles}
\alias{merge_state_shapefiles}
\title{Spatial overlay of facilities points and US States shapefiles to add STATE as column in facility table}
\usage{
merge_state_shapefiles(facs, shapefile)
}
\arguments{
\item{facs}{facilities LONG LAT table}

\item{shapefile}{shapefile of States}
}
\value{
Returns the facs that was passed to this function, but with a new column, STATE, that has the name of the State each point is inside of
}
\description{
Spatial overlay of facilities points and US States shapefiles to add STATE as column in facility table
}
