% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS.R
\docType{data}
\name{NAICS}
\alias{NAICS}
\title{named list of all NAICS code numbers and industry name for each}
\description{
named list of all NAICS code numbers and industry name for each
}
\details{
see \link{https://naics.com}
}
\seealso{
\link{naicstable} \code{\link[=naics_from_any]{naics_from_any()}} \code{\link[=naics_categories]{naics_categories()}} \link{NAICS}
}
