% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_find.R
\name{naics2children}
\alias{naics2children}
\title{See NAICS codes queried plus all children of any of those
Used by NAICS_find()}
\usage{
naics2children(codes, allcodes = EJAM::NAICS)
}
\arguments{
\item{codes}{vector of numerical or character}

\item{allcodes}{Optional (already loaded with package) - dataset with all the codes}
}
\value{
vector of codes and their names
}
\description{
See NAICS codes queried plus all children of any of those
Used by NAICS_find()
}
\details{
start with shortest (highest level) codes. since tied for nchar, these branches have zero overlap, so do each.
for each of those, get its children = all rows where parentcode == substr(allcodes, 1, nchar(parentcode))
put together list of all codes we want to include so far.
now for the next longest set of codes in original list of codes,
do same thing.
etc. until did it for 5 digit ones to get 6digit children.
take the unique(allthat)
table(nchar(as.character(NAICS)))
2    3    4    5    6
17   99  311  709 1057
}
\seealso{
NAICS_find() NAICS
}
