% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naics_counts.R
\docType{data}
\name{naics_counts}
\alias{naics_counts}
\title{naics_counts (DATA) data.frame with counts for each NAICS code, with and without subcodes, and labels that include the site counts}
\description{
This is a data frame containing 2200 NAICS codes (as of 3/2024), the count of sites for each of them in the frs data, both on their own and including all subcodes. Used by EJAM shiny app for dropdown menu.

@seealso \link{frs_by_naics}
}
