% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_find.R
\name{NAICS_find}
\alias{NAICS_find}
\title{Search for an industrial sector in the list of NAICS codes, see subsectors}
\usage{
NAICS_find(
  query,
  matchonly = FALSE,
  dataset = ifelse(exists(dataset), dataset, NULL),
  ignore.case = TRUE
)
}
\arguments{
\item{query}{a single word or phrase such as "chemical manufacturing" or "cement"}

\item{matchonly}{default is FALSE which shows}

\item{dataset}{Should default to the dataset NAICS, installed with this package. see \link{NAICS}}

\item{ignore.case}{default TRUE, ignoring whether query is upper or lower case}
}
\description{
Just a utility, quick way to view NAICS industrial sectors that contain queried word or phrase,
but can also see all the subcategories within the matching one.
}
\details{
See \url{https://www.census.gov/naics/}
 NOTE: By default, this shows the highest level category (smallest number of digits in NAICS, like 2 or 3)
 that matches (contains) the query via grep(),
 and also shows all the children categories (subcategories within that category).
 By default, it does NOT only show the entries that match the query.
 So NAICS_find('soap') shows "325612 - Polish and Other Sanitation Good Manufacturing",
 and others, not just "3256 - Soap, Cleaning Compound, and Toilet Preparation Manufacturing",
 because 3256 matches 'soap' and 325612 is a subcategory of 3256.
}
\examples{
 #NAICS_find('pulp')
 #NAICS_find('paper')
}
\seealso{
\link{NAICS_categories} \link{NAICS}
}
