% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_findwebscrape.R
\name{naics_findwebscrape}
\alias{naics_findwebscrape}
\title{for query term, show list of roughly matching NAICS, scraped from web
This finds more than just \code{\link[=naics_find]{naics_find()}} does, since that needs an exact match
but this looks at naics.com website which lists various aliases for a sector.}
\usage{
naics_findwebscrape(query)
}
\arguments{
\item{query}{text like "gasoline" or "copper smelting"}
}
\value{
data.frame of info on what was found, naics and title
}
\description{
for query term, show list of roughly matching NAICS, scraped from web
This finds more than just \code{\link[=naics_find]{naics_find()}} does, since that needs an exact match
but this looks at naics.com website which lists various aliases for a sector.
}
\examples{
 naics_find("copper smelting", search_on_naics_website=FALSE)
 naics_find("copper smelting", search_on_naics_website=TRUE)
  naics_url_of_query("copper smelting")
  \dontrun{
  naics_findwebscrape("copper smelting")
  browseURL(naics_url_of_query("copper smelting"))
  browseURL(naics_url_of_code(326))
  }
}
\seealso{
\code{\link[=naics_find]{naics_find()}} \code{\link[=naics_url_of_query]{naics_url_of_query()}}
}
