% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_from_federalregister.R
\name{naics_from_federalregister}
\alias{naics_from_federalregister}
\title{Try to extract which NAICS could be affected by a rule published in the Federal Register
by reading the NAICS listed near the top of the preamble - DRAFT WORK IN PROGRESS}
\usage{
naics_from_federalregister(naics_text_copy_from_fr)
}
\arguments{
\item{naics_text_copy_from_fr}{}
}
\description{
Try to extract which NAICS could be affected by a rule published in the Federal Register
by reading the NAICS listed near the top of the preamble - DRAFT WORK IN PROGRESS
}
\keyword{internal}
