% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_url_of_code.R
\name{naics_url_of_code}
\alias{naics_url_of_code}
\title{Get URL for page with info about industry sector(s) by NAICS
See (https://naics.com) for more information on NAICS codes}
\usage{
naics_url_of_code(naics)
}
\arguments{
\item{naics}{vector of one or more NAICS codes, like 11,"31-33",325}
}
\value{
vector of URLs as strings like https://www.naics.com/six-digit-naics/?v=2017&code=22
}
\description{
Get URL for page with info about industry sector(s) by NAICS
See (https://naics.com) for more information on NAICS codes
}
