% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_url_of_query.R
\name{naics_url_of_query}
\alias{naics_url_of_query}
\title{Get URL for page with info about industry sectors by text query term
See \link{naics.com} for more information on NAICS codes}
\usage{
naics_url_of_query(query)
}
\arguments{
\item{query}{string query term like "gasoline" or "copper smelting"}
}
\value{
URL as string
}
\description{
Get URL for page with info about industry sectors by text query term
See \link{naics.com} for more information on NAICS codes
}
