% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_validation.R
\name{NAICS_validation}
\alias{NAICS_validation}
\title{Validate NAIC uploads}
\usage{
NAICS_validation(NAICS_enter, NAIC_select)
}
\arguments{
\item{NAICS}{upload validate missing and/or improper inputs}
}
\value{
boolean value (valid or not valid)
}
\description{
Validates and prepares echo uploads
}
