% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naicstable.R
\docType{data}
\name{naicstable}
\alias{naicstable}
\title{data.table with all NAICS code numbers and industry name for each}
\description{
data.table with all NAICS code numbers and industry name for each
}
\details{
see \url{https://naics.com}
}
\seealso{
\code{\link[=naics_from_any]{naics_from_any()}} \code{\link[=naics_categories]{naics_categories()}} \link{NAICS} \code{\link[=naics_findwebscrape]{naics_findwebscrape()}}
}
