% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_climate.R
\docType{data}
\name{names_climate}
\alias{names_climate}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 4.
}
\usage{
names_climate
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
