% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_criticalservice.R
\docType{data}
\name{names_criticalservice}
\alias{names_criticalservice}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 5.
}
\usage{
names_criticalservice
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
