% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_criticalservice_pctile.R
\docType{data}
\name{names_criticalservice_pctile}
\alias{names_criticalservice_pctile}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 2.
}
\usage{
names_criticalservice_pctile
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
