% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_d_count.R
\docType{data}
\name{names_d_count}
\alias{names_d_count}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 7.
}
\usage{
names_d_count
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
