% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_d_demogindexstate.R
\docType{data}
\name{names_d_demogindexstate}
\alias{names_d_demogindexstate}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 2.
}
\usage{
names_d_demogindexstate
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
