% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_d_subgroups_nh.R
\docType{data}
\name{names_d_subgroups_nh}
\alias{names_d_subgroups_nh}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 8.
}
\usage{
names_d_subgroups_nh
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
