% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_e.R
\docType{data}
\name{names_e}
\alias{names_e}
\title{list of environmental indicator names}
\description{
list of environmental indicator names
}
\seealso{
EJAMejscreenapi dataset called map_headernames \link{names_d} \link{names_e} \link{namez}
}
