% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_e_other.R
\docType{data}
\name{names_e_other}
\alias{names_e_other}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 5.
}
\usage{
names_e_other
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
