% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_e_ratio_to_avg.R
\docType{data}
\name{names_e_ratio_to_avg}
\alias{names_e_ratio_to_avg}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 13.
}
\usage{
names_e_ratio_to_avg
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
