% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_ej_supp.R
\docType{data}
\name{names_ej_supp}
\alias{names_ej_supp}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 13.
}
\usage{
names_ej_supp
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
