% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_featuresinarea.R
\docType{data}
\name{names_featuresinarea}
\alias{names_featuresinarea}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 3.
}
\usage{
names_featuresinarea
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
