% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_geo.R
\docType{data}
\name{names_geo}
\alias{names_geo}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 25.
}
\usage{
names_geo
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
