% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesbyvarlist.R
\name{namesbyvarlist}
\alias{namesbyvarlist}
\title{Get indicator names within a varlist like names_d}
\usage{
namesbyvarlist(varlist, nametype = c("rname", "longname", "apiname")[1])
}
\arguments{
\item{varlist}{one character string like "names_d", or a vector of them}

\item{nametype}{vector of 1 or more names of columns in map_headernames, or a shortcut type that can be
api, csv, r, original, long, shortlabel}
}
\value{
a data.frame one row per indicator, one col per nametype and a column identifying the varlist
}
\description{
Get indicator names within a varlist like names_d
}
\details{
Note names_d_friendly refers to the semi-long names to use in labels
}
\examples{
 unique(map_headernames$varlist)
 
 namesbyvarlist('names_e_avg', 'rname')
 namesbyvarlist('names_d')
 namesbyvarlist('names_d', 'r')
 namesbyvarlist('names_d', 'friendly')
 namesbyvarlist('names_d', 'shortlabel')
 
 namesbyvarlist( 'names_e_pctile', c('r', 'longname'))    
 namesbyvarlist(c('names_e_pctile', 'names_e_state_pctile'), 
   c('varlist', 'r', 'api',   'names_friendly', 'longname'))

}
\keyword{internal}
