% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_namez.R
\docType{data}
\name{namez}
\alias{namez}
\title{list of lists of indicator names}
\description{
list of lists of indicator names
}
\seealso{
EJAMejscreenapi dataset called map_headernames \link{names_d} \link{names_e}
}
