% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_ratios.R
\name{plot_barplot_ratios}
\alias{plot_barplot_ratios}
\title{Barplot of ratios of demographic (or other) scores to averages (or other references)}
\usage{
plot_barplot_ratios(
  ratio.to.us.d.overall,
  names2plot_friendly = NULL,
  mycolorsavailable = c("gray", "yellow", "orange", "red"),
  main = "Demographics at the Analyzed Locations Compared to US Overall"
)
}
\arguments{
\item{ratio.to.us.d.overall}{named list of a few ratios to plot, but see \code{\link[=ejam2barplot]{ejam2barplot()}}
for an easier way to specify which indicator to show.}

\item{names2plot_friendly}{names to use for plot - should be same length as named list ratio.to.us.d.overall}

\item{mycolorsavailable}{leave as default}

\item{main}{title for plot, like "Demographics at the Analyzed Locations Compared to US Overall"}
}
\value{
ggplot should be returned
}
\description{
Barplot of ratios of demographic (or other) scores to averages (or other references)
}
\details{
\strong{SOME GENERAL NOTES, DURING EJAM DEVELOPMENT}

For plots in general, see:
\itemize{
\item \url{https://echarts4r.john-coene.com/articles/themes.html}
\item \url{https://exts.ggplot2.tidyverse.org/gallery}
}

\strong{For BARPLOTS, see/ merge/consolidate:}
\itemize{
\item output$view1_summary_plot <- renderPlot({v1_summary_plot()}) and v1_summary_plot <- reactive({ })
in EJAM server for Short Report if  bar type
\item output$summ_display_bar <- renderPlot({  }) contains its own plot code not a reactive
in EJAM server for tab showing barplots in Detailed Results
\item plot_barplot_ratios() drafted function in EJAM
}

\strong{For BOXPLOTS, see:}
\itemize{
\item v1_summary_plot <- reactive({ })     and output$view1_summary_plot <- renderPlot({v1_summary_plot()})
\itemize{
\item in EJAM server for SHORT report if box type, and
\item in EJAM server for LONG report passed as a parameter
}
\item boxplots_ratios()   in EJAMejscreenapi
(NOT in EJAM server for Detailed Results interactive views)
\item ejscreenapi_script() code also relevant? in EJAMejscreenapi
\item box/scatter examples in ggplot, \url{https://r-graph-gallery.com/89-box-and-scatter-plot-with-ggplot2.html}
\item boxplots in base R, \url{https://www.r-bloggers.com/2023/09/how-to-reorder-boxplots-in-r-a-comprehensive-guide}
}

\strong{For HISTOGRAMS, see:}
\itemize{
\item output$summ_display_hist <- renderPlot   in EJAM server for interactive views
\item the histograms code and discussion  in EJAMbatch.summarizer package
}
}
\examples{
  
  ejam2barplot(testoutput_ejamit_100pts_1miles)
  
  plot_barplot_ratios(unlist(testoutput_ejamit_1000pts_1miles$results_overall[ , c(..names_d_ratio_to_avg , ..names_d_subgroups_ratio_to_avg) ]))

}
\seealso{
\code{\link[=table_xls_format]{table_xls_format()}} \code{\link[=ejam2ratios]{ejam2ratios()}} \code{\link[=ejam2barplot]{ejam2barplot()}}
}
