% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demogshare_by_distance.R
\name{plot_demogshare_by_distance}
\alias{plot_demogshare_by_distance}
\title{plot_demogshare_by_distance - work in progress}
\usage{
plot_demogshare_by_distance(
  results_bybg_people,
  demogvarname = names_d[1],
  myids = unique(results_bybg_people$ejam_uniq_id),
  show.lowess = F,
  show.lm = TRUE,
  show.line = TRUE,
  ...
)
}
\arguments{
\item{results_bybg_people}{}

\item{demogvarname}{}

\item{show.lowess}{F}

\item{show.lm}{linefit}

\item{show.line}{linefit}

\item{...}{passed to plot}

\item{ejam_uniq_id}{}
}
\description{
plot_demogshare_by_distance - work in progress
}
\details{
Could also consider plotting something like boxplot(demogvar ~ round(distance, 1))

See notes on plots at \code{\link[=plot_barplot_ratios]{plot_barplot_ratios()}}
}
