% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{plot_distance_by_group}
\alias{plot_distance_by_group}
\title{Avg distance of each demog group (of multiple groups) - Table or Plot}
\usage{
plot_distance_by_group(
  results_bybg_people,
  demogvarname = NULL,
  demoglabel = fixcolnames(demogvarname, "r", "shortlabel"),
  graph = TRUE,
  returnwhat = "table"
)
}
\value{
see \code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}}
}
\description{
Avg distance of each demog group (of multiple groups) - Table or Plot
}
