% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{plot_distance_cdf_by_group}
\alias{plot_distance_cdf_by_group}
\title{plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances}
\usage{
plot_distance_cdf_by_group(
  results_bybg_people,
  radius_miles =
    round(max(results_bybg_people$distance_min_avgperson[!is.infinite(results_bybg_people$distance_min_avgperson)],
    na.rm = T), 1),
  demogvarname = c(namez$d_subgroups),
  demoglabel = NULL,
  colorlist = colorspace::diverging_hcl(length(demogvarname)),
  coloroverall = "black",
  ...
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{radius_miles}{miles radius that was max distance analyzed}

\item{demogvarname}{names of columns in results_bybg_people, e.g., "pctlowinc"}

\item{demoglabel}{friendly text names for labelling graphic, like "Low income residents"}

\item{colorlist}{colors like "red" etc. for the demographic groups of interest}

\item{coloroverall}{color like "gray" for everyone as a whole}
}
\value{
invisibly returns full table of sorted distances of blockgroups, cumulative count of demog groups at that block group's distance
}
\description{
plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plots(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=distance_by_groups]{distance_by_groups()}} \code{\link[=ejamit]{ejamit()}} for examples
}
