% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{plot_distance_mean_by_group}
\alias{plot_distance_mean_by_group}
\title{plot_distance_mean_by_group
Barplot of Average Proximity to sites, by Demographic Group (vs everyone else)}
\usage{
plot_distance_mean_by_group(
  results_bybg_people,
  demogvarname = c(namez$d, namez$d_subgroups),
  demoglabel = NULL,
  graph = TRUE
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{demogvarname}{vector of column names like "pctlowinc" etc.}

\item{demoglabel}{vector of labels like "Low Income Residents" etc.}
}
\value{
data.frame with group, ratio, avg_distance_for_group, avg_distance_for_nongroup
}
\description{
Note ratio shown is ratio of distance among others to distance in group,
so values below 1 mean the given demographic group lives closer to facilities.
}
\seealso{
\code{\link[=distance_by_group]{distance_by_group()}}

\code{\link[=distance_by_group_plot]{distance_by_group_plot()}}
}
