% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridgeline_ratios.R
\name{plot_ridgeline_ratios}
\alias{plot_ridgeline_ratios}
\title{plot_ridgeline_ratios
Make ridgeline plot of ratios of demographic score to its average}
\usage{
plot_ridgeline_ratios(ratio.to.us.d.bysite, names2plot_friendly = NULL)
}
\arguments{
\item{names2plot_friendly}{names to use for plot - should be same length as named list ratio.to.us.d.overall}

\item{ratio.to.us.d.overall}{named list of a few ratios to plot}
}
\description{
plot_ridgeline_ratios
Make ridgeline plot of ratios of demographic score to its average
}
