% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_from_ejscreen.R
\name{popup_from_ejscreen}
\alias{popup_from_ejscreen}
\title{Map popups - Create the popup text for maps of EJ results}
\usage{
popup_from_ejscreen(
  out,
  linkcolname = "EJScreen Report",
  linkcolname2 = "EJScreen Map",
  linkcolname3 = "EJScreenACS",
  verbose = FALSE
)
}
\arguments{
\item{out}{raw data in data.frame form, with results of EJ buffer analysis}

\item{linkcolname}{Name of one column in the table that has links to some URL}

\item{linkcolname2}{Another like linkcolname}

\item{linkcolname3}{another}

\item{verbose}{TRUE or FALSE, can see more details reported when function is used.}
}
\value{
HTML ready to be used for map popups
}
\description{
This creates the HTML text that appears in map popups.
}
\details{
Popup shows up in a window when you
click on a site on the map, when viewing the results of EJ analysis of
each site.

THIS IS CURRENTLY HARD CODED TO USE EJScreen VARIABLE NAMES.

It provides raw scores (but not for EJ indexes) and US and State percentiles if available:
\itemize{
\item some site id info fields if found
\item latitude, longitude, and size of area around site
\item Demographic Indicators like population count, Demographic Indicator, etc.
\item Environmental Indicators
\item EJ Indexes
\item web link(s) to map or report
}
}
\examples{
\dontrun{
  out <- testoutput_ejscreenapi_plus_50
  x <- popup_from_ejscreen(out)
  popup_print(x[1])
  mapfastej(out)
}

}
\keyword{internal}
