% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_EJAM_for_excel.R
\name{prep_EJAM_for_excel}
\alias{prep_EJAM_for_excel}
\title{Format batch results for excel
See EJAM code related to this also!}
\usage{
prep_EJAM_for_excel(
  df,
  hyperlink_cols = NULL,
  heatmap_colnames = NULL,
  heatmap_cuts = c(80, 90, 95),
  heatmap_colors = c("yellow", "orange", "red")
)
}
\arguments{
\item{df}{data.frame, table of batch buffer results}

\item{hyperlink_cols}{vector of names of columns in df to get treated as hyperlinks in excel}

\item{heatmap_colnames}{vector of names of columns in df to apply conditional formatting to,
by coloring like a heatmap.}

\item{heatmap_cuts}{vector of color names for heatmap bins, same length as
heatmap_cuts, where first color is for those >= 1st cutpoint, but <2d,
second color is for those >=2d cutpoint but <3d, etc.}

\item{heatmap_colors}{vector of colors corresponding to cuts}
}
\value{
A workbook via openxlsx::writeData() ready to be saved via openxlsx::saveWorkbook()
}
\description{
Format batch results for excel
See EJAM code related to this also!
}
