% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxistat2.R
\name{proxistat2}
\alias{proxistat2}
\title{Calculate a proximity score for every blockgroup
Indicator of proximity of each blockgroups to some set of facilities or sites.
Proximity score is sum of (1/d) where each d is distance of a given site in km,
summed over all sites within 5km, as in EJScreen.

*** Still need area of each block to fix this func proxistat2()}
\usage{
proxistat2(pts, cutoff = 8.04672, quadtree)
}
\arguments{
\item{pts}{data.table of lat lon}

\item{cutoff}{distance max, in miles, default is 5km (8.04672 miles)
which is the EJScreen max search range for proximity scores}

\item{quadtree}{must be localtree from EJAM::}
}
\value{
data.table with proximityscore, bgfips, lat, lon, etc.
}
\description{
Calculate a proximity score for every blockgroup
Indicator of proximity of each blockgroups to some set of facilities or sites.
Proximity score is sum of (1/d) where each d is distance of a given site in km,
summed over all sites within 5km, as in EJScreen.

*** Still need area of each block to fix this func proxistat2()
}
\examples{
 # pts <- testpoints_50
 # x <- proxistat2(pts = pts[1:1000,], quadtree = localtree) 
 #
 # summary(x$proximityscore)
 # # analyze.stuff::pctiles(x$proximityscore)
 # plot(x$lon, x$lat)
 # tops = x$proximityscore > 500 & !is.infinite(x$proximityscore) & !is.na(x$proximityscore)
 # points(x$lon[tops], x$lat[tops], col="red")
 
}
