% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_radiobox.R
\name{radiobox}
\alias{radiobox}
\title{DRAFT - shiny gadget for interactive RStudio use - lets you pick 1 option via radio buttons in popup dialog}
\usage{
radiobox(
  choiceNames = c("Points", "Shapes", "FIPS"),
  choiceValues = c("latlon", "shp", "fips"),
  label = "Choose one:",
  title = "",
  height = 250,
  width = 100
)
}
\arguments{
\item{choiceNames}{vector of options displayed}

\item{choiceValues}{vector of corresponding values as returned by the function}

\item{label}{Appears just above the list of choices}

\item{title}{Appears at top of dialog box and between cancel and done}

\item{height}{of box in pixels}

\item{width}{of box in pixels}
}
\value{
one of the choiceValues (if not cancelled/ error)
}
\description{
DRAFT - shiny gadget for interactive RStudio use - lets you pick 1 option via radio buttons in popup dialog
}
\details{
uses shiny::runGadget()

*** Note: Cannot really use within nontrivial scripts or functions as currently written
because the stopApp() seems to interrupt other processes and cause problems -
and seems related to a quirk seen if a script or function calls radiobox() twice -
it will work the first time but show a blank popup window the 2d time...
e.g., if you  try to do this:
radius1 <- radiobox()
radius2 <- radiobox()
May all be related to this: https://github.com/rstudio/rstudio/issues/13394
}
\examples{
\dontrun{
radius <- EJAM:::radiobox(c("Far (3 miles)", "Medium (2 miles)", "Near (1 mile)"), c(3,2,1),
  label = "Radius")
cat("The radius will be", radius, "miles. \n")
}

}
\keyword{internal}
