% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdattr.R
\name{rdattr}
\alias{rdattr}
\title{helper function in updating the package metadata}
\usage{
rdattr(
  packages = NULL,
  which = c("census_version", "acs_version", "acs_releasedate", "ACS",
    "ejscreen_version", "ejscreen_releasedate", "ejscreen_pkg_data", "year", "released"),
  loadifnotloaded = TRUE
)
}
\arguments{
\item{which}{Optional vector (not list) of strings, the attributes. 
Default is some typical ones used in EJAM-related packages currently.}

\item{loadifnotloaded}{Optional to control if func should load packages not already loaded.}

\item{package}{Optional. e.g. 'EJAMejscreendata', or can be a vector of character strings, 
and if not specified, default is to report on all packages with EJ as part of their name, 
like EJAMblockdata or ejscreenapi}
}
\description{
Quick and dirty helper during development, to check all the 
  attributes of all the data files in relevant packages. 
  It loads unloaded packages as needed, which you might not want it to do, 
  but it is not coded to be able to check attributes without doing that.
}
