% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_naics.R
\name{regid_from_naics}
\alias{regid_from_naics}
\title{Find EPA-regulated facilities in FRS by NAICS code (industrial category)}
\usage{
regid_from_naics(naics, id_only = FALSE)
}
\arguments{
\item{naics}{a vector of naics codes, or
a data.table with column named code, as with output of \code{\link[=naics_from_any]{naics_from_any()}}}
}
\value{
see  \code{\link[=latlon_from_naics]{latlon_from_naics()}}
}
\description{
Get lat lon, Registry ID, given NAICS industry code(s)
Find all EPA Facility Registry Service (FRS) sites with this exact NAICS code (not subcategories)
}
\details{
The EPA also provides a \href{https://www.epa.gov/frs/frs-query#industrial}{FRS Facility Industrial Classification Search tool}
where you can find facilities based on NAICS or SIC.

NOTE: many FRS sites lack NAICS code!

Also, this function does not find the sites
identified by FRS data as being in a child NAICS (subcategory of your exact query)!

Relies on  frs_by_naics (a data.table)

See info about NAICS industry codes at \url{https://www.naics.com/search}
}
\examples{
  regid_from_naics(321114)
  latlon_from_naics(321114)
  latlon_from_naics(EJAM::naics_from_any("cheese")[,code] )
  head(latlon_from_naics(c(3366, 33661, 336611), id_only=TRUE))
  # mapfast(frs_from_naics(336611)) # simple map

}
\seealso{
\code{\link[=frs_from_naics]{frs_from_naics()}}  \code{\link[=frs_from_sic]{frs_from_sic()}} \code{\link[=latlon_from_sic]{latlon_from_sic()}} \code{\link[=regid_from_naics]{regid_from_naics()}} \code{\link[=naics_from_any]{naics_from_any()}}
}
