% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionstats.R
\docType{data}
\name{regionstats}
\alias{regionstats}
\title{data.table of 100 percentiles and means for each EPA Region.}
\description{
data.table of 100 percentiles and means for each EPA Region (> 1,000 rows)
  for all the block groups in that zone (e.g., block groups in [blockgroupstats])
  for a set of indicators such as percent low income.
  Each column is one indicator (or specifies the percentile).
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
}
