% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrtable.R
\name{rrtable}
\alias{rrtable}
\title{rrtable
Calculate risk ratios of avg E in group D / in everyone not in group D}
\usage{
rrtable(x)
}
\arguments{
\item{x}{blockgroupstats}
}
\value{
table of ratios and max per row and max per col
}
\description{
rrtable
Calculate risk ratios of avg E in group D / in everyone not in group D
}
