% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames.R
\name{setdiff_yx}
\alias{setdiff_yx}
\title{setdiff_yx - UTILITY - see what is in y not x
utility just like setdiff except for y,x instead of x,y}
\usage{
setdiff_yx(x, y)
}
\description{
setdiff_yx - UTILITY - see what is in y not x
utility just like setdiff except for y,x instead of x,y
}
