% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blockpoints_in_shape.R
\name{shapefile2blockpoints}
\alias{shapefile2blockpoints}
\title{Find all Census blocks in a polygon, using internal point of block}
\usage{
shapefile2blockpoints(...)
}
\value{
same as \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}}
}
\description{
Find all Census blocks in a polygon, using internal point of block
}
