% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_clean}
\alias{shapefile_clean}
\title{shapefile_clean  -  drop invalid rows and warn if all invalid}
\usage{
shapefile_clean(shp)
}
\arguments{
\item{shp}{a shapefile object using sf::read_sf()}
}
\value{
a shapefile object using sf::read_sf()
}
\description{
shapefile_clean  -  drop invalid rows and warn if all invalid
}
\details{
it assumes crs = 4269 when doing sf::st_transform(shp, crs = 4269)
}
\seealso{
\code{\link[=shapefile_from_folder]{shapefile_from_folder()}}
}
