% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_filepaths_from_folder}
\alias{shapefile_filepaths_from_folder}
\title{shapefile_filepaths_from_folder  -  get list of valid filenames comprising shapefile including paths}
\usage{
shapefile_filepaths_from_folder(folder)
}
\arguments{
\item{folder}{path of folder that contains the files (.shp, .shx, .dbf, and .prj)}
}
\value{
string vector of filenames including full paths
}
\description{
shapefile_filepaths_from_folder  -  get list of valid filenames comprising shapefile including paths
}
\seealso{
\code{\link[=shapefile_from_folder]{shapefile_from_folder()}}
}
