% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_from_folder}
\alias{shapefile_from_folder}
\title{shapefile_from_folder  -  read shapefile from a folder}
\usage{
shapefile_from_folder(folder = NULL)
}
\arguments{
\item{folder}{path of folder that contains the files (.shp, .shx, .dbf, and .prj)}
}
\value{
a shapefile object using sf::read_sf()
}
\description{
shapefile_from_folder  -  read shapefile from a folder
}
\examples{
  testfolder <- system.file("testdata/shapes/Portland_neighborhoods", package = "EJAM")
  testshape <- shapefile_from_folder(testfolder)
  
  testpaths <- shapefile_filepaths_from_folder(testfolder)
  testshape <- shapefile_from_filepaths(testpaths)
  
  ## if interactive(), R user can point to right folder or select the right set of files:
  # testshape <- shapefile_from_filepaths()
  # testshape <- shapefile_from_folder()
  
}
