% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_from_gdbzip}
\alias{shapefile_from_gdbzip}
\title{read .zip that contains geodatabase file via unzip and st_read}
\usage{
shapefile_from_gdbzip(fname)
}
\arguments{
\item{fname}{path to .zip file that contains a .gdb file}
}
\value{
see \code{\link[=shapefile_from_gdb]{shapefile_from_gdb()}}
}
\description{
read .zip that contains geodatabase file via unzip and st_read
}
