% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{shapes_counties_from_countyfips}
\alias{shapes_counties_from_countyfips}
\title{use API to get boundaries of US Counties to map them}
\usage{
shapes_counties_from_countyfips(
  countyfips = "10001",
  outFields = "",
  myservice =
    c("https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Boundaries_2022/FeatureServer/2/query",
    "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Counties_and_States_with_PR/FeatureServer/0/query",
    "https://services.arcgis.com/cJ9YHowT8TU7DUyn/ArcGIS/rest/services/EJScreen_2_22_US_Percentiles_Tracts/FeatureServer/query")[1]
)
}
\arguments{
\item{countyfips}{FIPS codes as 5-character strings (or numbers) in a vector
as from fips_counties_from_state_abbrev("DE")}

\item{outFields}{can be "*" for all, or can be
just some variables like SQMI, POPULATION_2020, etc., or none}

\item{myservice}{URL of feature service to get shapes from.
Only default was tested}
}
\value{
spatial object via sf::read_sf()
}
\description{
use API to get boundaries of US Counties to map them
}
