% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIC.R
\docType{data}
\name{SIC}
\alias{SIC}
\title{named list of all SIC code numbers and category name for each}
\description{
named list of all SIC code numbers and category name for each
}
\details{
see \url{https://siccode.com}
}
