% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_from_any.R
\name{siteid_from_sic}
\alias{siteid_from_sic}
\title{latlon_from_sic - Find EPA-regulated facilities in FRS by SIC code (industrial category)
Get lat lon, Registry ID, given SIC industry code(s)
Find all EPA Facility Registry Service (FRS) sites with this exact SIC code (not subcategories)}
\usage{
siteid_from_sic(sic, id_only = FALSE)
}
\arguments{
\item{sic}{a vector of SIC codes, or
a data.table with column named code, as with output of \code{\link[=sic_from_any]{sic_from_any()}}}
}
\value{
A data.table (not just data.frame) with columns called
lat, lon, REGISTRY_ID, SIC (but see the id_only parameter)
}
\description{
latlon_from_sic - Find EPA-regulated facilities in FRS by SIC code (industrial category)
Get lat lon, Registry ID, given SIC industry code(s)
Find all EPA Facility Registry Service (FRS) sites with this exact SIC code (not subcategories)
}
\details{
NOTE: many FRS sites lack SIC code!

Also, this function does not find the sites
identified by FRS data as being in a child SIC (subcategory of your exact query)!

Relies on  frs_by_sic (a data.table)

See info about SIC industry codes at \url{https://www.naics.com/search}
}
\examples{
  siteid_from_sic('7300')
  latlon_from_sic('7300')
  latlon_from_sic(sic_from_any("cheese")[,code] )
  head(latlon_from_sic(c('6150', '6300', '5995'), id_only=TRUE))
  # mapfast(frs_from_sic('6150')) # simple map
}
