% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepoints_example.R
\docType{data}
\name{sitepoints_example}
\alias{sitepoints_example}
\title{data.table of points as example of sitepoints for EJAM}
\description{
data.table of points as example of sitepoints for EJAM
}
