% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sites2blocks_example100pts_1miles.R
\docType{data}
\name{sites2blocks_example100pts_1miles}
\alias{sites2blocks_example100pts_1miles}
\title{test output of getblocksnearby(), and is an input to doaggregate()}
\description{
test output of getblocksnearby(), and is an input to doaggregate()
}
\details{
This is the output of getblocksnearby(testpoints_100, radius = 1)
This is the same as  \link{testoutput_getblocksnearby_100pts_1miles}
}
\seealso{
\code{\link[=getblocksnearby]{getblocksnearby()}}  \code{\link[=doaggregate]{doaggregate()}}  \link{testpoints_100}
}
