% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sites2blocks_example_1000pts_1miles.R
\docType{data}
\name{sites2blocks_example_1000pts_1miles}
\alias{sites2blocks_example_1000pts_1miles}
\title{test output of getblocksnearby() or input to doaggregate()}
\description{
test output of getblocksnearby() or input to doaggregate()
}
