% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedtest.R
\name{speedtest}
\alias{speedtest}
\title{speedtest
Runs EJAM analysis for several radius values for many sitepoints, recording how long each step took.}
\usage{
speedtest(
  n = 10,
  sitepoints = NULL,
  weighting = "frs",
  radius = c(1, 3.106856, 5, 10, 20)[1],
  do_batch.summarize = TRUE,
  logging = TRUE,
  logfolder = getwd(),
  logfilename = "log_n_datetime.txt",
  honk_when_ready = TRUE
)
}
\arguments{
\item{n}{optional, how many random points to test, or set to 0 to interactively pick file of points in RStudio (n is ignored if sitepoints provided)}

\item{sitepoints}{optional,  (use if you do not want random points) data.frame of points or path/file with points, where columns are lat and lon in decimal degrees}

\item{weighting}{optional, if using random points, how to weight them, such as facilities, people, or blockgroups. see \code{\link[=testpoints_n]{testpoints_n()}}}

\item{radius}{optional, one or more radius values in miles to use in creating circular buffers when findings residents nearby each of sitepoints}

\item{do_batch.summarize}{optional, if want to also do that function}

\item{logging}{logical optional, whether to save log file with timings of steps}

\item{logfolder}{optional, name of folder for log file}

\item{logfilename}{optional, name of log file to go in folder}

\item{honk_when_ready}{optional, self-explanatory}
}
\value{
EJAM results similar to as from the web app  \code{\link[=ejamit]{ejamit()}}
}
\description{
speedtest
Runs EJAM analysis for several radius values for many sitepoints, recording how long each step took.
}
