% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_blockid}
\alias{state_from_blockid}
\title{given vector of blockids, get state abbreviation of each}
\usage{
state_from_blockid(blockid)
}
\arguments{
\item{blockid}{vector of blockid values as from EJAM in a table called blockpoints}
}
\value{
vector of ST info like AK, CA, DE, etc.
}
\description{
given vector of blockids, get state abbreviation of each
}
\examples{
EJAM:::state_from_blockid(c(8174952, blockpoints$blockid[5:6]))

}
\keyword{internal}
