% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_blocktable}
\alias{state_from_blocktable}
\title{state_from_blocktable - was used only in some special cases of using testpoints_n()
given data.table with blockid column, get state abbreviation of each - not used?}
\usage{
state_from_blocktable(dt_with_blockid)
}
\arguments{
\item{dt_with_blockid}{}
}
\value{
vector of ST info like AK, CA, DE, etc.
}
\description{
state_from_blocktable - was used only in some special cases of using testpoints_n()
given data.table with blockid column, get state abbreviation of each - not used?
}
\examples{
state_from_blocktable(blockpoints[45:49,])
}
