% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_fips}
\alias{state_from_fips}
\title{state_from_fips}
\usage{
state_from_fips(fips)
}
\arguments{
\item{fips}{Census FIPS codes vector, numeric or char, 2-digit, 5-digit, etc. OK}
}
\value{
vector of 2-character state abbreviations like CA,MD,TX,OH
}
\description{
state_from_fips
}
