% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_from_latlon.R
\name{state_from_latlon}
\alias{state_from_latlon}
\title{find what state is where each point is located}
\usage{
state_from_latlon(lat, lon, states_shapefile = EJAM::states_shapefile)
}
\arguments{
\item{lat}{latitudes vector}

\item{lon}{longitudes vector}

\item{shapefile}{shapefile of US States, in package already}
}
\value{
Returns data.frame: ST, statename, FIPS.ST, REGION,
each as long as lat or lon
}
\description{
find what state is where each point is located
}
