% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateregions.R
\docType{data}
\name{stateregions}
\alias{stateregions}
\title{data.table that shows which states are in which EPA regions}
\description{
51 rows (50 states plus DC, no PR, no Island Areas),
3 columns: ST with 2-character state abbreviation,
REGION with EPA Region number 1-10, and STATENAME.
}
