% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_infer.R
\name{states_infer}
\alias{states_infer}
\title{states_infer - Identify US State that each site is in (given ST, FIPS, or lat/lon)}
\usage{
states_infer(x)
}
\arguments{
\item{x}{data.frame or data.table, with one row per site, and
column(s) that are either "ST" (2-letter abbreviation of State),
"lat" and "lon", or "fips" or "bgfips"
and optionally a column like  "ejam_uniq_id" or "n"}
}
\value{
the input table as a data.frame, but with these
new columns if ST was not already a column:
ejam_uniq_id, ST, statename, FIPS.ST, REGION, n
}
\description{
Identify US State that each site is in (given ST, lat/lon, or FIPS)
}
\examples{
  states_infer(testpoints_10)
  states_infer(testoutput_ejamit_10pts_1miles$results_bysite[, .(ejam_uniq_id, ST, pop)])
  states_infer(testoutput_ejamit_10pts_1miles$results_bysite[, .(ST, pop)])
  states_infer(testoutput_ejamit_10pts_1miles$results_bysite[, .(ST, lat, lon, pop)])
  
}
\seealso{
\code{\link[=state_from_latlon]{state_from_latlon()}} \code{\link[=state_from_fips]{state_from_fips()}}
}
