% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_states_shapefile.R
\docType{data}
\name{states_shapefile}
\alias{states_shapefile}
\title{states_shapefile (DATA) US States boundaries 2020 shapefile from TIGER}
\description{
This is used to figure out which state contains each point (facility/site).
}
\details{
This is used by \code{\link[=state_from_latlon]{state_from_latlon()}} to find which state is associated with each point
that the user wants to analyze. That is needed to report indicators in
the form of State-specific percentiles
(e.g., a score that is at the 80th percentile within Texas).
It is created by the package via a script at EJAM/data-raw/datacreate_states_shapefile.R
which downloads the data from Census Bureau.
}
\seealso{
\code{\link[=state_from_latlon]{state_from_latlon()}} \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}}
}
