% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_structure.of.output.list.R
\name{structure.of.output.list}
\alias{structure.of.output.list}
\title{Utility to print summary info about the output of ejamit() or doaggregate()}
\usage{
structure.of.output.list(x, maxshown = 10)
}
\arguments{
\item{x}{the output of \code{\link[=ejamit]{ejamit()}} or of \code{\link[=doaggregate]{doaggregate()}},
a list of objects holding results of analysis}

\item{maxshown}{shows only first 10 elements of list by default}
}
\value{
data.frame summarizing names of list, whether each element is a
data.table, data.frame, or vector, and rows/cols/length info
}
\description{
Utility to print summary info about the output of ejamit() or doaggregate()
}
\examples{
 
  structure.of.output.list(testpoints_10)
  structure.of.output.list(testoutput_getblocksnearby_10pts_1miles)
  structure.of.output.list(testoutput_doaggregate_10pts_1miles)
  structure.of.output.list(testoutput_ejamit_10pts_1miles)
  structure.of.output.list(testoutput_ejscreenapi_plus_5)
  structure.of.output.list(testoutput_ejscreenit_5)
  
}
\keyword{internal}
