% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_blockcount.R
\name{summarize_blockcount}
\alias{summarize_blockcount}
\title{Get summary stats on counts of blocks (unique vs doublecounted) near sites}
\usage{
summarize_blockcount(x)
}
\arguments{
\item{x}{The output of getblocksnearby()}
}
\value{
A list of stats
}
\description{
Get summary stats on counts of blocks (unique vs doublecounted) near sites
}
