% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_blockcount.R
\name{summarize_blocks_per_site}
\alias{summarize_blocks_per_site}
\title{Get summary stats on counts of blocks near various sites}
\usage{
summarize_blocks_per_site(x, varname = "siteid")
}
\arguments{
\item{x}{The output of getblocksnearby()}

\item{varname}{colname of variable in data.table x that is the one to summarize by}
}
\value{
invisibly, a list of stats
}
\description{
Tells you # of blocks near avg site, 
how many sites have only 1 block nearby, or have <30 nearby, etc.
}
