% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_blockcount.R
\name{summarize_sites_per_block}
\alias{summarize_sites_per_block}
\title{Get summary stats on how many sites are near various blocks (residents)}
\usage{
summarize_sites_per_block(x, varname = "blockid")
}
\arguments{
\item{x}{The output of getblocksnearby()}
}
\value{
invisibly, a list of stats
}
\description{
Get summary stats on how many sites are near various blocks (residents)
}
