% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeForFacilities.R
\name{summarizeForFacilities}
\alias{summarizeForFacilities}
\title{Wrapper for other functions doing the work - seems to duplicate getblocksnearby_and_doaggregate.R}
\usage{
summarizeForFacilities(sitepoints, cutoff, maxcutoff, avoidorphans, ...)
}
\arguments{
\item{sitepoints}{see \link{getrelevantCensusBlocksviaQuadTree_Clustered} or other such functions}

\item{cutoff}{see \link{getrelevantCensusBlocksviaQuadTree_Clustered} or other such functions}

\item{maxcutoff}{see \link{getrelevantCensusBlocksviaQuadTree_Clustered} or other such functions}

\item{avoidorphans}{see \link{getrelevantCensusBlocksviaQuadTree_Clustered} or other such functions}

\item{...}{see \link{getrelevantCensusBlocksviaQuadTree_Clustered} or other such functions}
}
\description{
Wrapper for other functions doing the work - seems to duplicate getblocksnearby_and_doaggregate.R
}
