% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_blockcount.R
\name{summary_of_blockcount}
\alias{summary_of_blockcount}
\title{Get summary stats on counts of unique and non-unique blocks nearby}
\usage{
summary_of_blockcount(x)
}
\arguments{
\item{x}{The output of getblocksnearby()}
}
\value{
A list of stats
}
\description{
Get summary stats on counts of unique and non-unique blocks nearby
}
