% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ratios_from_ejamit-DRAFT.R
\name{table_ratios_from_ejamit}
\alias{table_ratios_from_ejamit}
\title{DRAFT - Quick view of summary stats by type of stat, but lacks rounding specific to each type, etc.}
\usage{
table_ratios_from_ejamit(
  ejamitout,
  vartypes = c("ratio", "pctile", "pct", "all")[1],
  stats = c("Average site", "Average person", "Median site", "Median person", "Min",
    "Max"),
  decimals = 1
)
}
\arguments{
\item{ejamitout}{na}

\item{vartypes}{na}

\item{stats}{na}

\item{decimals}{na}
}
\value{
na
}
\description{
DRAFT - Quick view of summary stats by type of stat, but lacks rounding specific to each type, etc.
}
