% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt_.R
\name{table_validated_ejamit_row}
\alias{table_validated_ejamit_row}
\title{table_validated_ejamit_row - Cleans/validates EJAM results for 1 place or overall
This is a first step in formatting results in nice tables}
\usage{
table_validated_ejamit_row(ejamit_results_1row = NULL)
}
\arguments{
\item{ejamit_results_1row}{1-row data.table like testoutput_ejamit_100pts_1miles$results_overall,

as would come from ejamit(testpoints_10)$results_overall

or testoutput_ejamit_100pts_1miles$results_bysite\link{ 1, }}
}
\value{
Returns the input as a 1-row data.table, indicators etc. in the columns.
If not a 1 row table, or colnames are not what is expected, it returns correct structure filled with NA values.
}
\description{
table_validated_ejamit_row - Cleans/validates EJAM results for 1 place or overall
This is a first step in formatting results in nice tables
}
\examples{
  x <- table_validated_ejamit_row(testoutput_ejamit_100pts_1miles$results_bysite[ 1, ])
  x <- table_validated_ejamit_row(testoutput_ejamit_100pts_1miles$results_overall)
  
}
