% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testdata_doaggregate_output_100.R
\docType{data}
\name{testdata_doaggregate_output_100}
\alias{testdata_doaggregate_output_100}
\title{test output of doaggregate(sites2blocks_example_100pts_1miles)}
\description{
test output of doaggregate(sites2blocks_example_100pts_1miles)
}
