% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testdata_ejamit_output_100pts_1mile.R
\docType{data}
\name{testdata_ejamit_output_100pts_1mile}
\alias{testdata_ejamit_output_100pts_1mile}
\title{test output of ejamit(testpoints_100, radius = 1)}
\description{
test output of ejamit(testpoints_100, radius = 1)
}
