% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testdata.R
\docType{data}
\name{testdata_sites2blocks}
\alias{testdata_sites2blocks}
\title{test data for 100 points to use as input to doaggregate()}
\description{
test data for 100 points to use as input to doaggregate()
}
\details{
data.table of output of \code{\link[=getblocksnearby]{getblocksnearby()}}, using testpoints_100_dt
each row is a unique site-block-distance.
Can be used as input to \code{\link[=doaggregate]{doaggregate()}}
}
\seealso{
\link{testdata_sites2blocks} is the same as \link{sites2blocks_example}
}
