% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testdata.R
\docType{data}
\name{testdata_sites2blocks_1000}
\alias{testdata_sites2blocks_1000}
\title{test data for 1000 points to use as input to doaggregate()}
\description{
test data for 1000 points to use as input to doaggregate()
}
\details{
data.table of output of \code{\link[=getblocksnearby]{getblocksnearby()}}, using testpoints_1000_dt
each row is a unique site-block-distance.
Can be used as input to \code{\link[=doaggregate]{doaggregate()}}
}
\seealso{
\link{testdata_sites2blocks} \link{testdata_sites2blocks_1000}  \link{testpoints_1000_dt}
}
