% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejscreenapi_plus_500.R
\docType{data}
\name{testoutput_ejscreenapi_plus_500}
\alias{testoutput_ejscreenapi_plus_500}
\title{test data examples of output from \code{\link[=ejscreenapi_plus]{ejscreenapi_plus()}} using \link{testpoints_500}, radius = 1}
\description{
test data examples of output from \code{\link[=ejscreenapi_plus]{ejscreenapi_plus()}} using \link{testpoints_500}, radius = 1
}
\details{
Just for convenience, installed with the package.
Has header row plus 500 rows, and about 200+ columns of buffer summary results.
}
\seealso{
\link{testpoints_50}                   \link{testpoints_500}

\link{testoutput_ejscreenit_50}       \link{testoutput_ejscreenit_500}

\link{testoutput_ejscreenapi_plus_50} \link{testoutput_ejscreenapi_plus_500}
}
