% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejscreenit_5.R
\docType{data}
\name{testoutput_ejscreenit_5}
\alias{testoutput_ejscreenit_5}
\title{test data examples of output from \link{ejscreenit} using (testpoints_5, radius = 1)}
\description{
test data examples of output from \link{ejscreenit} using (testpoints_5, radius = 1)
}
\details{
Just for convenience, installed with the package.
Has header row plus 5 rows, and about 200+ columns of buffer summary results.
}
\seealso{
\link{testpoints_50}                   \link{testpoints_500}

\link{testoutput_ejscreenit_50}       \link{testoutput_ejscreenit_500}

\link{testoutput_ejscreenapi_plus_50} \link{testoutput_ejscreenapi_plus_500}
}
