% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_10.R,
%   R/data_testpoints_conus5.R
\docType{data}
\name{testpoints_10}
\alias{testpoints_10}
\title{test points data.frame with columns sitenumber, lat, lon}
\description{
test points data.frame with columns sitenumber, lat, lon

test points data.frame with columns siteid, lat, lon
}
