% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_1000.R
\docType{data}
\name{testpoints_1000}
\alias{testpoints_1000}
\title{test points data.frame with columns sitenumber, lat, lon}
\description{
test points data.frame with columns sitenumber, lat, lon
}
