% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoints_blockpoints.R
\name{testpoints_blockpoints}
\alias{testpoints_blockpoints}
\title{Get some random US locations as points to try out/ for testing}
\usage{
testpoints_blockpoints(n = 10, weighting = "geo", ST = is.null, as.dt = TRUE)
}
\arguments{
\item{n}{how many points do you want?}

\item{weighting}{geo means each block is equally likely, pop means the points
are population weighted (Census 2020 pop) so they represent a random sample of
where US residents live - the average person.}

\item{ST}{can be a character vector of 2 letter State abbreviations to pick from only some States}

\item{as.dt}{if TRUE (default), a data.table, but if FALSE then a data.frame}
}
\value{
see as.dt paramter. It returns a table with columns blockid, lat, lon
}
\description{
Get some random US locations as points to try out/ for testing
}
