% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_conus5.R
\docType{data}
\name{testpoints_conus5}
\alias{testpoints_conus5}
\title{test points data.frame with columns lat, lon, sitenumber, sitename}
\description{
test points data.frame with columns lat, lon, sitenumber, sitename
}
