% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilaterate.R
\name{trilaterate_sites2blocks}
\alias{trilaterate_sites2blocks}
\title{trilaterate_sites2blocks - Estimate lat,lon of each siteid, from outputs of getblocksnearby()
get data.table with siteid, lat,lon of each site (eg for when you did not save sitepoints info)}
\usage{
trilaterate_sites2blocks(s2b)
}
\arguments{
\item{s2b}{like \link{sites2blocks_example_2pts_1miles}}
}
\value{
a data.table with one row per unique siteid from input dt,
plus lat,lon columns
}
\description{
trilaterate_sites2blocks - Estimate lat,lon of each siteid, from outputs of getblocksnearby()
get data.table with siteid, lat,lon of each site (eg for when you did not save sitepoints info)
}
\examples{
s2b = copy(sites2blocks_example_2pts_1miles)
s2b_located = latlon_join_on_blockid(s2b) # done by trilaterate also
inferred_sites = trilaterate_sites2blocks(s2b)
inferred_sites
plotblocksnearby(s2b_located)

}
