% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames.R
\name{unshared}
\alias{unshared}
\title{unshared aka setdiff2 - UTILITY - see what is only in x or y but not both
utility just like setdiff except for y,x and also x,y
Just shows which elements are in one and only one of the sets x and y}
\usage{
unshared(x, y)
}
\description{
unshared aka setdiff2 - UTILITY - see what is only in x or y but not both
utility just like setdiff except for y,x and also x,y
Just shows which elements are in one and only one of the sets x and y
}
