% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_bookmark-DRAFT.R
\name{url_bookmark_save}
\alias{url_bookmark_save}
\title{url_bookmark_save
save bookmarked EJScreen session (map location and indicator)}
\usage{
url_bookmark_save(..., file = "ejscreenbookmark.json")
}
\arguments{
\item{...}{passed to \code{\link[=url_bookmark_text]{url_bookmark_text()}}}

\item{file}{path and name of .json file you want to save locally}
}
\value{
URL for 1 bookmarked EJScreen map location and variable displayed on map
}
\description{
url_bookmark_save
save bookmarked EJScreen session (map location and indicator)
}
\details{
WORK IN PROGRESS - NOT USED AS OF EARLY 2023.
You can use this function to create and save a json file that is a bookmark
for a specific place/ map view/ data layer in EJScreen.
You can later pull up that exact map in EJScreen by launching EJScreen,
clicking Tools, Save Session, Load from File.

***Units are not lat lon: "spatialReference":{"latestWkid":3857,"wkid":102100}

Note:
(1) The number of sessions that can be saved depends on the browser cache size.
(2) Session files, if saved, are available from the default Downloads folder on your computer.
(3) Users should exercise caution when saving sessions that may contain sensitive or confidential data.
}
