% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_bookmark.R
\name{url_bookmark_text}
\alias{url_bookmark_text}
\title{url_bookmark_text
URL for 1 bookmarked EJScreen session (map location and indicator)}
\usage{
url_bookmark_text(
  x = c(-13232599.1784247, -13085305.0249191),
  y = c(3970069.24597194, 4067373.582979),
  name = "BookmarkedEJScreenMap",
  title = "Socioeconomic Indicators",
  renderField = "B_UNEMPPCT",
  pctlevel = "nation",
  xmin = 1.1 * min(x),
  xmax = 0.9 * min(x),
  ymin = 0.9 * min(y),
  ymax = 1.1 * min(y),
  urlrest = paste0("https://geopub.epa.gov/arcgis/rest/services",
    "/ejscreen/ejscreen_v2022_with_AS_CNMI_GU_VI/MapServer")
)
}
\arguments{
\item{x}{vector of approx topleft, bottomright longitudes in some units EJScreen uses?
Units are not lat lon: "spatialReference":{"latestWkid":3857,"wkid":102100}}

\item{y}{vector of approx topleft, bottomright latitudes in some units EJScreen uses?
Units are not lat lon: "spatialReference":{"latestWkid":3857,"wkid":102100}}

\item{name}{Your name for the map bookmark}

\item{title}{Your name for the map like Socioeconomic Indicators  or  Pollution and Sources}

\item{renderField}{name of variable shown on map, like B_UNEMPPCT for map color bins of percent unemployed
or B_PTRAF for traffic indicator}

\item{pctlevel}{nation or state}

\item{xmin}{calculated bounding box for map view}

\item{xmax}{calculated bounding box for map view}

\item{ymin}{calculated bounding box for map view}

\item{ymax}{calculated bounding box for map view}

\item{urlrest}{Just use the default but it changes each year}
}
\value{
URL for 1 bookmarked EJScreen map location and variable displayed on map
}
\description{
url_bookmark_text
URL for 1 bookmarked EJScreen session (map location and indicator)
}
\details{
WORK IN PROGRESS - NOT USED AS OF EARLY 2023.
You can use this function to create and save a json file that is a bookmark
for a specific place/ map view/ data layer in EJScreen.
You can later pull up that exact map in EJScreen by launching EJScreen,
clicking Tools, Save Session, Load from File.

Note:
(1) The number of sessions that can be saved depends on the browser cache size.
(2) Session files, if saved, are available from the default Downloads folder on your computer.
(3) Users should exercise caution when saving sessions that may contain sensitive or confidential data.
}
\examples{
\dontrun{
  url_bookmark_text()
  url_bookmark_save(
    x=c(-10173158.179197036, -10128824.702791695), 
    y=c(3548990.034736070,3579297.316451102), 
    file="./mysavedejscreensession1.json")
  }
}
\seealso{
\code{\link[=url_bookmark_save]{url_bookmark_save()}}
}
