% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_ejscreenmap}
\alias{url_ejscreenmap}
\title{Get URL(s) for EJScreen map centered at given point(s)}
\usage{
url_ejscreenmap(lon, lat, as_html = FALSE, linktext, wherestr = "")
}
\arguments{
\item{lon}{one or more longitudes}

\item{lat}{one or more latitudes}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{wherestr}{passed to API in URL as wherestr= , if lat/lon not used. Can be State abbrev like "NY"
or can be some other options - not sure that part of the EJScreen URL-encoded map request is documented.}
}
\value{
URL(s)
}
\description{
Get URL(s) for EJScreen map centered at given point(s)
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreen_acs_report]{url_ejscreen_acs_report()}}   \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=url_frs_report]{url_frs_report()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}  \code{\link[=url_envirofacts_data]{url_envirofacts_data()}}
}
