% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_getacs_epaquery-DRAFT.R
\name{url_get_eparest_chunked_by_id}
\alias{url_get_eparest_chunked_by_id}
\title{experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API}
\usage{
url_get_eparest_chunked_by_id(objectIds, chunksize = 200, ...)
}
\arguments{
\item{objectIds}{see API}

\item{chunksize}{see API}

\item{...}{passed to url_getacs_epaquery()}
}
\value{
a table
}
\description{
experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API
}
