% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_get_via_url-DRAFT.R
\name{url_get_via_url}
\alias{url_get_via_url}
\title{url_get_via_url - helper function work in progress: GET json via url of ejscreen ejquery map services}
\usage{
url_get_via_url(url)
}
\arguments{
\item{url}{the url for an EJScreen ejquery request}
}
\value{
json
}
\description{
url_get_via_url - helper function work in progress: GET json via url of ejscreen ejquery map services
}
