% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_naics.com.R
\name{url_naics.com}
\alias{url_naics.com}
\title{url_naics.com - Get URL for page with info about industry sectors by text query term
See (https://naics.com) for more information on NAICS codes}
\usage{
url_naics.com(query, as_html = FALSE, linktext)
}
\arguments{
\item{query}{string query term like "gasoline" or "copper smelting"}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}
}
\value{
URL as string
}
\description{
url_naics.com - Get URL for page with info about industry sectors by text query term
See (https://naics.com) for more information on NAICS codes
}
