% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/url_to_any_rest_services_ejscreen_ejquery-DRAFT.R
\name{url_to_any_rest_services_ejscreen_ejquery}
\alias{url_to_any_rest_services_ejscreen_ejquery}
\title{utility - draft - generic function wrapping ejscreen/ejquery API calls}
\usage{
url_to_any_rest_services_ejscreen_ejquery(servicenumber = 7)
}
\arguments{
\item{servicenumber}{which service - depends on API}
}
\description{
utility - draft - generic function wrapping ejscreen/ejquery API calls
}
\details{
Disadvantage of this generic approach is
it does not help you by showing a list of parameters,
since those are specific to the service like 7 vs 71.

see links to documentation on using APIs

dev/intersect-distance/arcgis/ArcGIS REST API basics.txt
}
\section{notes - examples}{
url_getacs_epaquery(  objectIds = 1:4,                 outFields = 'STCNTRBG', justurl = TRUE)

t(url_getacs_epaquery(objectIds = sample(1:220000,2),  outFields = '*'))

t(url_getacs_epaquery(objectIds = sample(1:220000,2)))

url_getacs_epaquery(  objectIds = sample(1:220000,10), outFields = c('STCNTRBG', 'STATE', 'COUNTY', 'TRACT', 'BLKGRP'), justurl = FALSE)

y <- url_get_via_url(url_to_any_rest_services_ejscreen_ejquery(         servicenumber = 71, lat = 30.494982, lon = -91.132107, miles = 1))

x <- url_get_via_url(url_to_get_nearby_blocks_rest_services_ejscreen_ejquery_MapServer_71(lat = 30.494982, lon = -91.132107, miles = 1))

z <- url_get_via_url(url_to_get_ACS2019_rest_services_ejscreen_ejquery_MapServer_7())
}

\keyword{internal}
