% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{urls_clusters_and_sort_cols}
\alias{urls_clusters_and_sort_cols}
\title{Add Links to ejscreenapi output, to launch EJScreen report for given point(s)}
\usage{
urls_clusters_and_sort_cols(results_table)
}
\arguments{
\item{results_table}{from ejscreenapi() function for example}
}
\value{
the input table but with extra columns
}
\description{
Add or update, and reorder, columns with results
}
\details{
Creates weblinks to maps, as EJScreenMAP column,

Adds weblinks to pdf-like reports, as new column

Adds a column to flag sites that are close to other sites, and

Puts certain columns first.
}
\seealso{
\code{\link[=url_ejscreenmap]{url_ejscreenmap()}} \code{\link[=near_eachother]{near_eachother()}}
}
\keyword{internal}
