% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usastats.R
\docType{data}
\name{usastats}
\alias{usastats}
\title{data.table of 100 percentiles and means, in the USA overall, across all locations (e.g., block groups)
  for a set of indicators such as percent low income
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
  and same as the lookup tables in ejscreen::lookupUSA  or lookupStates}
\description{
data.table of 100 percentiles and means, in the USA overall, across all locations (e.g., block groups)
  for a set of indicators such as percent low income
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
  and same as the lookup tables in ejscreen::lookupUSA  or lookupStates
}
