% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{usastats_queryd}
\alias{usastats_queryd}
\title{see statestats_query()}
\usage{
usastats_queryd(varnames = EJAM::names_d, PCTILES = NULL, dig = 2)
}
\arguments{
\item{varnames}{see statestats_query()}

\item{PCTILES}{see statestats_query()}

\item{dig}{see statestats_query()}
}
\description{
see statestats_query()
}
\examples{
see statestats_query()
}
