% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_is_numeric_ish.R
\name{var_is_numeric_ish}
\alias{var_is_numeric_ish}
\title{var_is_numeric_ish - see which columns seem numeric and could be rounded, e.g. - DRAFT NOT FULLY TESTED}
\usage{
var_is_numeric_ish(
  x,
  only.if.already.numeric = FALSE,
  strip.characters.before.coerce = FALSE
)
}
\arguments{
\item{x}{data.table, data.frame, or vector}

\item{only.if.already.numeric}{logical, if TRUE, only reports TRUE for a column (or element) if  is.numeric() is TRUE for that one}

\item{strip.characters.before.coerce}{logical, if TRUE, tries to remove spaces and percentage signs before trying to coerce to numeric}
}
\value{
logical vector as long as NCOL(x) i.e., is length(x), if x is table, or length(x) if vector
}
\description{
var_is_numeric_ish - see which columns seem numeric and could be rounded, e.g. - DRAFT NOT FULLY TESTED
}
\seealso{
\code{\link[=table_round]{table_round()}}
}
