% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{varname2color_ejam}
\alias{varname2color_ejam}
\title{helper function - for color coding excel sheet columns
Wrapper for varname2vartype_ejam() and vartype2color_ejam() to simplify use case}
\usage{
varname2color_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{things like us.avg.pctlowinc}
}
\value{
vector of colors
}
\description{
helper function - for color coding excel sheet columns
Wrapper for varname2vartype_ejam() and vartype2color_ejam() to simplify use case
}
\seealso{
varname2vartype_ejam() vartype2color_ejam()
}
