% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{varname2vartype_ejam}
\alias{varname2vartype_ejam}
\title{varname2vartype_ejam
helper function - given indicator names, look up what type each is}
\usage{
varname2vartype_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{vector of 1 or more names}

\item{varnameinfo}{data.frame with info on type of each variable}
}
\value{
vector same size as varname
}
\description{
varname2vartype_ejam
helper function - given indicator names, look up what type each is
}
\details{
The types are things like raw data count for indicator, average, percentile, etc.
Variable names are stored in column of varnameinfo called newnames_ejscreenapi
Types are stored in column of varnameinfo called jsondoc_vartype
}
\seealso{
vartype2color_ejam() varname2color_ejam()
}
