% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{vartype2color_ejam}
\alias{vartype2color_ejam}
\title{vartype2color_ejam
helper function - assign fill color to shade excel cells by indicator type
Use color shading to make spreadsheet easier to use, grouping the indicators}
\usage{
vartype2color_ejam(vartype)
}
\arguments{
\item{vartype}{must be one found in varnameinfo$jsondoc_vartype,
ie "percentile", "average", or "raw data for indicator"
NA if not found.}
}
\value{
vector of colors like c('lightblue', 'gray') matching length of vartype
}
\description{
vartype2color_ejam
helper function - assign fill color to shade excel cells by indicator type
Use color shading to make spreadsheet easier to use, grouping the indicators
}
\seealso{
varname2vartype_ejam() varname2color_ejam()
}
