% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_output_styled.R
\name{workbook_output_styled}
\alias{workbook_output_styled}
\title{workbook_output_styled
Format EJAM tabular outputs for saving as Excel spreadsheet}
\usage{
workbook_output_styled(
  overall,
  eachsite,
  graycolnums = NULL,
  narrowcolnums = NULL,
  graycolor = "gray",
  narrow6 = 6,
  ...
)
}
\arguments{
\item{overall}{table to save in one tab, from EJAM analysis of indicators overall (one row)}

\item{eachsite}{table to save in another tab, from EJAM analysis site by site (one row per site)}

\item{graycolnums}{which columns to deemphasize}

\item{narrowcolnums}{which columns to make narrow}

\item{graycolor}{color used to deemphasize some columns}

\item{narrow6}{how narrow}

\item{...}{other params passed along to \code{\link[openxlsx:writeData]{openxlsx::writeData()}}}
}
\value{
a workbook, ready to be saved in spreadsheet format, with tabs like "Overall" and "Each Site"
}
\description{
workbook_output_styled
Format EJAM tabular outputs for saving as Excel spreadsheet
}
\details{
NEED TO MERGE THIS WITH EJAMejscreenapi's  prep_for_excel()
}
\examples{
\dontrun{
  wb <- workbook_output_styled(datasetResults()$results_overall, datasetResults()$results_bysite)
  openxlsx::saveWorkbook(wb, "results.xlsx", overwrite = TRUE)
}
}
