% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pctiles_lookup.R
\name{write_pctiles_lookup}
\alias{write_pctiles_lookup}
\title{create lookup table of percentiles 0 to 100 and mean for each indicator by State or USA total}
\usage{
write_pctiles_lookup(
  x,
  zone.vector = NULL,
  zoneOverallName = "USA",
  wts = NULL
)
}
\arguments{
\item{x}{data.frame with numeric data. Each column will be examined to calculate
mean, sd, and percentiles, for each zone}

\item{zone.vector}{optional names of states or regions, for example. same length as wts, or rows in mydf}

\item{zoneOverallName}{optional. Default is USA.}

\item{wts}{not used in EJScreen percentiles anymore}
}
\description{
create lookup table of percentiles 0 to 100 and mean for each indicator by State or USA total
}
