% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{xls_formatting2}
\alias{xls_formatting2}
\title{xls_formatting2
Format EJAM tabular outputs for saving as Excel spreadsheet}
\usage{
xls_formatting2(
  overall,
  eachsite,
  longnames = NULL,
  bybg = NULL,
  plotfilename = NULL,
  summary_plot = NULL,
  plotlatest = TRUE,
  analysis_title = "EJAM analysis",
  buffer_desc = "Selected Locations",
  radius_or_buffer_in_miles = NULL,
  hyperlink_cols = NULL,
  testing = FALSE,
  launchexcel = FALSE,
  saveas = NULL,
  heatmap_colnames = NULL,
  heatmap_cuts = c(80, 90, 95),
  heatmap_colors = c("yellow", "orange", "red"),
  graycolnums = NULL,
  narrowcolnums = NULL,
  graycolor = "gray",
  narrow6 = 6,
  ...
)
}
\arguments{
\item{overall}{table to save in one tab, from EJAM analysis of indicators overall (one row)}

\item{eachsite}{table to save in another tab, from EJAM analysis site by site (one row per site)}

\item{longnames}{vector of indicator names to display in Excel table}

\item{plotfilename}{the full path including name of .png file to insert}

\item{summary_plot}{optional plot passed from EJAM shiny app to save in 'Plot' sheet of Excel table}

\item{plotlatest}{optional logical. If TRUE, the most recently displayed plot will be inserted into a tab called plot2}

\item{analysis_title}{optional title passed from Shiny app to 'Notes' sheet}

\item{buffer_desc}{optional description of buffer used in analysis, passed to 'Notes' sheet}

\item{hyperlink_cols}{which to treat as URLs that should be hyperlinks}

\item{testing}{optional for testing only}

\item{launchexcel}{Set to TRUE to have this function launch Excel immediately, showing the final workbook created here.}

\item{saveas}{If not NULL, and a valid path with filename.xlsx is provided,
the workbook will be saved locally at that path and name. Warning: it will overwrite an existing file.}

\item{heatmap_colnames}{optional vector of colnames to apply heatmap colors}

\item{heatmap_cuts}{vector of values to separate heatmap colors, between 0-100}

\item{heatmap_colors}{vector of color names for heatmap bins, same length as
heatmap_cuts, where first color is for those >= 1st cutpoint, but <2d,
second color is for those >=2d cutpoint but <3d, etc.}

\item{graycolnums}{which column numbers to deemphasize}

\item{narrowcolnums}{which column numbers to make narrow}

\item{graycolor}{color used to deemphasize some columns}

\item{narrow6}{how narrow}

\item{...}{other params passed along to \code{\link[openxlsx:writeData]{openxlsx::writeData()}}}

\item{radius_miles}{If provided, miles buffer distance (from polygon or from point if circular buffers)}
}
\value{
a workbook, ready to be saved in spreadsheet format, with tabs like "Overall" and "Each Site"
}
\description{
xls_formatting2
Format EJAM tabular outputs for saving as Excel spreadsheet
}
\details{
Already took and put here most or all of code from xls_formatting() or xls_formatting_api()
}
\examples{
\dontrun{
  wb <- xls_formatting2(testoutputs_$results_overall, datasetResults()$results_bysite,
    saveas =  "results.xlsx")
}
}
