% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run_EJAM.R
\name{app_run_EJAM}
\alias{app_run_EJAM}
\alias{run_app}
\title{app_run_EJAM - identical to run_app(), just an alias}
\usage{
app_run_EJAM(
  onStart = NULL,
  options = list(),
  enableBookmarking = "server",
  uiPattern = "/",
  ...
)
}
\description{
launch Shiny web app from RStudio
}
