% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties_as_sites.R
\name{bg_from_county}
\alias{bg_from_county}
\title{Analyze US Counties as if they were sites, to get EJ indicators summary for each county}
\usage{
bg_from_county(fips)
}
\arguments{
\item{fips}{County FIPS vector as character not numeric values}
}
\value{
data.table with all pairs of county fips - bgid, and a unique siteid assigned to each county
}
\description{
Analyze US Counties as if they were sites, to get EJ indicators summary for each county
}
\examples{
 bg_from_county(c('01001','72153'))
 # Largest US Counties by ACS Population Totals:
 blockgroupstats[ , .(ST = ST[1], countypop = sum(pop)),
  by=.(FIPS = substr(bgfips,1,5))][order(-countypop),][1:20, .(
  CountyPopulation = prettyNum(countypop, big.mark = ","), FIPS, ST)]
 
}
