% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgej.R
\docType{data}
\name{bgej}
\alias{bgej}
\title{bgej (DATA) EJScreen EJ Indexes for Census block groups}
\description{
bgej (DATA) EJScreen EJ Indexes for Census block groups
}
\details{
\itemize{
\item As of 08/2023 it was the EJScreen 2.2 version of data, which used ACS 2017-2021.
}

Each year this should be re-created as for the latest version.

See \url{https://www.epa.gov/ejscreen}
}
