% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgej.R
\docType{data}
\name{bgej}
\alias{bgej}
\title{bgej (DATA) EJScreen EJ Indexes for Census block groups}
\description{
bgej (DATA) EJScreen EJ Indexes for Census block groups
}
\details{
For documentation on the demographic and environmental data and indicators used by
EJScreen and EJAM, see \url{https://www.epa.gov/ejscreen/understanding-ejscreen-results}.
}
