% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockgroupstats.R
\docType{data}
\name{blockgroupstats}
\alias{blockgroupstats}
\title{EJScreen demographic and enviromental indicators for Census block groups}
\description{
The EJScreen dataset (demographic, environmental, EJ indicators),
plus more demographic subgroups.
}
\details{
As of 10/2022 it is the EJScreen 2.1 version of data, which uses ACS 2016-2020.
EJScreen 2.1 was released October 2022.
As of  4/2022 it was the EJScreen 2.0 version of data, which used ACS 2015-2019.
EJScreen 2.0 was released 2/18/2022 (raw data download avail 2/22/2022).

NOTE: It also has the race/ethnic subgroups that add up to minority or people of color.

Each year this could be created as for the latest version.
See attributes(blockgroupstats)
and see EJSCREEN_Full_with_AS_CNMI_GU_VI  # which is in the package EJAMejscreendata
but there are differences in which columns are kept.

It is a data.table of US Census blockgroups (not blocks). With PR, 242,335 rows, approx 175 columns.
See \url{https://www.epa.gov/ejscreen}

column names include bgfips, bgid (for join to blockwt$bgid), pop, pctlowinc, pcthisp, etc.

see source code and notes in EJAM/inst/notes_datasets/ which has  create_blockgroupstats()

See maybe the notes on cleaning up and changing the dataset starting from ejscreen::bg22plus
}
