% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_community_report.R
\name{build_community_report}
\alias{build_community_report}
\title{build_community_report}
\usage{
build_community_report(
  output_df,
  analysis_title,
  totalpop,
  locationstr,
  include_ejindexes = FALSE,
  in_shiny = FALSE,
  filename = NULL
)
}
\arguments{
\item{output_df, }{single row of results table from doaggregate - either results_overall or one row of bysite}

\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{include_ejindexes, }{whether to build tables for EJ and EJ supp. indexes}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}

\item{filename, }{path to file to save HTML content to; if null, returns as string (used in Shiny app)}
}
\description{
build_community_report
}
