% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sites_with_n_high_scores.R
\name{count_sites_with_n_high_scores}
\alias{count_sites_with_n_high_scores}
\title{count_sites_with_n_high_scores}
\usage{
count_sites_with_n_high_scores(
  scores,
  thresholds = c(1.01, 1.5, 2, 3, 5, 10),
  xwide = c("statewide", "nationwide")[1]
)
}
\arguments{
\item{scores}{score}

\item{thresholds}{thresholds}

\item{xwide}{xwide}
}
\description{
count_sites_with_n_high_scores
}
