% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{counties_as_sites}
\alias{counties_as_sites}
\title{Analyze US Counties as if they were sites, to get EJ indicators summary for each county}
\usage{
counties_as_sites(fips)
}
\arguments{
\item{fips}{County FIPS vector (ideally as character not numeric values),
like fips_counties_from_state_abbrev("DE")}
}
\value{
provides table similar to the output of getblocksnearby(),
data.table with one row per blockgroup in these counties, or
all pairs of county fips - bgid, and ejam_uniq_id (1 through N) assigned to each county
but missing blockid and distance so not ready for doaggregate().
}
\description{
Analyze US Counties as if they were sites, to get EJ indicators summary for each county
}
\details{
This function provides one row per blockgroup.
\code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}} provides one row per block.
See more below under "Value"
}
\examples{
 counties_as_sites(c('01001','72153'))
 # Largest US Counties by ACS Population Totals:
 blockgroupstats[ , .(ST = ST[1], countypop = sum(pop)),
  by=.(FIPS = substr(bgfips,1,5))][order(-countypop),][1:20, .(
  CountyPopulation = prettyNum(countypop, big.mark = ","), FIPS, ST)]
 
}
\seealso{
\code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}}
}
