% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_local.R
\name{dataload_from_local}
\alias{dataload_from_local}
\title{dataload_from_local
utility for R analysts/developers to store large block data locally instead of redownloading from AWS}
\usage{
dataload_from_local(
  varnames = c("bgid2fips", "blockid2fips", "blockpoints", "blockwts", "quaddata"),
  ext = c(".arrow", ".rda")[1],
  fun = c("arrow::read_ipc_file", "load")[1],
  envir = globalenv(),
  folder_local_source = "~/../Downloads",
  justchecking = FALSE,
  testing = FALSE
)
}
\arguments{
\item{varnames}{use defaults}

\item{ext}{use defaults}

\item{fun}{use defaults}

\item{envir}{use defaults}

\item{folder_local_source}{Your local folder path}

\item{justchecking}{use defaults}

\item{testing}{use defaults}
}
\description{
dataload_from_local
utility for R analysts/developers to store large block data locally instead of redownloading from AWS
}
\details{
rm(bgid2fips, blockid2fips, blockpoints, blockwts, quaddata)

dataload_from_local(folder_local_source = '.')
}
