% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapack.R
\name{datapack}
\alias{datapack}
\title{See info about the data sets in one or more packages
Wrapper for data() - just shows info in console and silently returns a data.frame}
\usage{
datapack(pkg = ejampackages, len = 30)
}
\arguments{
\item{pkg}{a character vector giving the package(s) to look in for data sets}

\item{len}{Only affects what is printed to console - specifies the
number of characters to limit Title to, making it easier to see in the console.}
}
\value{
data.frame with Item and Title as columns
}
\description{
See info about the data sets in one or more packages
Wrapper for data() - just shows info in console and silently returns a data.frame
}
\examples{
 datapack("datasets")
 datapack("MASS")
 datapack(ejampackages)
}
