% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{distance_by_group}
\alias{distance_by_group}
\title{See plot_distance_mean_by_group()}
\usage{
distance_by_group(
  results_bybg_people,
  demogvarname = NULL,
  demoglabel = fixcolnames(demogvarname, "r", "shortlabel"),
  returnwhat = "table",
  graph = FALSE
)
}
\value{
See \code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}}
}
\description{
See plot_distance_mean_by_group()
}
