% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{distance_by_group}
\alias{distance_by_group}
\title{distance_by_group - Avg distance of each demog group (of multiple groups)
Same as \code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}} but no plot by default}
\usage{
distance_by_group(
  results_bybg_people,
  demogvarname = "Demog.Index",
  demoglabel = demogvarname
)

distance_by_group(
  results_bybg_people,
  demogvarname = "Demog.Index",
  demoglabel = demogvarname
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{demogvarname}{e.g., "pctlowinc"}

\item{demoglabel}{e.g., "Low Income Residents"}
}
\value{
list of 2 numbers: avg_distance_for_group and avg_distance_for_nongroup
}
\description{
Note ratio shown is ratio of distance among others to distance in group,
so values below 1 mean the given demographic group lives closer to facilities.
A value of 0.85 would mean the group is only 85\% as far from a site as
@details Note it is in miles assuming input was in miles, and the distance for each resident is
actually the average distance of all residents within their Census block (not block group),
and when a site is very close to the block internal point (like a centroid)
relative to the size of the block, the distance to the average resident in the block is
estimated as 90 percent of the effective radius, which is what the radius of the block
would be if it were the same area in square meters or miles but circular in shape.
This is the approach used in EJScreen to estimate average proximity of a block resident in
cases where the block is extremely close to the site or the site may actually be inside the block,
or exactly on top of the internal point of the block, in which case zero would not be an
appropriate estimate of the distance, hence this adjustment is made in EJAM getblocksnearby()
}
\details{
see examples in \code{\link[=plot_distance_cdf_by_group]{plot_distance_cdf_by_group()}}

Note on Avg Distance and range of distances in each Demog group, & \%D as function of distance:

We have info on each blockgroup near each site, which means some small \% of those bgs are duplicated in this table:

\if{html}{\out{<div class="sourceCode">}}\preformatted{results_bybg_people
}\if{html}{\out{</div>}}

Mostly we want overall (not by site) to know avg and cum distrib of distances in each demog,

(and also \%D as a function of continuous distance),

and for those stats we would want to take only unique blockgroups from here,
using the shorter distance, so the distribution of distances does not doublecount people.

But we might also want to see that distribution of distances by D for just 1 site?

And we might also want to see the \%D as a function of continuous distance at just 1 site?

So to retain flexibility doaggregate() reports all instances of blockgroup-site pairings.
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plot(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}}  \code{\link[=distance_by_group]{distance_by_group()}}
}
