% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{distance_by_group}
\alias{distance_by_group}
\title{distance_by_group  -  JUST ONE GROUP
Get average distance for ONE demographic group versus everyone else}
\usage{
distance_by_group(
  results_bybg_people,
  demogvarname = "Demog.Index",
  demoglabel = demogvarname
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{demogvarname}{e.g., "pctlowinc"}

\item{demoglabel}{e.g., "Low Income Residents"}
}
\value{
list of 2 numbers: avg_distance_for_group and avg_distance_for_nongroup
}
\description{
distance_by_group  -  JUST ONE GROUP
Get average distance for ONE demographic group versus everyone else
}
\details{
Note on Avg Distance and range of distances in each Demog group, & \%D as function of distance:

We have info on each blockgroup near each site, which means some small \% of those bgs are duplicated in this table:

\if{html}{\out{<div class="sourceCode">}}\preformatted{results_bybg_people
}\if{html}{\out{</div>}}

Mostly we want overall (not by site) to know avg and cum distrib of distances in each demog,

(and also \%D as a function of continuous distance),

and for those stats we would want to take only unique blockgroups from here,
using the shorter distance, so the distribution of distances does not doublecount people.

But we might also want to see that distribution of distances by D for just 1 site?

And we might also want to see the \%D as a function of continuous distance at just 1 site?

So to retain flexibility doaggregate() reports all instances of blockgroup-site pairings.
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plots(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=plot_distance_mean_by_group]{plot_distance_mean_by_group()}}  \code{\link[=distance_by_groups]{distance_by_groups()}}
}
