% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_mean_by_group.R
\name{distance_by_group}
\alias{distance_by_group}
\title{distance_by_group
Get average distance for ONE demographic group versus everyone else}
\usage{
distance_by_group(
  results_bybg_people,
  demogvarname = "Demog.Index",
  demoglabel = demogvarname
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{demogvarname}{e.g., "pctlowinc"}

\item{demoglabel}{e.g., "Low Income Residents"}
}
\value{
list of 2 numbers: avg_distance_for_group and avg_distance_for_nongroup
}
\description{
distance_by_group
Get average distance for ONE demographic group versus everyone else
}
\details{
Note on Avg Distance and range of distances in each Demog group, & \%D as function of distance:

We have info on each blockgroup near each site, which means some small \% of those bgs are duplicated in this table:

\if{html}{\out{<div class="sourceCode">}}\preformatted{results_bybg_people
}\if{html}{\out{</div>}}

Mostly we want overall (not by site) to know avg and cum distrib of distances in each demog,

(and also \%D as a function of continuous distance),

and for those stats we would want to take only unique blockgroups from here,
using the shorter distance, so the distribution of distances does not doublecount people.

But we might also want to see that distribution of distances by D for just 1 site?

And we might also want to see the \%D as a function of continuous distance at just 1 site?

So to retain flexibility doaggregate() reports all instances of blockgroup-site pairings.
}
\seealso{
\code{\link[=distance_by_group_plot]{distance_by_group_plot()}} \code{\link[=distance_by_groups]{distance_by_groups()}}
}
