% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{distance_cdf_by_group_plot}
\alias{distance_cdf_by_group_plot}
\title{What percentage of this demographic group's population lives less than X miles from a site?}
\usage{
distance_cdf_by_group_plot(
  results_bybg_people,
  radius_miles = round(max(results_bybg_people$distance_min_avgperson, na.rm = T), 1),
  demogvarname = "Demog.Index",
  demoglabel = demogvarname,
  color1 = "red",
  color2 = "black"
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{radius_miles}{miles radius that was max distance analyzed}

\item{demogvarname}{name of column in results_bybg_people, e.g., "pctlowinc"}

\item{demoglabel}{friendly text name for labelling graphic, like "Low income residents"}

\item{color1}{color like "red" for demographic group of interest}

\item{color2}{color like "gray" for everyone else}
}
\value{
invisibly returns full table of sorted distances of blockgroups, cumulative count of demog group at that block group's distance,
and cumulative count of everyone else in that block group
}
\description{
This plots the cumulative share of residents found within each distance,
for a single demographic group.

This function, distance_cdf_by_group_plot(), is based on ejamit()$results_bybg_people,
which provides only block group resolution information about distance.
For block resolution analysis of distance by group, see \code{\link[=plot_distance_by_pctd]{plot_distance_by_pctd()}}.
}
\details{
The function distance_cdf_by_group_plot is SLOW - ***needs to be optimized
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people,
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plot(y$results_bybg_people,
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = distance_by_group_plot(y$results_bybg_people) #
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile,
   compared to 18.7\% of Asian who do.

}
\seealso{
\code{\link[=distance_by_group]{distance_by_group()}} \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} for examples
}
