% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{distance_cdf_by_group_plot}
\alias{distance_cdf_by_group_plot}
\title{distance_cdf_by_group_plot - SLOW - needs to be optimized
Plot a graphic showing cumulative shares of ONE demographic group that are within each distance}
\usage{
distance_cdf_by_group_plot(
  results_bybg_people,
  radius_miles = round(max(x$distance_min_avgperson, na.rm = T), 1),
  demogvarname = "Demog.Index",
  demoglabel = demogvarname,
  color1 = "red",
  color2 = "black"
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{radius_miles}{miles radius that was max distance analyzed}

\item{demogvarname}{name of column in results_bybg_people, e.g., "pctlowinc"}

\item{demoglabel}{friendly text name for labelling graphic, like "Low income residents"}

\item{color1}{color like "red" for demographic group of interest}

\item{color2}{color like "gray" for everyone else}
}
\value{
invisibly returns full table of sorted distances of blockgroups, cumulative count of demog group at that block group's distance,
and cumulative count of everyone else in that block group
}
\description{
distance_cdf_by_group_plot - SLOW - needs to be optimized
Plot a graphic showing cumulative shares of ONE demographic group that are within each distance
}
\seealso{
\code{\link[=distance_by_group]{distance_by_group()}} \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} for examples
}
