% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doaggregate.R
\name{doaggregate}
\alias{doaggregate}
\title{Summarize indicators in each buffer (given the blocks in each buffer and indicators for each block)}
\usage{
doaggregate(
  sites2blocks,
  sites2states_or_latlon = NA,
  countcols = NULL,
  popmeancols = NULL,
  calculatedcols = NULL,
  testing = FALSE,
  include_ejindexes = FALSE,
  ...
)
}
\arguments{
\item{sites2blocks}{data.table of distances in miles between all sites (facilities) and
nearby Census block internal points, with columns siteid, blockid, distance,
created by getblocksnearby  function.
See sites2blocks_example dataset in package, as input to this function}

\item{sites2states_or_latlon}{data.table or just data.frame, with columns siteid (each unique one in sites2blocks) and ST (2-character State abbreviation) or lat and lon}

\item{countcols}{character vector of names of variables  to aggregate within a buffer
using a sum of counts, like, for example, the number of people for whom a
poverty ratio is known, the count of which is the exact denominator needed
to correctly calculate percent low income.}

\item{popmeancols}{character vector of names of variables to aggregate within a buffer
using population weighted mean.}

\item{calculatedcols}{character vector of names of variables to aggregate within a buffer
using formulas that have to be specified.}

\item{testing}{used while testing this function}

\item{include_ejindexes}{not yet implemented}

\item{...}{more to pass to another function? Not used currently.}
}
\description{
This updated 2022 code takes a set of facilities and the set of blocks that are near each,
(as identified previously, in other code that has identified which blocks are nearby)
and combines those with indicator scores for block groups.
}
\details{
This function aggregates the blockgroup scores to create a summary of each indicator,
as a raw score and US percentile and State percentile,
in each buffer (i.e., near each facility):
\itemize{
\item \strong{SUMS OF COUNTS}: for population count, or number of households or Hispanics, etc.
\item \strong{POPULATION-WEIGHTED MEANS}: for  Environmental indicators.

\emph{\strong{EJ Indexes}:} These could be in theory recalculated via formula, but the way EJScreen
does this is apparently finding the pop wtd mean of EJ Index raw scores,
not the EJ Index formula applied to the summarized demographic score and aggregated envt number.
\item \strong{CALCULATED BY FORMULA}: Buffer or overall score calculated via formulas using aggregated counts,
such as percent low income = sum of counts low income / sum of counts of denominator,
which in this case is the count of those for whom the poverty ratio is known.
\item \strong{LOOKED UP}: Aggregated scores are converted into percentile terms via lookup tables (US or State version).
}

This function requires the following as data lazy loaded for example from EJAMblockdata package:
\itemize{
\item blockwts: data.table with these columns: blockid , bgid, blockwt
\item quaddata, and blockquadtree: data.table and quad tree, for indexes of block points
(and localtree that is created when package is loaded)
\item EJAM::blockgroupstats - A data.table (such as EJSCREEN demographic and environmental data by blockgroup?)
}
}
\examples{
\dontrun{
  testsites <- EJAMfrsdata::frs[sample(1:nrow(EJAMfrsdata::frs), 1e3),]
  x <- getblocksnearby(testsites, cutoff = 3.1, quadtree = localtree)
  stinfo <- data.frame(siteid=1:nrow(testsites),  state_from_latlon(lat=testsites$lat, lon=testsites$lon))
  y <- doaggregate(x, sites2states = stinfo)


}
}
