% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_latest_arrow_data.R
\name{download_latest_arrow_data}
\alias{download_latest_arrow_data}
\title{Download latest versions of arrow datasets if user doesn't have them}
\usage{
download_latest_arrow_data(
  varnames = c("blockpoints", "blockwts", "quaddata", "bgej", "bgid2fips",
    "blockid2fips", "frs", "frs_by_mact", "frs_by_naics", "frs_by_programid",
    "frs_by_sic")
)
}
\arguments{
\item{varnames}{use defaults, or vector of names like "bgej" or use "all" to get all available}
}
\description{
Used when EJAM is loaded
}
