% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdupefilenames.R
\name{dupenames}
\alias{dupenames}
\title{look at several packages to spot conflicting exported names (functions or data)
See what objects are exported by a given package}
\usage{
dupenames(pkg = EJAM::ejampackages, sortbypkg = FALSE)
}
\arguments{
\item{pkg}{one or more package names as vector of strings.
If "all" it checks all installed pkgs, but takes very very long potentially.}

\item{sortbypkg}{If TRUE, just returns same thing but sorted by package name}
}
\value{
data.frame with columns Package, Object name (or NA if no dupes)
}
\description{
look at several packages to spot conflicting exported names (functions or data)
See what objects are exported by a given package
}
\details{
This can help find duplicates/conflicts within source code
and make sure they are on search path, for when renaming / moving functions/packages
}
