% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames.R
\name{dupenames}
\alias{dupenames}
\title{dupenames  -  helper to look at several packages to spot conflicting exported names
See what objects (functions or data) are exported by a given (installed) package}
\usage{
dupenames(
  pkg = EJAM::ejampackages,
  sortbypkg = FALSE,
  compare.functions = TRUE
)
}
\arguments{
\item{pkg}{one or more package names as vector of strings.
If "all" it checks all installed pkgs, but takes very very long potentially.}

\item{sortbypkg}{If TRUE, just returns same thing but sorted by package name}

\item{compare.functions}{If TRUE, sends to console inf about whether body and formals
of the functions are identical between functions of same name from different packages.
Only checks the first 2 copies, not any additional ones (where 3+ pkgs use same name)}
}
\value{
data.frame with columns Package, Object name (or NA if no dupes)
}
\description{
dupenames  -  helper to look at several packages to spot conflicting exported names
See what objects (functions or data) are exported by a given (installed) package
}
\details{
This can help find duplicates/conflicts within source code
and make sure they are on search path, for when renaming / moving functions/packages
}
\seealso{
\code{\link[=all_equal_functions]{all_equal_functions()}}
}
