% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames.R
\name{dupeRfiles}
\alias{dupeRfiles}
\title{dupeRfiles - UTILITY - check conflicting sourcefile names}
\usage{
dupeRfiles(folder1 = "../EJAM/R", folder2 = "./R")
}
\arguments{
\item{folder1}{path to other folder with R source files}

\item{folder2}{path to a folder with R source files, defaults to "./R"}
}
\description{
See what same-named .R files are in 2 sourcecode folders
}
\details{
useful for shiny app that is not a package,
as ejamlite and EJAMejscreenapi had copies of some EJAM files.

See \code{\link[=dupenames]{dupenames()}} for when they are all packages.
}
