% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2excel.R
\name{ejam2excel}
\alias{ejam2excel}
\title{Save EJAM results in a spreadsheet}
\usage{
ejam2excel(
  ejamitout,
  fname = NULL,
  save_now = TRUE,
  overwrite = TRUE,
  launchexcel = FALSE,
  interactive_console = TRUE,
  ok2plot = TRUE,
  in.testing = FALSE,
  in.analysis_title = "EJAM analysis",
  react.v1_summary_plot = NULL,
  radius_or_buffer_in_miles = NULL,
  buffer_desc = "Selected Locations",
  radius_or_buffer_description =
    "Miles radius of circular buffer (or distance used if buffering around polygons)",
  hyperlink_colnames = c("EJScreen Report", "EJScreen Map", "ECHO report"),
  ...
)
}
\arguments{
\item{ejamitout}{output of \code{\link[=ejamit]{ejamit()}}}

\item{fname}{optional name or full path and name of file to save locally, like "out.xlsx"}

\item{save_now}{optional logical, whether to save as a .xlsx file locally or just return workbook object
that can later be written to .xlsx file using \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{overwrite}{optional logical, passed to \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{launchexcel}{optional logical, passed to \code{\link[=table_xls_format]{table_xls_format()}}, whether to launch browser to see spreadsheet immediately}

\item{interactive_console}{optional - should set to FALSE when used in code or server. If TRUE,
prompts RStudio user interactively asking where to save the downloaded file}

\item{ok2plot}{optional logical, passed to  \code{\link[=table_xls_format]{table_xls_format()}}, whether safe to try and plot or set FALSE if debugging plot problems}

\item{in.testing}{optional logical}

\item{in.analysis_title}{optional title as character string}

\item{react.v1_summary_plot}{optional - a plot object}

\item{radius_or_buffer_in_miles}{optional radius in miles}

\item{radius_or_buffer_description}{optional text phrase describing places analyzed}

\item{hyperlink_colnames}{optional names of columns with URLs}

\item{...}{optional additional parameters passed to \code{\link[=table_xls_format]{table_xls_format()}}, such as
heatmap_colnames, heatmap_cuts, heatmap_colors, etc.}
}
\value{
same as what \code{\link[=table_xls_from_ejam]{table_xls_from_ejam()}} returns
}
\description{
ejam2excel() takes the output of something like ejamit() and
creates a spreadsheet with an overall summary tab, a site by site table tab,
as well as other tabs such as map, plot, notes, etc.
It is an alias for table_xls_from_ejam()
}
