% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2shapefile.R
\name{ejam2shapefile}
\alias{ejam2shapefile}
\title{DRAFT - export EJAM results as shapefile for use in ArcPro, EJScreen, etc.}
\usage{
ejam2shapefile(out, fname = "bysite.shp", folder = ".", crs = 4269)
}
\arguments{
\item{out}{output of EJAM such as from \code{\link[=ejamit]{ejamit()}}}

\item{fname}{a filename no path and must be .shp extension}

\item{folder}{optional folder (directory), path to where it should be saved}

\item{crs}{optional coord ref system}
}
\value{
path to saved .zip file
}
\description{
DRAFT - export EJAM results as shapefile for use in ArcPro, EJScreen, etc.
}
\details{
see
https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm
}
\examples{
  # x <- ejam2shapefile(ejamit(testpoints_100 , radius = 3.1), 
  #    fname = "test100_3miles.shp", folder = "~/../Downloads")
  out = testoutput_ejamit_10pts_1miles
  # ejam2shapefile(out,
  #   fname = "ejamoutput.shp", folder = "~/../Downloads")
}
