% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_groups_of_places.R
\name{ejamit_compare_types_of_places}
\alias{ejamit_compare_types_of_places}
\title{Compare subsets (types) of places that are all from one list}
\usage{
ejamit_compare_types_of_places(sitepoints, typeofsite = NULL, ...)
}
\arguments{
\item{sitepoints}{see \code{\link[=ejamit]{ejamit()}}}

\item{typeofsite}{vector of length same as NROW(sitepoints), where each unique value defines a group of sites}

\item{...}{see \code{\link[=ejamit]{ejamit()}}}
}
\value{
similar to ejamit output but results_overall has one row per unique typeofsite
}
\description{
*** DRAFT - May change but works as currently drafted.
e.g., change output formats of results_bytype vs results_overall
}
\examples{
  out <- ejamit_compare_types_of_places(testpoints_10[1:4, ], 
  typeofsite <- c("A", "B", "B", "C"))
  cbind(Rows_or_length = sapply(out, NROW))
  
  x <- EJAM::state_from_latlon(lat = testpoints_1000$lat, lon = testpoints_1000$lon)
  x <- data.frame(testpoints_1000, x)
  out_bystate <- ejamit_compare_types_of_places(x, x$ST)
  
}
