% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_see1map}
\alias{ejscreen_vs_ejam_see1map}
\title{EJAM/EJSCREEN comparisons - see map and tables of blocks near a site to explain discrepancy in pop and blocks}
\usage{
ejscreen_vs_ejam_see1map(
  n = 1,
  x,
  overlay_blockgroups = FALSE,
  radius = NULL,
  ...
)
}
\arguments{
\item{n}{row number in x$EJAM of site to check}

\item{x}{results from  x <- ejscreen_vs_ejam(testpoints_10, radius =3, include_ejindexes = TRUE)}

\item{overlay_blockgroups}{optional, set TRUE to see overlay of boundaries of parent blockgroups,
noting that you have to click to turn off the layer for block point info popups to work}

\item{...}{passed to \code{\link[=plotblocksnearby]{plotblocksnearby()}}}
}
\value{
Just draws map and shows tables and returns the table of blocks near the site
and info on how much the distance and pop count differ from what EJAM thinks
is the actual exact radius, with possible explanation of discrepancy
between ejscreen api and ejam estimate,
but you may also want the output of ejscreen_vs_ejam_see1()
}
\description{
EJAM/EJSCREEN comparisons - see map and tables of blocks near a site to explain discrepancy in pop and blocks
}
\examples{
dontrun{
  z <- ejscreen_vs_ejam(testpoints_10, radius = 3, include_ejindexes = TRUE)
  ejscreen_vs_ejam_see1map(3, z, overlay_blockgroups = TRUE)
 }

}
\keyword{internal}
