% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vscript}
\alias{ejscreen_vscript}
\title{EJAM/EJSCREEN comparisons - for interactive RStudio use}
\usage{
ejscreen_vscript(
  defdir = ".",
  n,
  newpts,
  pts,
  radius,
  savedejscreentableoutput,
  x100fix = TRUE,
  x100varnames = names_pct_as_fraction_ejamit
)
}
\arguments{
\item{defdir}{folder}

\item{n}{how many points}

\item{newpts}{logical, if need new set of random locations}

\item{pts}{data.frame of points with columns lat,lon}

\item{radius}{miles}

\item{savedejscreentableoutput}{is a data.table from ejscreenit()$table}
}
\value{
a list of data frames, with names
EJSCREEN, EJAM, EJSCREEN_shown, EJAM_shown, same_shown,
ratio, diff, absdiff, pctdiff, abspctdiff

diff is EJAM - EJSCREEN

ratio is EJAM / EJSCREEN

pctdiff is ratio - 1

abs is absolute value

For each data.frame, colnames are indicators like pop, blockcount_near_site, etc.
and rows represent sites analyzed.
}
\description{
EJAM/EJSCREEN comparisons - for interactive RStudio use
}
\details{
THIS IS A SCRIPT FOR INTERACTIVE RSTUDIO CONSOLE USE
TO RUN A SET OF POINTS THROUGH BOTH EJSCREEN AND EJAM
AND SAVE STATS ON THE DIFFERENCES
-- including letting you use saved ejscreenapi results and input points
so you can iterate and rerun just the EJAM portion and compare to the saved benchmark data.
}
