% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi2ejam_format}
\alias{ejscreenapi2ejam_format}
\title{Convert output of ejscreenapi_plus to format of ejamit table of sites
Makes it easier to compare EJScreen and EJAM results}
\usage{
ejscreenapi2ejam_format(
  ejscreenapi_plus_out,
  fillmissingcolumns = FALSE,
  ejamcolnames = NULL
)
}
\arguments{
\item{ejscreenapi_plus_out}{results of ejscreenapi_plus() or also
could be results of ejscreenit()$table even though the colnames differ,
because they get converted here in that case.}

\item{fillmissingcolumns}{optional. set to TRUE if you want the output
to have exactly all the same columns as the EJAM table would, and
fill in with NA values all the columns not provided by EJScreen API.}

\item{ejamcolnames}{optional. if specified as vector of colnames, it
overrides the assumed colnames that would have been taken to be
colnames(testoutput_ejamit_10pts_1miles$results_bysite).
Any colnames you specify here will be the colnames of the output
if fillmissingcolumns = TRUE, or else those not in names(ejscreenapi_plus_out)
will be omitted.}
}
\value{
A data.table not just data.frame, with some or all of the columns
found in output of ejamit()$results_bysite
}
\description{
Convert output of ejscreenapi_plus to format of ejamit table of sites
Makes it easier to compare EJScreen and EJAM results
}
\examples{
\dontrun{
   x <- ejscreenapi_plus(testpoints_10[1:2, ], radius = 1)
   y <- ejscreenapi2ejam_format(x)
   ejamvars <- names(testoutput_ejamit_10pts_1miles$results_bysite)
   all.equal(
     names(y), 
     ejamvars[ejamvars \%in\% names(y)]
  )
  
  z <- ejscreenapi2ejam_format(x, fillmissingcolumns = T)
  all.equal(names(z), ejamvars)
  
 y2 <- ejamit(testpoints_10[1:2, ], radius = 1)
 y2 <- y2$results_bysite
 ejamvars <- names(y2)
 all.equal(names(z), ejamvars)
 sitenum <- 1
 z <- data.frame(EJAM = t(y2[sitenum, -1:-3]), EJSCREEN = t(z[sitenum, -1:-3 ]))
 z
 z[!is.na(z$EJSCREEN), ]
  }
  
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1]{ejscreenapi_vs_ejam1()}}
}
