% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam1}
\alias{ejscreenapi_vs_ejam1}
\title{compare EJScreen API vs EJAM stats near one site}
\usage{
ejscreenapi_vs_ejam1(latlon, radius = 3, nadrop = FALSE, ...)
}
\arguments{
\item{latlon}{1-row data.table or data.frame with colnames lat and lon}

\item{radius}{in miles, used in ejamit() and ejscreenapi_plus()}

\item{nadrop}{whether to drop indicators for which EJScreen API returns NA}

\item{...}{passed to ejamit() as any additional parameters,
like include_ejindexes = FALSE}
}
\value{
a data.frame with columns EJSCREENAPI and EJAM,
rownames are indicators like pop, blockcount_near_site, etc.
}
\description{
compare EJScreen API vs EJAM stats near one site
}
\examples{
 \dontrun{
  z <- ejscreenapi_vs_ejam1(testpoints_100[5, ], radius = 3)
  z <- ejscreenapi_vs_ejam1(testpoints_100[27, ], radius = 3, nadrop = T)
  z$longname <- fixcolnames(rownames(z), 'r', 'long')
  z[names_these, ]
  }
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1_alreadyrun]{ejscreenapi_vs_ejam1_alreadyrun()}}
}
