% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenapi2ejam_format.R
\name{ejscreenapi_vs_ejam1_alreadyrun}
\alias{ejscreenapi_vs_ejam1_alreadyrun}
\title{compare EJScreen API vs EJAM stats near one site (if results already run)}
\usage{
ejscreenapi_vs_ejam1_alreadyrun(apisite, ejamsite, nadrop = FALSE)
}
\arguments{
\item{apisite}{1-row table output of ejscreenit()$table,
or ejscreenapi_plus()}

\item{ejamsite}{1-row table output of ejamit()$results_bysite}

\item{nadrop}{whether to drop indicators for which EJScreen API returns NA}
}
\value{
a data.frame with columns EJSCREEN and EJAM,
rownames are indicators like pop, blockcount_near_site, etc.
}
\description{
compare EJScreen API vs EJAM stats near one site (if results already run)
}
\examples{
 \dontrun{
 
 # requires data.table
   pts <- testpoints_100[1:5, ]
   #api1 <- ejscreenit(pts, radius = 3)
   api1 <- api1$table[5, ]
   ejam1 <- ejamit(pts, radius = 3, include_ejindexes = TRUE)
   ejam1 <- ejam1$results_bysite[5, ]
   z <- ejscreenapi_vs_ejam1_alreadyrun(api1, ejam1)
   
   # Reported key indicators - which ones do or don't match
   # when comparing EJSCREEN and EJAM results?
   keyreportnames <- c('pop', names_these, names_pctile, names_state_pctile)
   z[z$rname \%in\% keyreportnames &  z$same_shown, ]
   z[z$rname \%in\% keyreportnames & !z$same_shown & !is.na(z$EJSCREEN), ]
   
   # Reported (rounded) numbers match:
   z[z$same_shown , -1]
   # Reports disagree: 
   #  (and not just because of percentages being of 100 vs of 1.00)
   z[!z$same_shown & !is.na(z$EJSCREEN) & z$ratio != 0.01, -1] 
   # Reports disagree if percentages reported as 0-100 vs fractions 0-1.00
   z[z$ratio == 0.01 & !is.na(z$ratio), -1]
 }
}
\seealso{
\code{\link[=ejscreenapi_vs_ejam1]{ejscreenapi_vs_ejam1()}}
}
