% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenit_for_ejam.R
\name{ejscreenit_for_ejam}
\alias{ejscreenit_for_ejam}
\title{Get EJScreen results directly from EJScreen servers via their API}
\usage{
ejscreenit_for_ejam(sitepoints, radius = 3, fillmissingcolumns = TRUE, ...)
}
\arguments{
\item{sitepoints}{table with lat and lon columns}

\item{radius}{in miles}

\item{...}{passed to ejscreenit()}
}
\value{
a data.table that looks like output of ejamit()$results_bysite
}
\description{
A wrapper for EJAM to use ejscreenit()
}
\seealso{
\code{\link[=ejscreen_vs_ejam]{ejscreen_vs_ejam()}} \code{\link[=ejscreenapi2ejam_format]{ejscreenapi2ejam_format()}} which it uses.
}
