% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreenit_for_ejam.R
\name{ejscreenit_for_ejam}
\alias{ejscreenit_for_ejam}
\title{ejscreenit_for_ejam
Wrapper for EJAMejscreenapi::ejscreenit(), to use in EJAM app - NOT USED CURRENTLY}
\usage{
ejscreenit_for_ejam(sitepoints, radius = 3, ...)
}
\arguments{
\item{sitepoints}{table with lat and lon columns}

\item{radius}{in miles}

\item{...}{passed to ejscreenit() but not tested and probably should not use/ not needed}
}
\value{
list of results from \code{\link[EJAMejscreenapi:ejscreenit]{EJAMejscreenapi::ejscreenit()}}
i.e., list of these: results_bysite (data.table), map, plot, us.ratios
}
\description{
ejscreenit_for_ejam
Wrapper for EJAMejscreenapi::ejscreenit(), to use in EJAM app - NOT USED CURRENTLY
}
