% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epa_programs.R
\docType{data}
\name{epa_programs}
\alias{epa_programs}
\title{epa_programs (DATA) named vector with program counts}
\format{
An object of class \code{character} of length 97.
}
\usage{
epa_programs
}
\description{
This is a named vector of EPA programs that lists the programs and has names with programs and program counts. Used by EJAM shiny app for dropdown menu.
}
\details{
Also see \href{https://www.epa.gov/frs/frs-data-sources}{EPA documentation describing each program code} aka data source.
\subsection{For a table sorted by how many facilities from FRS are listed for each program:}{

\code{epa_programs_counts <- frs_by_programid[, .N, by="program"][order(N), ]}
}

\subsection{Top 25 Federal ones (by trying to exclude State databases from the list)}{

\code{epa_programs_counts$fed = !grepl("^[A-Z]{2}[-]", as.vector(epa_programs_counts$program),1,3)}

\code{epa_programs_counts[fed == TRUE, ][order(-N), ][1:25, ]}

Also see \link{frs_by_programid}
}
}
\keyword{datasets}
