% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epa_programs.R
\docType{data}
\name{epa_programs}
\alias{epa_programs}
\title{epa_programs (DATA) named vector with program counts}
\format{
An object of class \code{character} of length 6.
}
\usage{
epa_programs
}
\description{
Named vector of EPA programs and facility count for each.
}
\details{
This lists selected data sources for the facility registry services (FRS) database, as
documented in \href{https://www.epa.gov/frs/frs-data-sources}{EPA documentation describing each FRS program code}.

The named list is used by the shiny app for a dropdown menu used to pick a program.

Also see \link{frs_by_programid}
}
\keyword{datasets}
