% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips2state_abbrev}
\alias{fips2state_abbrev}
\title{fips2state_abbrev - convert Census FIPS to 2-character state abbreviation}
\usage{
fips2state_abbrev(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of abbreviations like "NY", "LA", "DE", etc.
}
\description{
fips2state_abbrev - convert Census FIPS to 2-character state abbreviation
}
\examples{
  cbind(fips_lead_zero(1:80), fips2state_abbrev(1:80), fips2statename(1:80))
}
