% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_x_from_y.R
\name{fips_counties_from_statename}
\alias{fips_counties_from_statename}
\title{Get ALL county fips in specified states}
\usage{
fips_counties_from_statename(statename)
}
\arguments{
\item{statename}{vector of state names like c("New York","Georgia")}
}
\value{
vector of 5-digit character string county FIPS of all unique counties in those states
}
\description{
Get ALL county fips in specified states
}
\examples{
fips_counties_from_statename("Delaware")
  
}
