% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_lead_zero.R
\name{fips_lead_zero}
\alias{fips_lead_zero}
\title{fips_lead_zero
Add leading zeroes to fips codes if missing, replace with NA if length invalid
Note it does NOT VALIDATE FIPS -
It does NOT check if FIPS is valid other than checking its length seems OK,
i.e., it might be a state, county, tract, blockgroup, or block FIPS code.}
\usage{
fips_lead_zero(fips)
}
\arguments{
\item{fips}{vector of numeric or character US FIPS codes}
}
\value{
vector of same length
}
\description{
fips_lead_zero
Add leading zeroes to fips codes if missing, replace with NA if length invalid
Note it does NOT VALIDATE FIPS -
It does NOT check if FIPS is valid other than checking its length seems OK,
i.e., it might be a state, county, tract, blockgroup, or block FIPS code.
}
