% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fipstype.R
\name{fipstype}
\alias{fipstype}
\title{fipstype  -  identify what type of Census geography is each FIPS code (block, county, etc.)}
\usage{
fipstype(fips)
}
\arguments{
\item{fips}{vector of one or more Census FIPS with or without leading zeroes, as strings or numeric}
}
\value{
vector of types: "block", "blockgroup", "tract", "county", or "state"
}
\description{
fipstype  -  identify what type of Census geography is each FIPS code (block, county, etc.)
}
\examples{
 fips_counties_from_statename(c("Connecticut", "Delaware") )
 # [1] "09001" "09003" "09005" "09007" "09009" "09011" "09013" "09015" "10001" "10003" "10005"
 fipstype(9001)
 fipstype("10001")
 # note blockid2fips is a large file, but can be obtained via [dataload_from_pins()]
 \dontrun{
 fipsexamples <- c(
   fips_state_from_statename("Alaska"),
   fips_counties_from_state_abbrev("DE")[1],
    substr(blockid2fips$blockfips[1],1,11),
    blockgroupstats$bgfips[1],
    blockid2fips$blockfips[1]
 )
 cbind(fipsexamples, type = fipstype(fipsexamples))
}
}
