% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_pctcols_x100.R
\name{fix_pctcols_x100}
\alias{fix_pctcols_x100}
\title{utility to multiply certain percentage columns by 100 to convert 0-1.00 into 0-100
multiple some data to rescale percentages stored as 0 to 1, into 0-100}
\usage{
fix_pctcols_x100(df, cnames = NULL)
}
\arguments{
\item{df}{data.frame but can be data.table}

\item{cnames}{colnames in df of indicators to multiply by 100, like those in

names_pct_as_fraction_ejamit,

names_pct_as_fraction_blockgroupstats, or

names_pct_as_fraction_ejscreenit}
}
\value{
df with data in specified columns multiplied by 100
}
\description{
utility to multiply certain percentage columns by 100 to convert 0-1.00 into 0-100
multiple some data to rescale percentages stored as 0 to 1, into 0-100
}
\examples{
    y <- testoutput_ejamit_10pts_1miles$results_bysite
    y1 <- fix_pctcols_x100(y)
    y2 <- fix_pctcols_x100(y,  names_pct_as_fraction_ejscreenit)
    
}
