% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmapheadernamescolname.R
\name{fixmapheadernamescolname}
\alias{fixmapheadernamescolname}
\title{utility to convert aliases to proper colnames of map_headernames}
\usage{
fixmapheadernamescolname(x)
}
\arguments{
\item{x}{character vector of colnames of map_headernames, or aliases like "long"}
}
\value{
vector where aliases are replaced with actual colnames and unmatched ones left as-is
}
\description{
utility to convert aliases to proper colnames of map_headernames
}
\examples{
  fixmapheadernamescolname(c('long', 'csv', 'api', 'r'))

}
