% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_diagnostics}
\alias{getblocks_diagnostics}
\title{getblocks_diagnostics - Get summary stats on counts of blocks (unique vs doublecounted) near sites}
\usage{
getblocks_diagnostics(x, detailed = FALSE, see_plot = FALSE)
}
\arguments{
\item{x}{The output of \code{\link[=getblocksnearby]{getblocksnearby()}} like testoutput_getblocksnearby_10pts_1miles}

\item{detailed}{if TRUE, also shows in console a long table of frequencies via \code{\link[=getblocks_summarize_blocks_per_site]{getblocks_summarize_blocks_per_site()}}}

\item{see_plot}{set TRUE to draw for each site a boxplot of distances of nearby blocks}
}
\value{
A list of stats
}
\description{
getblocks_diagnostics - Get summary stats on counts of blocks (unique vs doublecounted) near sites
}
\examples{
 getblocks_diagnostics(testoutput_getblocksnearby_10pts_1miles)
}
\seealso{
This relies on  \code{\link[=getblocks_summarize_blocks_per_site]{getblocks_summarize_blocks_per_site()}} and \code{\link[=getblocks_summarize_sites_per_block]{getblocks_summarize_sites_per_block()}}
}
