% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_diagnostics}
\alias{getblocks_diagnostics}
\title{getblocks_diagnostics - Lots of stats about # of blocks, # of sites, etc.}
\usage{
getblocks_diagnostics(x, detailed = FALSE, see_pctiles = FALSE)
}
\arguments{
\item{x}{The output of \code{\link[=getblocksnearby]{getblocksnearby()}} like testoutput_getblocksnearby_10pts_1miles}

\item{detailed}{if TRUE, also shows in console a long table of frequencies via \code{\link[=getblocks_summarize_blocks_per_site]{getblocks_summarize_blocks_per_site()}}}

\item{see_pctiles}{set to TRUE to see 20 percentiles of distance in a table}
}
\value{
A list of stats
}
\description{
getblocks_diagnostics - Lots of stats about # of blocks, # of sites, etc.
}
\examples{
  getblocks_diagnostics(testoutput_getblocksnearby_10pts_1miles)
  pts <- rbind(data.table(lat = 40.3, lon = -96.23),
    testpoints_10[ , .(lat, lon)])
 z <- getblocksnearbyviaQuadTree(pts, 1, quadtree = localtree, quiet = T)
 z[ , .(blocks = .N) , keyby = 'siteid']
 plotblocksnearby(pts, radius = 1, sites2blocks = z)
 zz <- getblocks_diagnostics(z, detailed = T, see_pctiles = T)
cbind(stats = zz)

}
\seealso{
This relies on  \code{\link[=getblocks_summarize_blocks_per_site]{getblocks_summarize_blocks_per_site()}} and \code{\link[=getblocks_summarize_sites_per_block]{getblocks_summarize_sites_per_block()}}
}
