% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_diagnostics}
\alias{getblocks_diagnostics}
\title{Get summary stats on counts of blocks (unique vs doublecounted) near sites}
\usage{
getblocks_diagnostics(x)
}
\arguments{
\item{x}{The output of \code{\link[=getblocksnearby]{getblocksnearby()}} like sites2blocks_example}
}
\value{
A list of stats
}
\description{
Get summary stats on counts of blocks (unique vs doublecounted) near sites
}
\examples{
 getblocks_diagnostics(sites2blocks_example)
}
