% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby.R
\name{getblocksnearby}
\alias{getblocksnearby}
\title{Key buffering function - wrapper redirecting to the right version of the getblocksnearby function}
\usage{
getblocksnearby(
  sitepoints,
  cutoff = 1,
  maxcutoff = 31.07,
  uniqueonly = FALSE,
  avoidorphans = TRUE,
  quadtree,
  ...
)
}
\arguments{
\item{sitepoints}{see \link{getblocksnearbyviaQuadTree} or other such functions}

\item{cutoff}{see \link{getblocksnearbyviaQuadTree} or other such functions}

\item{maxcutoff}{see \link{getblocksnearbyviaQuadTree} or other such functions}

\item{uniqueonly}{see \link{getblocksnearbyviaQuadTree} or other such functions}

\item{avoidorphans}{see \link{getblocksnearbyviaQuadTree} or other such functions}

\item{...}{see \link{getblocksnearbyviaQuadTree_Clustered} or other such functions}
}
\description{
As written it assumes that certain things are already in global environment.
}
