% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearby.R
\name{getblocksnearby}
\alias{getblocksnearby}
\title{Fast way to find nearby points - finds distance to each Census block centroid nearby}
\usage{
getblocksnearby(
  sitepoints,
  cutoff = 1,
  maxcutoff = 31.07,
  avoidorphans = TRUE,
  quadtree,
  ...
)
}
\arguments{
\item{sitepoints}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{cutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{maxcutoff}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{avoidorphans}{see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}} or other such functions}

\item{quadtree}{a large quadtree object created from the SearchTree package example:
SearchTrees::createTree(EJAMblockdata::quaddata, treeType = "quad", dataType = "point")}

\item{...}{see \code{\link[=getblocksnearbyviaQuadTree_Clustered]{getblocksnearbyviaQuadTree_Clustered()}} or other such functions}
}
\description{
Fast way to find nearby points - finds distance to each Census block centroid nearby
}
\details{
wrapper redirecting to the right version like \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}}
Census block "internal points" are actually what it looks for, like a centroid.
The blocks are pre-indexed for the whole USA, via the data object quadtree or localtree
}
\seealso{
\code{\link[=getblocksnearby2]{getblocksnearby2()}} that was work in progress
}
