% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{map_blockgroups_over_blocks}
\alias{map_blockgroups_over_blocks}
\title{map_blockgroups_over_blocks - Overlay blockgroups near 1 site, after plotblocksnearby()
Overlay blockgroups near 1 site, after plotblocksnearby()}
\usage{
map_blockgroups_over_blocks(y)
}
\arguments{
\item{y}{output of \code{\link[=plotblocksnearby]{plotblocksnearby()}}}
}
\value{
leaflet map widget
}
\description{
map_blockgroups_over_blocks - Overlay blockgroups near 1 site, after plotblocksnearby()
Overlay blockgroups near 1 site, after plotblocksnearby()
}
\examples{
dontrun{
 y <- plotblocksnearby(testpoints_10[5,], 
        radius = 3,
        returnmap = TRUE)
 map_blockgroups_over_blocks(y)
  }
}
\seealso{
\code{\link[=map_blockgroups]{map_blockgroups()}}
}
