% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_blockgroups.R
\name{map_shapes_leaflet}
\alias{map_shapes_leaflet}
\title{Create a new leaflet map from shapefile data}
\usage{
map_shapes_leaflet(shapes, color = "green", popup = shapes$NAME)
}
\arguments{
\item{shapes}{like from shapes_counties_from_countyfips(fips_counties_from_state_abbrev("DE"))}

\item{color}{passed to leaflet::addPolygons()}

\item{popup}{passed to leaflet::addPolygons()}
}
\value{
html widget from leaflet::leaflet()
}
\description{
Create a new leaflet map from shapefile data
}
