% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfast.R
\name{mapfast}
\alias{mapfast}
\title{Get a leaflet html widget map of data.frame with lat lon}
\usage{
mapfast(mydf, radius = 3, column_names = "all", labels = column_names)
}
\arguments{
\item{mydf}{data.frame or data.table with lat and lon columns or
columns that \code{\link[=latlon_infer]{latlon_infer()}} can infer to be that}

\item{radius}{in miles, converted to meters and passed to leaflet::addCircles()}

\item{column_names}{If "ej" then nice popup made based on just key EJScreen
indicators. If "all" then every column in the entire mydf table is shown
in the popup. If a vector of colnames, only those are shown in popups.}

\item{labels}{The labels used before the column_names, for map popups,
like  label: column_name  (ignored if column_names is ej or all)}
}
\value{
plots a leaflet map with popups with all the columns from mydf,
and returns html widget
}
\description{
Get a leaflet html widget map of data.frame with lat lon
}
\examples{
 mapfast(testpoints_500)
 \dontrun{
 # see random sample of 1k FRS facilities
 if (!exists("frs")) {
   x <- try(require(EJAM)); if (!inherits(x, "try-error")) {
     dataload_from_pins("frs")
   }
   mydf <-  frs[sample(1:NROW( frs), 1000), 1:5]
   mapfast(mydf)
 }
 # out <- EJAMejscreenapi::testoutput_ejscreenit_50$table
 # names(out) <- fixcolnames(names(out), 'long', 'r')
 out <- testoutput_ejscreenapi_plus_50
 mapfastej(out)
 mapfast(out, column_names = 'ej')
 mapfast(out)

 # Save .html file and view it in your browser

 mytable <- testoutput_ejscreenit_5$table
 # names(mytable) <- fixcolnames(names(mytable), 'long', 'r')
 ## now, that renaming is done by mapfast function
 mymap <- mapfastej(mytable)
 tfile <- paste0(tempfile(), '.html')
 htmltools::save_html(mymap, file = tfile)
 browseURL(tfile)
  }

}
\seealso{
\code{\link[=popup_from_df]{popup_from_df()}} \code{\link[=mapfastej]{mapfastej()}}
}
