% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfast.R
\name{mapfastej}
\alias{mapfastej}
\title{Get a leaflet html widget map of data.frame with lat lon}
\usage{
mapfastej(...)
}
\value{
like what \code{\link[=mapfast]{mapfast()}} returns
}
\description{
Like \code{\link[=mapfast]{mapfast()}} but with column_names = "ej"
}
