% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{mapfastej_counties}
\alias{mapfastej_counties}
\title{Map - County polygons / boundaries - Create leaflet or static map of EJ results}
\usage{
mapfastej_counties(
  mydf,
  colorvarname = "pctile.Demog.Index.Supp",
  static_not_leaflet = FALSE,
  main = "Selected Counties",
  ...
)
}
\arguments{
\item{mydf}{something like  ejamit(fips = fips_counties_from_statename("Kentucky"), radius = 0)$results_bysite}

\item{colorvarname}{colname of indicator in mydf that drives color-coding}

\item{static_not_leaflet}{set TRUE to use \code{\link[=map_shapes_plot]{map_shapes_plot()}} instead of \code{\link[=map_shapes_leaflet]{map_shapes_leaflet()}}}

\item{main}{title for map}

\item{...}{passed to map_shapes_plot() if relevant}
}
\value{
leaflet html widget (but if static_not_leaflet=T,
returns just shapes_counties_from_countyfips(mydf$ejam_uniq_id))
}
\description{
Map - County polygons / boundaries - Create leaflet or static map of EJ results
}
\details{
THIS ASSUMES THAT mydf$ejam_unique_id is the county FIPS codes
}
\examples{
\dontrun{
 fips_ky <- fips_counties_from_statename("Kentucky")
 x <- ejamit(fips = fips_ky, radius = 0)
 mapfastej_counties(x$results_bysite)
 }
 # map_shapes_leaflet(shapes = shapes_counties_from_countyfips(fips_ky))

}
\seealso{
\code{\link[=mapfastej]{mapfastej()}} \code{\link[=map_shapes_leaflet]{map_shapes_leaflet()}}
}
