% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata_add.R
\name{metadata_check}
\alias{metadata_check}
\title{helper function in updating the package metadata}
\usage{
metadata_check(
  packages = EJAM::ejampackages,
  which = c("date_saved_in_package", "ejscreen_version", "ejscreen_releasedate",
    "acs_releasedate", "acs_version", "census_version"),
  loadifnotloaded = TRUE
)
}
\arguments{
\item{packages}{Optional. e.g. 'EJAMejscreendata', or can be a vector of character strings,
and if not specified, default is to report on all packages with EJ as part of their name,
like EJAMejscreenapi}

\item{which}{Optional vector (not list) of strings, the attributes.
Default is some typical ones used in EJAM-related packages currently.}

\item{loadifnotloaded}{Optional to control if func should temporarily attach packages not already loaded.}

\item{datasets}{optional, "all" means all data objects exported?
can be a vector of character names of the ones to check like c("bgpts", "blockpoints")}

\item{grepdatasets}{optional, if set to TRUE, datasets should be a query to use
via grep to identify which datasets to check. It always uses ignore.case=TRUE for this.}
}
\description{
Quick and dirty helper during development, to check all the
attributes of all the data files in relevant packages.
It loads unloaded packages as needed, which you might not want it to do,
but it is not coded to be able to check attributes without doing that.
}
\examples{
  x <- metadata_check("EJAM")
  x[x$has_metadata == TRUE, ]
  table(x$has_metadata)

}
\seealso{
\code{\link[=functions_in_pkg]{functions_in_pkg()}}
}
\keyword{internal}
