% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics2children.R
\name{naics2children}
\alias{naics2children}
\title{See NAICS codes queried plus all children of any of those}
\usage{
naics2children(codes, allcodes = EJAM::NAICS, quiet = FALSE)
}
\arguments{
\item{codes}{vector of numerical or character}

\item{allcodes}{Optional (already loaded with package) - dataset with all the codes}

\item{quiet}{whether to avoid printing results to console}
}
\value{
vector of codes and their names
}
\description{
Used by naics_find()
}
\details{
start with shortest (highest level) codes. since tied for nchar, these branches have zero overlap, so do each.
for each of those, get its children = all rows where parentcode == substr(allcodes, 1, nchar(parentcode))
put together list of all codes we want to include so far.
now for the next longest set of codes in original list of codes,
do same thing.
etc. until did it for 5 digit ones to get 6digit children.
take the unique(allthat)
table(nchar(as.character(NAICS)))
2    3    4    5    6
17   99  311  709 1057
}
\examples{
  naics2children(211)
  naics_find(211, exactnumber=TRUE)
  naics_find(211, exactnumber=TRUE, add_children = TRUE)
  NAICS[211][1:3] # wrong
  NAICS[NAICS == 211]
  NAICS["211 - Oil and Gas Extraction"]

}
\seealso{
naics_find() NAICS
}
