% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_categories.R
\name{NAICS_categories}
\alias{NAICS_categories}
\title{see the top level (or other) industrial categories in the NAICS list}
\usage{
NAICS_categories(digits = 2, dataset = ifelse(exists(NAICS), NAICS, NULL))
}
\arguments{
\item{digits}{default is 2-digits NAICS, which is the top level, but could say 3 or more}

\item{dataset}{Should default to the dataset called NAICS, installed with this package.
see \link{NAICS}  Check attr(NAICS, 'year')}
}
\description{
See \url{https://www.census.gov/naics/}
}
\seealso{
\link{NAICS_find}  \link{NAICS}
}
