% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_find.R
\name{naics_find}
\alias{naics_find}
\title{Search for an industrial sector in the list of NAICS codes, see subsectors}
\usage{
naics_find(
  query,
  add_children = FALSE,
  naics_dataset = NULL,
  ignore.case = TRUE,
  exactnumber = FALSE,
  search_on_naics_website = FALSE
)
}
\arguments{
\item{query}{a single word or phrase such as "chemical manufacturing" or "cement"}

\item{add_children}{default is FALSE, so it does NOT chidren (subcategories) of those that match the query.}

\item{naics_dataset}{Should default to the dataset NAICS, installed with this package. see \link{NAICS}}

\item{ignore.case}{default TRUE, ignoring whether query is upper or lower case}

\item{exactnumber}{if TRUE, only return the exact match to (each) queried number (NAICS code)}

\item{search_on_naics_website}{if TRUE (not default),
returns URL of webpage at naics.com with info on the sector}
}
\description{
Just a utility, quick way to view NAICS industrial sectors that contain queried word or phrase,
but can also see all the subcategories within the matching one.
}
\details{
See \url{https://www.census.gov/naics/}
NOTE: By default, this shows NAICS that match the text query,
and also can include all the children NAICS even if they do not match based on text query.
So it first finds NAICS that match the text (or code) query via grep(),
and then can also include all subcategories within those categories.

So naics_find('soap', add_children=TRUE) shows "325612 - Polish and Other Sanitation Good Manufacturing",
and others, not just "3256 - Soap, Cleaning Compound, and Toilet Preparation Manufacturing",
because 3256 matches 'soap' and 325612 is a subcategory of 3256.

The format of NAICS, the naics_dataset, in this package is
dput(NAICS\link{1:4})
c(\verb{11 - Agriculture, Forestry, Fishing and Hunting} = 11, \verb{111 - Crop Production} = 111,
\verb{1111 - Oilseed and Grain Farming} = 1111, \verb{11111 - Soybean Farming} = 11111
)
}
\examples{
 naics_find(8111, exactnumber = FALSE)
 naics_find(8111, exactnumber = TRUE)
 naics_find(8111, exactnumber = TRUE, add_children = TRUE)
 
 naics_find("paper")
 naics_find("cement | concrete")
 cbind(naics_find("pig")
 naics_find("pulp", add_children = FALSE)
 naics_find("pulp", add_children = TRUE)
 naics_find("asdfasdf", add_children = TRUE)  
 naics_find("asdfasdf", add_children = FALSE)
 naics_find("copper smelting", search_on_naics_website=FALSE)
 naics_find("copper smelting", search_on_naics_website=TRUE)
 # browseURL(naics_find("copper smelting", search_on_naics_website=TRUE))
 
 EJAMfrsdata::frs[EJAMfrsdata::frs$REGISTRY_ID \%in\% unlist(
   EJAMfrsdata::get_siteid_from_naics(
     EJAM::naics_find("pulp", add_children = TRUE))[,"REGISTRY_ID"]), 1:5]
   
 EJAMejscreenapi::mapfast(EJAMfrsdata::frs[EJAMfrsdata::frs$REGISTRY_ID \%in\% unlist(
   EJAMfrsdata::get_siteid_from_naics(EJAM::naics_find("pulp"))[,"REGISTRY_ID"]),   ])
   
  naics_find(211, exactnumber=TRUE)
  naics_find(211, exactnumber=TRUE, add_children = TRUE)
  naics2children(211)
  NAICS[211][1:3] # wrong
  NAICS[NAICS == 211]
  NAICS["211 - Oil and Gas Extraction"]
}
\seealso{
\code{\link[EJAMfrsdata:get_siteid_from_naics]{EJAMfrsdata::get_latlon_from_naics()}} \link{naics_categories} \link{NAICS} \code{\link[=naics_findwebscrape]{naics_findwebscrape()}} get_facility_info_via_ECHO function \code{\link[=naics_url_of_code]{naics_url_of_code()}} \code{\link[=naics_url_of_query]{naics_url_of_query()}}
}
