% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_findwebscrape.R
\name{NAICS_findwebscrape}
\alias{NAICS_findwebscrape}
\title{for query term, show list of roughly matching NAICS, scraped from web
This finds more than just \code{\link[=NAICS_find]{NAICS_find()}} does, since that needs an exact match
but this looks at naics.com website which lists various aliases for a sector.}
\usage{
NAICS_findwebscrape(query)
}
\arguments{
\item{query}{text like "gasoline" or "copper smelting"}
}
\value{
data.frame of info on what was found, naics and title
}
\description{
for query term, show list of roughly matching NAICS, scraped from web
This finds more than just \code{\link[=NAICS_find]{NAICS_find()}} does, since that needs an exact match
but this looks at naics.com website which lists various aliases for a sector.
}
\examples{
 NAICS_find("copper smelting", search_on_naics_website=FALSE)
 NAICS_find("copper smelting", search_on_naics_website=TRUE)
  naics_url_of_query("copper smelting")
  \dontrun{
  NAICS_findwebscrape("copper smelting")
  browseURL(naics_url_of_query("copper smelting"))
  browseURL(naics_url_of_code(326))
  }
}
\seealso{
\code{\link[=NAICS_find]{NAICS_find()}} \code{\link[=NAICS_url_of_query]{NAICS_url_of_query()}}
}
