% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_from_federalregister.R
\name{naics_from_federalregister}
\alias{naics_from_federalregister}
\title{naics_from_federalregister - DRAFT WORK IN PROGRESS}
\usage{
naics_from_federalregister(naics_text_copy_from_fr)
}
\arguments{
\item{naics_text_copy_from_fr}{}
}
\description{
naics_from_federalregister - DRAFT WORK IN PROGRESS
}
