% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_validation.R
\name{naics_validation}
\alias{naics_validation}
\title{Validate NAICS uploads}
\usage{
naics_validation(naics_enter, naics_select)
}
\arguments{
\item{naics_enter}{vector of naics}

\item{naics_select}{single value}
}
\value{
boolean value
}
\description{
Validates and prepares echo uploads
}
\keyword{internal}
