% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_naicstable.R
\docType{data}
\name{naicstable}
\alias{naicstable}
\title{data.table of NAICS code(s) and industry names for each EPA-regulated site in Facility Registry Service
Also has the 2,3,4,5,and 6-digit NAICS that this code falls under, where relevant for given length}
\description{
This is similar to the data file EJAM::NAICS but in a more useful format and newer functions work with it.
}
\details{
see \url{https://naics.com}
}
\seealso{
\code{\link[=naics_from_any]{naics_from_any()}} \link{NAICS}  \code{\link[=naics_categories]{naics_categories()}}  \code{\link[=naics_findwebscrape]{naics_findwebscrape()}}
}
