% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{name2fips}
\alias{name2fips}
\alias{fips_from_name}
\alias{names2fips}
\title{Get FIPS codes from names of states or counties
inverse of fips2name(), 1-to-1 map statename, ST, countyname to FIPS of each}
\usage{
name2fips(x)
}
\arguments{
\item{x}{vector of 1 or more exact names of states or ST abbreviations or
countynames that include the comma and state abbrev., like
"Harris County, TX"
(not the same as where ST is separate in \code{\link[=fips_counties_from_countyname]{fips_counties_from_countyname()}})
Ignores case.}
}
\value{
vector of character fips codes
}
\description{
Get FIPS codes from names of states or counties
inverse of fips2name(), 1-to-1 map statename, ST, countyname to FIPS of each
}
\examples{
name2fips(c("de", "NY"))
name2fips("rhode island")
name2fips(c("delaware", "NY")) 
}
