% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_d.R
\docType{data}
\name{names_d}
\alias{names_d}
\title{names_d (DATA) list of demographic indicator names}
\description{
names_d (DATA) list of demographic indicator names
}
\seealso{
\code{\link[=varinfo]{varinfo()}}  \link{map_headernames}  \link{names_d} \link{names_e} \link{namez}
}
