% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_e.R
\docType{data}
\name{names_e}
\alias{names_e}
\title{names_e (DATA) list of environmental indicator names}
\description{
names_e (DATA) list of environmental indicator names
}
\seealso{
\code{\link[=varinfo]{varinfo()}}  \link{map_headernames} \link{names_d} \link{names_e} \link{namez}
}
