% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_ratios.R
\name{plot_barplot_ratios_ez}
\alias{plot_barplot_ratios_ez}
\title{Barplot of ratios of demographic (or other) scores to averages - simpler syntax}
\usage{
plot_barplot_ratios_ez(
  out,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  main = "Demographics at the Analyzed Locations Compared to US Overall",
  ...
)
}
\arguments{
\item{out}{like from \code{\link[=ejamit]{ejamit()}}}

\item{varnames}{vector of indicator names that are ratios to avg, like
c(names_d_ratio_to_avg , names_d_subgroups_ratio_to_avg)
but could be c(names_d_ratio_to_state_avg , names_d_subgroups_ratio_to_state_avg)}

\item{main}{title of plot - must change to note it vs. State if not comparing to US avg.}

\item{...}{passed to \code{\link[=plot_barplot_ratios]{plot_barplot_ratios()}}}
}
\value{
ggplot
}
\description{
Barplot of ratios of demographic (or other) scores to averages - simpler syntax
}
\examples{
plot_barplot_ratios_ez(testoutput_ejamit_100pts_1miles)
}
