% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_ratios.R
\name{plot_barplot_ratios_ez}
\alias{plot_barplot_ratios_ez}
\title{Barplot of ratios of demographic (or other) scores to averages - simpler syntax}
\usage{
plot_barplot_ratios_ez(
  out,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  ...
)
}
\arguments{
\item{out}{like from ejamit()}

\item{varnames}{vector of indicator names that are ratios to avg}

\item{...}{passed to plot_barplot_ratios()}
}
\value{
ggplot
}
\description{
Barplot of ratios of demographic (or other) scores to averages - simpler syntax
}
\examples{
plot_barplot_ratios_ez(testoutput_ejamit_100pts_1miles)
}
