% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_by_pctd.R
\name{plot_distance_by_pctd}
\alias{plot_distance_by_pctd}
\title{plot_distance_by_pctd  -  Plot percent demographics within X miles of a site}
\usage{
plot_distance_by_pctd(
  s2b = NULL,
  mysiteid = NULL,
  myvars = c(names_d_count, names_d_subgroups_count)[1],
  dpctvar = paste0("pct", myvars)
)
}
\arguments{
\item{s2b}{output of \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{mysiteid}{one number that is the siteid to look at in s2b}

\item{myvars}{a colname of a population count variable in blockgroupstats indicating which to plot, like "hisp" or "lowinc"
and only works for one indicator at a time so far.}

\item{dpctvar}{a colname of usastats and statestats that is the percentage version of myvars, like "pcthisp" or "pctlowinc"}
}
\value{
returns s2b but with more columns in it, like cumpop, cumdpop, pctdwithin
}
\description{
plot_distance_by_pctd  -  Plot percent demographics within X miles of a site
}
\examples{
  plot_distance_by_pctd()
}
