% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{plot_distance_cdf_by_group}
\alias{plot_distance_cdf_by_group}
\title{plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances}
\usage{
plot_distance_cdf_by_group(
  results_bybg_people,
  radius_miles = round(max(x$distance_avg, na.rm = T), 1),
  demogvarname = c(namez$d, namez$d_subgroups),
  demoglabel = NULL,
  colorlist = colors()[1:length(demogvarname)],
  coloroverall = "gray"
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{radius_miles}{miles radius that was max distance analyzed}

\item{demogvarname}{names of columns in results_bybg_people, e.g., "pctlowinc"}

\item{demoglabel}{friendly text names for labelling graphic, like "Low income residents"}

\item{colorlist}{colors like "red" etc. for the demographic groups of interest}

\item{coloroverall}{color like "gray" for everyone as a whole}
}
\value{
invisibly returns full table of sorted distances of blockgroups, cumulative count of demog groups at that block group's distance
}
\description{
plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances
}
\seealso{
\code{\link[=distance_by_groups]{distance_by_groups()}} \code{\link[=ejamit]{ejamit()}} for examples
}
