% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_cdf_by_group.R
\name{plot_distance_cdf_by_group}
\alias{plot_distance_cdf_by_group}
\title{plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances}
\usage{
plot_distance_cdf_by_group(
  results_bybg_people = NULL,
  radius_miles =
    round(max(results_bybg_people$distance_min_avgperson[!is.infinite(results_bybg_people$distance_min_avgperson)],
    na.rm = T), 1),
  subgroups_type = NULL,
  demogvarname = NULL,
  demoglabel = NULL,
  colorlist = colorspace::diverging_hcl(length(demogvarname)),
  coloroverall = "black",
  returnwhat = "table",
  ...
)
}
\arguments{
\item{results_bybg_people}{data.table from doaggregate()$results_bybg_people}

\item{radius_miles}{miles radius that was max distance analyzed}

\item{subgroups_type}{optional, can be set to "nh" or "alone".
Specifies types of race ethnicity subgroups to use for demogvarname
but only if demogvarname is not specified as a parameter.
If neither is specified it tries to use default_subgroups_type
if that is a variable set by global.R, since it cannot check the reactive variable input$subgroups_type
outside the context of the web app.}

\item{demogvarname}{optional way to specify names of columns to use from results_bybg_people,
e.g., c("pctlowinc", "pctmin"), or  namez$d, or
could be a vector of subgroups such as namez$d_subgroups_nh that includes "pctnhba" etc.
or namez$d_subgroups_alone that includes "pctba" etc.,
but if demogvarname is not specified here as a parameter,
this info could also be specified by the subgroups_type parameter here.
If neither is specified, the function will try to use a default
(which may not reflect any changes being made during development of EJAM if default_subgroups_type is in flux)}

\item{demoglabel}{friendly text names for labelling graphic, like "Low income residents"}

\item{colorlist}{colors like "red" etc. for the demographic groups of interest}

\item{coloroverall}{color like "gray" for everyone as a whole}

\item{returnwhat}{If returnwhat is "table", invisibly returns a
full table of sorted distances of blockgroups,
cumulative count of demog groups at that block group's distance.
If returnwhat is "plotfilename" then it returns the full path including filename of a .png in a tempdir
If returnwhat is "plot" then it returns the plot object as needed for xls_formatting2()}
}
\value{
see returnwhat parameter
}
\description{
plot_distance_cdf_by_group  - SLOW - needs to be optimized
CDF Line Plots of cumulative share of each demographic group, within each distance
Each groups distribution of distances
}
\examples{
 y <- ejamit(testpoints_100, radius = 3)
 plot_distance_mean_by_group(y$results_bybg_people) # or distance_mean_by_group() synonym
 print(distance_by_group(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income'))
 distance_by_group_plots(y$results_bybg_people, 
   demogvarname = 'pctlowinc', demoglabel = 'Low Income')
 xyz = plot_distance_cdf_by_group(y$results_bybg_people) #  
 tail(round(xyz,3))
 tail(xyz[xyz$pctwa <= 0.501, ]) #  Median distance to nearest site here 
   for White Alone is 2.15 miles, but >60\% of Black Alone have a site that close.
 tail(xyz[xyz$pctba <= 0.501, ]) #  Median distance to nearest site here 
   for Black Alone is 1.85 miles
 round(tail(xyz[xyz$dist <=1, ]), 3) #  11\% of White have a site within 1 mile, 
   compared to 18.7\% of Asian who do.
}
\seealso{
\code{\link[=distance_by_groups]{distance_by_groups()}} \code{\link[=ejamit]{ejamit()}} for examples
}
