% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vs_us - DRAFT.R
\name{plot_vs_us}
\alias{plot_vs_us}
\title{draft code to plot distribution of a variable among residents near set of sites vs nationwide
Not yet population weighted}
\usage{
plot_vs_us(
  bysite = ejamit()$results_bysite,
  varname = "Demog.Index",
  refdt = NULL,
  type = "box",
  colorfills = c("lightblue", "orange"),
  ...
)
}
\arguments{
\item{bysite}{table of results from ejamit()$results_bysite, like testoutput_ejamit_1000pts_1miles$results_bysite}

\item{varname}{name of column in bysite, like  "Demog.Index"}

\item{type}{"box", "plotly", or "ggplot"}

\item{colorfills}{two colors for boxplot}
}
\value{
plots
}
\description{
draft code to plot distribution of a variable among residents near set of sites vs nationwide
Not yet population weighted
}
\examples{
 # plot_vs_us(testoutput_ejamit_1000pts_1miles$results_bysite, type = 'box')
 # plot_vs_us(testoutput_ejamit_1000pts_1miles$results_bysite, varname = "pctlingiso", type =  'box', ylim=c(0,20))
 # plot_vs_us(testoutput_ejamit_1000pts_1miles$results_bysite, varname = "pctlingiso", type =  'ggplot')
 # plot_vs_us(testoutput_ejamit_1000pts_1miles$results_bysite, varname = "pctnhaa", type =  'ggplot')
 # plot_vs_us(testoutput_ejamit_1000pts_1miles$results_bysite, varname = "pctnhaa", type = 'box', ylim = c(0,20))
}
