% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotblocksnearby.R
\name{plotblocksnearby}
\alias{plotblocksnearby}
\title{plotblocksnearby - Map view of Census blocks (their centroids) near one or more sites
Utility to quickly view one or more facility points on map with the blocks found nearby}
\usage{
plotblocksnearby(
  sitepoints,
  radius = 3,
  sites2blocks,
  usemapfast = TRUE,
  returnmap = FALSE,
  ...
)
}
\arguments{
\item{sitepoints}{table of points with lat, lon in decimal degrees (data.frame or data.table),
but also could just be the output of getblocksnearby() if that has already been done.}

\item{radius}{in miles}

\item{sites2blocks}{If provided, used as sites2blocks like \link{testoutput_getblocksnearby_10pts_1miles}
If neither sites2blocks nor sitepoints is provided it cannot plot and returns error.
If sites2blocks and sitepoints are both provided, it uses them both to plot blocks and sites (centers of circles).
If sites2blocks not provided, but sitepoints alone is provided, checks if sitepoints is actually sites2blocks, and uses as such.
If sites2blocks not provided, but sitepoints alone is provided, and sitepoints is really sitepoints, it runs getblocksnearby() to create sites2blocks.
If sites2blocks is provided, but sitepoints is not, it could only do a bad approximation of sitepoints so it will not draw the circle or site at center of the circle.}

\item{usemapfast}{optional. simpler plot if FALSE}

\item{returnmap}{optional. if set TRUE, returns the leaflet map object instead of tabular info.}

\item{...}{passed to mapfast() or plot() depending on usemapfast}
}
\value{
invisibly returns sites2blocks like getblocksnearby() does
}
\description{
plotblocksnearby - Map view of Census blocks (their centroids) near one or more sites
Utility to quickly view one or more facility points on map with the blocks found nearby
}
\details{
Uses \code{\link[=getblocksnearby]{getblocksnearby()}} if lat,lon points provided as sitepoints,
but skips it if looks like user passed output of getblocksnearby(),
and then displays a map of those blocks near the specified point.
}
\examples{
  #  see all Census Blocks within 1 mile of 1 site, if already had run getblocksnearby()
   z =  plotblocksnearby(sitepoints = testpoints_10, 
   sites2blocks = testoutput_getblocksnearby_10pts_1miles[siteid == 1,], radius = 1  )
   # see two sites if already had run getblocksnearby()
   z =  plotblocksnearby(sitepoints = testpoints_10[c(4,10),], 
   sites2blocks = testoutput_getblocksnearby_10pts_1miles[siteid \%in\% c(4,10),], radius = 1  )
  \dontrun{
  # See one randomly selected regulated facility from FRS and all Census Blocks within 2 miles:
    plotblocksnearby(testpoints_n(1), 2) 
  # See two sites and all Census Blocks within 5 kilometers
    plotblocksnearby(testpoints_2, radius = convert_units(5, from = "km", towhat = "miles"))
  # See 100 sites and all blocks within 1 mile of each - 
  # Note you have to specify radius here or it uses default that may not match intent 
  # - and this is a bit slow
  plotblocksnearby(testdata_ejamit_output_100pts_1miles$results_bysite[,.(siteid, lat,lon)], radius = 1) 
  }
}
