% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotblocksnearby.R
\name{plotblocksnearby}
\alias{plotblocksnearby}
\title{Map view of Census blocks (their centroids) near one or more sites
Utility to quickly view one or more facility points on map with the blocks found nearby}
\usage{
plotblocksnearby(sitepoints, radius = 3, usemapfast = TRUE, ...)
}
\arguments{
\item{sitepoints}{table of points with lat, lon in decimal degrees (data.frame or data.table)}

\item{radius}{in miles}

\item{usemapfast}{optional. simpler plot if FALSE}

\item{...}{passed to mapfast() or plot() depending on usemapfast}
}
\value{
invisibly returns sites2blocks like getblocksnearby() does
}
\description{
Map view of Census blocks (their centroids) near one or more sites
Utility to quickly view one or more facility points on map with the blocks found nearby
}
\examples{
plotblocksnearby(testpoints_n(1), 2) 
}
