% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxistat2.R
\name{proxistat2}
\alias{proxistat2}
\title{create proximity score for every blockgroup
indicator of each blockgroups proximity to user set of facilities or sites
Still need area of each block to fix this func proxistat2()}
\usage{
proxistat2(pts, cutoff = 8.04672, quadtree)
}
\arguments{
\item{pts}{data.table of lat lon}

\item{cutoff}{distance max, in miles, default is 5km}

\item{quadtree}{must be localtree from EJAM::}

\item{area}{A number or vector of numbers giving size of each spatial unit with FIPS.pop,
in square miles or square kilometers depending on the \code{units} parameter.
Optional. Default is 0, in which case no adjustment is made for small distance,
which can cause unrealistically large scores.
But if point to point distance is zero distance, and area=0,
NA instead of Inf will be returned for the score.}
}
\value{
data.table with proximityscore, bgfips, lat, lon, etc.
}
\description{
create proximity score for every blockgroup
indicator of each blockgroups proximity to user set of facilities or sites
Still need area of each block to fix this func proxistat2()
}
\examples{
 # x <- proxiscore_bg(pts = pts[1:1000,], quadtree = localtree) 
 #
 # summary(x$proximityscore)
 # # analyze.stuff::pctiles(x$proximityscore)
 # plot(x$lon, x$lat)
 # cat("problem = still have infinite or huge values since function is not yet correcting for tiny distances \n")
 # tops = x$proximityscore > 500 & !is.infinite(x$proximityscore) & !is.na(x$proximityscore)
 # points(x$lon[tops], x$lat[tops], col="red")
 
}
