% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionstats.R
\docType{data}
\name{regionstats}
\alias{regionstats}
\title{data.table of 100 percentiles and means, within each EPA Region, across all locations (e.g., block groups)
  for a set of indicators such as percent low income
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
  and like the lookup tables in ejscreen::lookupUSA}
\description{
data.table of 100 percentiles and means, within each EPA Region, across all locations (e.g., block groups)
  for a set of indicators such as percent low income
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
  and like the lookup tables in ejscreen::lookupUSA
}
