% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rmost.R
\name{rmost}
\alias{rmost}
\title{rmost - utility to rm(list=ls()) but not remove key datasets EJAM uses}
\usage{
rmost(
  notremove = c("rmost", "localtree", "blockgroupstats", "usastats", "statestats",
    "bgid2fips", "blockid2fips", "blockpoints", "blockwts", "quaddata")
)
}
\description{
rmost - utility to rm(list=ls()) but not remove key datasets EJAM uses
}
