% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_clean}
\alias{shapefile_clean}
\title{Drop invalid rows and warn if all invalid}
\usage{
shapefile_clean(shp, crs = 4269)
}
\arguments{
\item{shp}{a shapefile object using sf::read_sf()}

\item{crs}{used in shp <- sf::st_transform(shp, crs = crs), default is crs = 4269 or Geodetic CRS NAD83}
}
\value{
a shapefile object using sf::read_sf()
}
\description{
Drop invalid rows and warn if all invalid
}
\seealso{
\code{\link[=shapefile_from_folder]{shapefile_from_folder()}}
}
