% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2blocks_example.R
\docType{data}
\name{sites2blocks_example}
\alias{sites2blocks_example}
\title{data.table of output of [getblocknearby()], each row is a unique site-block-distance}
\description{
data.table of output of [getblocknearby()], each row is a unique site-block-distance
}
