% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_blockid}
\alias{state_from_blockid}
\title{given vector of blockids, get state abbreviation of each
unused. Not needed if you have sites2blocks table that includes a bgid column}
\usage{
state_from_blockid(blockid)
}
\arguments{
\item{blockid}{vector of blockid values as from EJAM in a table called blockpoints}
}
\value{
vector of ST info like AK, CA, DE, etc.
}
\description{
given vector of blockids, get state abbreviation of each
unused. Not needed if you have sites2blocks table that includes a bgid column
}
\details{
This is much faster than using latlon to identify state

test1000 = blockpoints\link{sample(seq_len(NROW(blockpoints)), 1000), }

system.time({ state_from_blockid(test1000$blockid) })

system.time({ state_from_latlon(lat = test1000$lat, lon = test1000$lon) })
}
\examples{
x = sample(blockpoints$blockid, 3)
state_from_blockid(x)[]
mapfast(blockpoints[blockid \%in\% x, ])

all.equal(state_from_blockid(x), state_from_blockid_table(blockpoints[blockid \%in\% x, ]))

## How many of the 8 million+ blockpoints are in each state?

data.table(ST = state_from_blockid(blockpoints$blockid))[, .(blockcount = .N), by = "ST"]

test1000 = blockpoints[sample(seq_len(NROW(blockpoints)), 999), ]
test1000 = rbind(test1000, blockpoints[blockid == 915525, ]) # at least 1 from CA
test1000$ST = state_from_blockid(test1000$blockid) 
 mapfast(test1000, color = "green") \%>\% 
   addCircles(
     lat = test1000$lat[test1000$ST == "CA"], 
     lng = test1000$lon[test1000$ST == "CA"], 
     color = "red")
     
}
\seealso{
unexported state_from_blockid_table()
}
\keyword{internal}
