% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_latlon}
\alias{state_from_latlon}
\title{find what state is where each point is located
Takes 3 seconds to find state for 1k points, so a faster alternative would be useful}
\usage{
state_from_latlon(lon, lat, states_shapefile = EJAM::states_shapefile)
}
\arguments{
\item{lon}{longitudes vector}

\item{lat}{latitudes vector}

\item{shapefile}{shapefile of US States, in package already}
}
\value{
Returns data.frame: ST, statename, FIPS.ST, REGION, n
as many rows as elements in lat or lon
}
\description{
find what state is where each point is located
Takes 3 seconds to find state for 1k points, so a faster alternative would be useful
}
\examples{
 myprogram <- "CAMDBS" # 739 sites
 pts <- frs_from_program(myprogram)[ , .(lat, lon, REGISTRY_ID,  PRIMARY_NAME)]
 # add a column with State abbreviation
 pts[, ST := state_from_latlon(lon = lon, lat=lat)$ST]
 #map these points
 mapfast(pts[ST == 'TX',], radius = 1) # 1 miles radius circles
 
}
\seealso{
\link{states_shapefile} \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}} \code{\link[=states_infer]{states_infer()}}
}
