% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_latlon}
\alias{state_from_latlon}
\title{find what state is where each point is located
Takes 3 seconds to find state for 1k points, so a faster alternative would be useful}
\usage{
state_from_latlon(lat, lon, states_shapefile = EJAM::states_shapefile)
}
\arguments{
\item{lat}{latitudes vector}

\item{lon}{longitudes vector}

\item{shapefile}{shapefile of US States, in package already}
}
\value{
Returns data.frame: ST, statename, FIPS.ST, REGION,
each as long as lat or lon
}
\description{
find what state is where each point is located
Takes 3 seconds to find state for 1k points, so a faster alternative would be useful
}
\seealso{
\link{states_shapefile} \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}}
}
