% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_infer.R
\name{states_infer}
\alias{states_infer}
\title{states_infer - Get table of info on States (from latlon or FIPS)}
\usage{
states_infer(x)
}
\arguments{
\item{x}{data.frame or data.table with either ST column or lat and lon columns, or FIPS,
and optionally a column like  id or n}
}
\value{
data.frame with unique id, ST, etc.
}
\description{
Get cleaned table of US State etc. by ejam_uniq_id, from lat/lon, or from FIPS
}
\seealso{
\code{\link[=state_from_latlon]{state_from_latlon()}} \code{\link[=state_from_fips]{state_from_fips()}}
}
