% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_infer.R
\name{states_infer}
\alias{states_infer}
\title{states_infer
Get cleaned table of US State etc. by siteid, from lat/lon or other info}
\usage{
states_infer(x)
}
\arguments{
\item{x}{data.frame or data.table with either ST column or lat and lon columns,
and optionally a column with siteid or column called n}
}
\value{
data.frame with unique siteid, ST, etc.
}
\description{
states_infer
Get cleaned table of US State etc. by siteid, from lat/lon or other info
}
