% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_shapefile.R
\docType{data}
\name{states_shapefile}
\alias{states_shapefile}
\title{US States boundaries 2020 shapefile from TIGER}
\description{
US States boundaries 2020 shapefile from TIGER
}
\details{
setwd("~/../../R/mysource/EJAM/data-raw")

download.file("https://www.census.gov/cgi-bin/geo/shapefiles/index.php?year=2020&layergroup=States+\%28and+equivalent\%29", destfile = "tl_2020_us_state.zip" )

dir.create("./shp")

unzip("tl_2020_us_state.zip", exdir = "./shp")

states_shapefile <- sf::st_read("./shp")
}
