% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2ratios.R
\name{table_ratios_from_ejamit}
\alias{table_ratios_from_ejamit}
\title{Quick view of summary stats by type of stat, but lacks rounding specific to each type, etc.}
\usage{
table_ratios_from_ejamit(ejamitout, decimals = 1)
}
\arguments{
\item{ejamitout}{list as from ejamit() that includes results_overall}

\item{decimals}{optional number of decimal places to round to}
}
\value{
prints to console and returns a simple data.frame
}
\description{
Quick view of summary stats by type of stat, but lacks rounding specific to each type, etc.
}
