% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata_sites2blocks}
\alias{testdata_sites2blocks}
\alias{sites2blocks_example}
\title{data.table of output of \code{\link[=getblocksnearby]{getblocksnearby()}}, each row is a unique site-block-distance}
\description{
data.table of output of \code{\link[=getblocksnearby]{getblocksnearby()}}, each row is a unique site-block-distance
}
\seealso{
\link{testdata_sites2blocks} \link{testdata_sites2blocks_1000}
}
