% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2blocks_example.R
\docType{data}
\name{testdata_sites2blocks}
\alias{testdata_sites2blocks}
\alias{sites2blocks_example}
\title{data.table of output of \code{\link[=getblocknearby]{getblocknearby()}}, each row is a unique site-block-distance}
\description{
data.table of output of \code{\link[=getblocknearby]{getblocknearby()}}, each row is a unique site-block-distance
}
\details{
library(EJAM)
testdata_sites2blocks <- sites2blocks_example
usethis::use_data(testdata_sites2blocks)
}
