% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata_sites2blocks_1000}
\alias{testdata_sites2blocks_1000}
\title{data.table of output of \link{getblocknearby(testpoints_1000,cutoff = 3.1)}, each row is a unique site-block-distance}
\description{
data.table of output of \link{getblocknearby(testpoints_1000,cutoff = 3.1)}, each row is a unique site-block-distance
}
\seealso{
\link{testdata_sites2blocks}
}
