% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trilaterate-DRAFT.R
\name{trilaterate_sites2blocks}
\alias{trilaterate_sites2blocks}
\title{DRAFT - Estimate lat,lon of each ejam_uniq_id, from outputs of getblocksnearby()}
\usage{
trilaterate_sites2blocks(s2b)
}
\arguments{
\item{s2b}{like \link{testoutput_getblocksnearby_10pts_1miles}}
}
\value{
a data.table with one row per unique ejam_uniq_id from input dt,
plus lat,lon columns
}
\description{
get data.table with ejam_uniq_id, lat,lon of each site (eg for when you did not save sitepoints info)
}
\examples{
s2b = copy(testoutput_getblocksnearby_10pts_1miles)
s2b_located = latlon_join_on_blockid(s2b) # done by trilaterate also
inferred_sites = trilaterate_sites2blocks(s2b)
inferred_sites
plotblocksnearby(s2b_located)

}
