% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_4table.R
\name{url_4table}
\alias{url_4table}
\title{Create URLs in columns}
\usage{
url_4table(lat, lon, radius, regid = NULL, as_html = TRUE)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}

\item{radius}{vector of values for radius in miles}

\item{regid}{optional vector of FRS registry IDs if available to use to create links
to detailed ECHO facility reports}

\item{as_html}{logical, optional.
passed to url_ejscreen_report() and url_ejscreenmap()
from EJAMejscreenapi package}
}
\value{
list of data.frames to append to the list of data.frames created by
ejamit() or doaggregate(),
list(results_bysite = results_bysite, results_overall = results_overall,
newcolnames=newcolnames)
}
\description{
Create URLs in columns
}
\seealso{
url_ejscreen_report() url_ejscreenmap() url_echo_facility_webpage()
from EJAMejscreenapi package
}
\keyword{internal}
