% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_4table}
\alias{url_4table}
\title{URL functions - Create URLs in columns, for EJAM}
\usage{
url_4table(lat, lon, radius, regid = NULL, as_html = TRUE)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}

\item{radius}{vector of values for radius in miles}

\item{regid}{optional vector of FRS registry IDs if available to use to create links
to detailed ECHO facility reports}

\item{as_html}{logical, optional.
passed to url_ejscreen_report() and url_ejscreenmap()}
}
\value{
list of data.frames to append to the list of data.frames created by
\code{\link[=ejamit]{ejamit()}} or \code{\link[=doaggregate]{doaggregate()}},

\verb{list(results_bysite = results_bysite, }
\verb{    results_overall = results_overall,}
\verb{      newcolnames=newcolnames)}
}
\description{
Could start to use this in server and ejamit(), and already used in table_xls_format()
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}} \code{\link[=url_ejscreenmap]{url_ejscreenmap()}} \code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}}
}
\keyword{internal}
