% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_4table.R
\name{url_4table}
\alias{url_4table}
\title{url_4table}
\usage{
url_4table(lon, lat, radius, regid = NULL, as_html = TRUE)
}
\arguments{
\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{radius}{vector of values for radius in miles}

\item{regid}{vector of FRS registry IDs if available to use to create links to detailed ECHO facility reports}

\item{as_html}{logical}
}
\value{
list of data.frames to append to the list of data.frames created by ejamit() or doaggregate(),
list(results_bysite = results_bysite, results_overall = results_overall, newcolnames=newcolnames)
}
\description{
url_4table
}
