% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_countyhealthrankings.R
\name{url_countyhealthrankings}
\alias{url_countyhealthrankings}
\title{url_countyhealthrankings}
\usage{
url_countyhealthrankings(fips, year = 2023)
}
\arguments{
\item{fips}{vector of fips codes of counties, 5 characters each, like "10003"}

\item{year}{2023}
}
\value{
vector of URLs
}
\description{
url_countyhealthrankings
}
