% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_get_via_url-DRAFT.R
\name{url_get_via_url}
\alias{url_get_via_url}
\title{utility - work in progress: GET query of ejscreen ejquery map services}
\usage{
url_get_via_url(url)
}
\arguments{
\item{url}{the url for an EJScreen ejquery request}
}
\value{
json format results
}
\description{
utility - work in progress: GET query of ejscreen ejquery map services
}
