% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usastats.R
\docType{data}
\name{usastats}
\alias{usastats}
\title{data.table of 100 percentiles and means}
\description{
data.table of 100 percentiles and means (about 100 rows)
  in the USA overall, across all locations (e.g., block groups in [blockgroupstats])
  for a set of indicators such as percent low income.
  Each column is one indicator (or specifies the percentile).
  
  This should be similar to the lookup tables in the gdb on the FTP site of EJScreen.
}
