% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{usastats_means}
\alias{usastats_means}
\title{convenient way to see USA MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators from lookup table}
\usage{
usastats_means(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=usastats_query]{usastats_query}}
  \describe{
    \item{\code{varnames}}{names of columns in lookup table, like "proximity.rmp"}
    \item{\code{PCTILES}}{vector of percentiles 0-100 and/or "mean"}
  }}
}
\description{
convenient way to see USA MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators from lookup table
}
