% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{usastats_means}
\alias{usastats_means}
\title{see statestats_query()}
\usage{
usastats_means(...)
}
\arguments{
\item{...}{see statestats_query()}
}
\description{
see statestats_query()
}
\examples{
see statestats_query()
}
