% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varinfo.R
\name{varinfo}
\alias{varinfo}
\title{varinfo - Get metadata for a variable, like its type, glossary definition, decimal places rounding, etc.}
\usage{
varinfo(var, info, varnametype = "rname")
}
\arguments{
\item{var}{vector of variable names such as c("pctlowinc", "cancer") or c(names_d, names_d_subgroups)}

\item{info}{one type of metadata/info needed, such as "decimals" or "long",
which should be the name of a column in map_headernames, to pass to fixcolnames()
which is the function that actually looks up the info.
Or a vector of those, in which case a table is returned, one column for each.}

\item{varnametype}{which column of map_headernames to use when looking for var, like "rname" or "api" or "long"}
}
\value{
named vector same size as var, with var as names;
unless info is a vector in which case a table is returned, one col per value of info.
Results can be character, numeric, etc. depending on what info is requested
}
\description{
varinfo - Get metadata for a variable, like its type, glossary definition, decimal places rounding, etc.
}
\examples{
 
varinfo("traffic.score", "decimals")
varinfo(names_d, "long")
myvars <- c(names_d, names_d_subgroups, names_e)
myinfo <- "percentage"
cbind(  is.a.percentage = varinfo(myvars, myinfo) )
cbind(varinfo(names_all_r, "pctile."))
myinfo <- "long"
cbind(varinfo(myvars, myinfo) )
table_rounding_info(names_e)
 
varinfo(
 var = c(names_these, names_d_pctile),
 info = c(
 "topic_root_term", "varcategory", "vartype", "percentage", "pctile.", "calculation_type"
))

varinfo(names_all_r, c("varcategory", "varlist", "in_api", "in_bgcsv"))

}
\seealso{
fixcolnames() \code{\link[=table_rounding_info]{table_rounding_info()}}
}
