% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{varname2color_ejam}
\alias{varname2color_ejam}
\title{varname2color_ejam - helper function - for color coding excel sheet columns
Convert R variable name of indicator to appropriate color for header row in Excel}
\usage{
varname2color_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{things like us.avg.pctlowinc}
}
\value{
vector of colors
}
\description{
varname2color_ejam - helper function - for color coding excel sheet columns
Convert R variable name of indicator to appropriate color for header row in Excel
}
\seealso{
\code{\link[=varname2vartype_ejam]{varname2vartype_ejam()}} \code{\link[=varname2varcategory_ejam]{varname2varcategory_ejam()}} \code{\link[=vartype_cat2color_ejam]{vartype_cat2color_ejam()}}
}
