% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_formatting2.R
\name{xls_formatting2}
\alias{xls_formatting2}
\title{xls_formatting2
Format EJAM tabular outputs for saving as Excel spreadsheet}
\usage{
xls_formatting2(
  overall,
  eachsite,
  longnames = NULL,
  formatted = NULL,
  bybg = NULL,
  plot_distance_by_group = TRUE,
  summary_plot = NULL,
  plotlatest = FALSE,
  plotfilename = NULL,
  mapadd = FALSE,
  ok2plot = TRUE,
  analysis_title = "EJAM analysis",
  buffer_desc = "Selected Locations",
  radius_or_buffer_in_miles = NULL,
  radius_or_buffer_description =
    "Miles radius of circular buffer (or distance used if buffering around polygons)",
  notes = NULL,
  heatmap_colnames = NULL,
  heatmap_cuts = c(80, 90, 95),
  heatmap_colors = c("yellow", "orange", "red"),
  heatmap2_colnames = NULL,
  heatmap2_cuts = c(1.009, 2, 3),
  heatmap2_colors = c("yellow", "orange", "red"),
  hyperlink_colnames = c("EJScreen Report", "EJScreen Map", "ECHO report"),
  graycolnames = NULL,
  narrowcolnames = NULL,
  graycolor = "gray",
  narrow6 = 6,
  testing = FALSE,
  launchexcel = FALSE,
  saveas = NULL,
  ...
)
}
\arguments{
\item{overall}{table to save in one tab, from ejamit()$overall, EJAM analysis of indicators overall (one row)}

\item{eachsite}{table to save in one tab, from ejamit()$overall, EJAM analysis site by site (one row per site)}

\item{longnames}{vector of indicator names to display in Excel table}

\item{formatted}{optional table to save in one tab, from ejamit()$overall, EJAM analysis overall in different format}

\item{bybg}{Optional large table of details of each block group that is only needed to analyze distances by group.}

\item{plot_distance_by_group}{logical, whether to try to add a plot of mean distance by group.
This requires that bybg be provided as a parameter input to this function.}

\item{summary_plot}{optional plot object passed from EJAM shiny app to save in 'Plot' sheet of Excel table}

\item{plotlatest}{optional logical. If TRUE, the most recently displayed plot (prior to this function being called) will be inserted into a tab called plot2}

\item{plotfilename}{the full path including name of .png file to insert}

\item{mapadd}{logical optional - try to include a map of the points}

\item{ok2plot}{can set to FALSE to prevent plots from being attempted, while debugging}

\item{analysis_title}{optional title passed from Shiny app to 'Notes' sheet}

\item{buffer_desc}{optional description of buffer used in analysis, passed to 'Notes' sheet}

\item{radius_or_buffer_description}{optional text saying if distance is radius or polygon buffer, passed to 'Notes' sheet}

\item{notes}{Text of additional notes to put in the notes tab, optional vector of character elements pasted in as one line each.}

\item{heatmap_colnames}{optional vector of colnames to apply heatmap colors}

\item{heatmap_cuts}{vector of values to separate heatmap colors, between 0-100}

\item{heatmap_colors}{vector of color names for heatmap bins, same length as
heatmap_cuts, where first color is for those >= 1st cutpoint, but <2d,
second color is for those >=2d cutpoint but <3d, etc.}

\item{hyperlink_colnames}{names of which to treat as URLs that should be hyperlinks}

\item{graycolnames}{which columns to deemphasize}

\item{narrowcolnames}{which column numbers to make narrow}

\item{graycolor}{color used to deemphasize some columns}

\item{narrow6}{how narrow}

\item{testing}{optional for testing only}

\item{launchexcel}{Set to TRUE to have this function launch Excel immediately, showing the final workbook created here.}

\item{saveas}{If not NULL, and a valid path with filename.xlsx is provided,
the workbook will be saved locally at that path and name. Warning: it will overwrite an existing file.}

\item{...}{other params passed along to \code{\link[openxlsx:writeData]{openxlsx::writeData()}}}

\item{radius_miles}{If provided, miles buffer distance (from polygon or from point if circular buffers)}
}
\value{
a workbook, ready to be saved in spreadsheet format, with tabs like "Overall" and "Each Site"
}
\description{
xls_formatting2
Format EJAM tabular outputs for saving as Excel spreadsheet
}
\details{
Already took and put here most or all of code from xls_formatting() or xls_formatting_api()
}
\examples{
\dontrun{
  wb <- xls_formatting2(testoutputs_$results_overall, datasetResults()$results_bysite,
    saveas =  "results.xlsx")
}
}
